themis: Extra Recipes Steps for Dealing with Unbalanced Data

A dataset with an uneven number of cases in each class is said to be unbalanced. Many models produce a subpar performance on unbalanced datasets. A dataset can be balanced by increasing the number of minority cases using SMOTE <arXiv:1106.1813>, BorderlineSMOTE and ADASYN. Or by decreasing the number of majority cases using NearMiss or Tomek link removal.

See also