Synthetic Dataset With a Circle

A random dataset with two classes one of which is inside a circle. Used for examples to show how the different methods handles borders.

Usage

circle_example

Format

A data frame with 200 rows and 4 variables:

x

Numeric.

y

Numeric.

class

Factor, values "Circle" and "Rest".

id

character, ID variable.