ROSE Algorithm

A thin wrapper around ROSE::ROSE() that generates a synthetic balanced sample by enlarging the feature space of minority and majority class examples.

Usage

rose(
  df,
  var,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1
)

Arguments

df

A data.frame or tibble. Must have 1 factor variable with exactly 2 levels and remaining numeric variables.

var

Character, name of variable containing the 2-level factor variable.

over_ratio

A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows as the majority level. See vignette("ratio", package = "themis") for more details.

minority_prop

A numeric value between 0 and 1 for the proportion of synthetic observations from the minority class. Defaults to 0.5, which generates an equal split of minority and majority synthetic observations. This parameter controls the class balance within the synthetic data, while over_ratio controls the total size of the synthetic data.

minority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the minority class. Defaults to 1.

majority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the majority class. Defaults to 1.

Value

A data.frame or tibble, depending on type of df.

Details

The factor variable used to balance around must only have 2 levels.

The ROSE algorithm works by selecting an observation belonging to class k and generating new examples in its neighborhood, which is determined by a smoothing matrix H_k. Smaller values of minority_smoothness and majority_smoothness shrink the entries of H_k, producing tighter neighborhoods. This is a cautious choice when there is a concern that excessively large neighborhoods could blur the boundaries between classes.

This function is a thin wrapper around ROSE::ROSE(). For full details on the underlying implementation, see that function's documentation.

References

Lunardon, N., Menardi, G., and Torelli, N. (2014). ROSE: a Package for Binary Imbalanced Learning. R Journal, 6:79–89.

Menardi, G. and Torelli, N. (2014). Training and assessing classification rules with imbalanced data. Data Mining and Knowledge Discovery, 28:92–122.

See also

Examples

rose(circle_example[, c("x", "y", "class")], var = "class")
#>               x          y  class
#> 1   12.72808609  5.1994429   Rest
#> 2    3.12284752 -0.8443850   Rest
#> 3   15.32744344 14.2217217   Rest
#> 4   13.45546873 14.1589815   Rest
#> 5   14.84556269  4.1383320   Rest
#> 6    2.51913907 16.4640250   Rest
#> 7    4.21097261 12.6813205   Rest
#> 8   16.70689627  5.8788516   Rest
#> 9    5.16936162 12.0857079   Rest
#> 10  11.17229162  7.1449629   Rest
#> 11   0.75304581  9.0673585   Rest
#> 12   9.26559538  2.7808313   Rest
#> 13   7.77020816 16.3020747   Rest
#> 14  11.50634574 13.9824557   Rest
#> 15  10.02338738 10.0193400   Rest
#> 16   1.16816309  2.9958277   Rest
#> 17   5.06595183  3.1554234   Rest
#> 18   5.15577273  9.5849127   Rest
#> 19   9.38212615 10.0533042   Rest
#> 20  14.59081852  0.9405091   Rest
#> 21  13.87940756  3.1322960   Rest
#> 22   4.07229536  8.1476346   Rest
#> 23   9.93538483 17.0170388   Rest
#> 24  12.95183105  5.9934518   Rest
#> 25  10.94350130  3.7990488   Rest
#> 26  11.07326697  1.7073000   Rest
#> 27   1.60304487  2.9392404   Rest
#> 28   5.61033737  9.1170233   Rest
#> 29   8.33454410 15.7626631   Rest
#> 30   6.71092452  3.0671454   Rest
#> 31   3.33110774 14.1152926   Rest
#> 32   4.08630561  8.1913240   Rest
#> 33   5.25988890  8.6813626   Rest
#> 34  11.88571436 13.7635202   Rest
#> 35  12.22711209  2.5454592   Rest
#> 36   5.36446821  2.7894891   Rest
#> 37  14.90068884  5.8564215   Rest
#> 38  11.68405800  4.3776805   Rest
#> 39   4.63260953  5.0135529   Rest
#> 40   6.33670254  0.7143084   Rest
#> 41  13.54071348 12.7368329   Rest
#> 42   9.65623593  8.2825155   Rest
#> 43  14.29469368  2.4902402   Rest
#> 44  10.56924461 15.5218057   Rest
#> 45  14.87765183  3.2203335   Rest
#> 46  12.29262048 12.8010627   Rest
#> 47   7.01760680 15.0339446   Rest
#> 48   4.67024432 12.0480087   Rest
#> 49   4.03343052  2.8127792   Rest
#> 50   4.85672100  2.3626811   Rest
#> 51  10.62728274 10.5057710   Rest
#> 52   7.00313001 12.4687280   Rest
#> 53   2.99634754 11.5135425   Rest
#> 54   5.83134957  5.3395039   Rest
#> 55   7.87480181 11.3172270   Rest
#> 56  12.82246206  7.6014434   Rest
#> 57   4.28708271 13.0593014   Rest
#> 58  11.35450622  8.4340338   Rest
#> 59   4.68671064 12.6578805   Rest
#> 60   4.82322565  1.7092954   Rest
#> 61   2.29646885  8.6138451   Rest
#> 62  16.61270686  1.1725743   Rest
#> 63   4.07626473  6.8407618   Rest
#> 64   6.33387001 15.0274849   Rest
#> 65   9.36543013 16.3364411   Rest
#> 66   3.37908173 10.4294329   Rest
#> 67   6.01707996  3.1685472   Rest
#> 68  12.89575944 14.7890970   Rest
#> 69   5.13871638 15.1509659   Rest
#> 70  10.43733293 11.4613018   Rest
#> 71   5.23529080 11.9657181   Rest
#> 72  12.05748564  3.8338531   Rest
#> 73   4.72678668  0.1815178   Rest
#> 74  16.44362662  0.6865874   Rest
#> 75  14.05376949  6.9301786   Rest
#> 76  14.75316378  7.5034034   Rest
#> 77  15.33590339  4.2241576   Rest
#> 78   4.35728760 -1.3544068   Rest
#> 79   0.50509383 12.9760309   Rest
#> 80   2.54717482 12.5866056   Rest
#> 81  12.74641098 11.4915596   Rest
#> 82   3.66667543  3.1906061   Rest
#> 83  11.99458181 10.7220154   Rest
#> 84   1.46797000  6.5387319   Rest
#> 85   8.50974092 13.4561890   Rest
#> 86   6.40006886  4.0223152   Rest
#> 87  13.75779457 14.4098970   Rest
#> 88  11.64979086  5.0741578   Rest
#> 89  16.28341492 14.3414143   Rest
#> 90   6.33069868 12.8782164   Rest
#> 91   7.64801451 15.2027085   Rest
#> 92   2.34024066  4.3239473   Rest
#> 93   3.00538145  9.5643124   Rest
#> 94  13.61113022 10.4380295   Rest
#> 95  15.14341540  9.9162070   Rest
#> 96   2.59949916 10.9170137   Rest
#> 97   8.28844974 12.4559449   Rest
#> 98   3.96886962 14.5313506   Rest
#> 99   5.01587852 11.0162896   Rest
#> 100  4.41836665  9.5189053   Rest
#> 101 12.25939456 17.1091387   Rest
#> 102  4.55451818 12.5844448   Rest
#> 103  3.99860026  6.2181622   Rest
#> 104  3.64061189 11.6644715   Rest
#> 105  6.02506172 14.2922302   Rest
#> 106 11.04470879 12.3851215   Rest
#> 107  7.12302733 15.9993353   Rest
#> 108  6.43827713 16.2603718   Rest
#> 109  3.11570936 13.1089648   Rest
#> 110  5.99725745 14.8344366   Rest
#> 111 11.10466906  2.6900050   Rest
#> 112  3.73841542  5.6862330   Rest
#> 113 12.60221871 10.5197926   Rest
#> 114  4.06782598 10.0091230   Rest
#> 115  2.63872687 11.7942205   Rest
#> 116  3.84391275  1.8403439   Rest
#> 117  4.07528453  7.8601635   Rest
#> 118 15.01471593  3.1799255   Rest
#> 119  0.50556355  8.8022060   Rest
#> 120  1.05175988 14.4993247   Rest
#> 121  8.17210903 16.0751738   Rest
#> 122  6.22454012  3.2933675   Rest
#> 123 -1.41375218  0.5726552   Rest
#> 124  8.14226101  3.5826655   Rest
#> 125 14.17208518  8.3759792   Rest
#> 126 14.87806321 11.1912257   Rest
#> 127 13.41532307  8.1391272   Rest
#> 128 13.19075835  1.8754148   Rest
#> 129 16.88521326  2.2185036   Rest
#> 130  7.52760881 14.5339476   Rest
#> 131  1.27094436  7.5794633   Rest
#> 132 11.75329156 12.5190978   Rest
#> 133  5.24616851  9.9189591   Rest
#> 134  2.86616584  7.3388027   Rest
#> 135  8.85216485  9.6387465   Rest
#> 136  2.98268754  7.5598868   Rest
#> 137  5.01379208  1.4338775   Rest
#> 138  4.97919902  3.7419075   Rest
#> 139 11.37358186  8.7188400   Rest
#> 140 10.47449469 10.3626422   Rest
#> 141  6.63101053  3.2460623   Rest
#> 142  2.10367984 15.9088403   Rest
#> 143  4.43243806 13.6495252   Rest
#> 144  8.86789223 13.9654626   Rest
#> 145  3.11397896  5.0834012   Rest
#> 146  2.77394599  9.9909106   Rest
#> 147 14.13090094  4.5682954   Rest
#> 148 11.46734503 10.5413952   Rest
#> 149 14.75243455  4.5819201   Rest
#> 150  2.12779054  4.0989063   Rest
#> 151 10.98673667 10.8221398   Rest
#> 152 10.54916496  4.2927034   Rest
#> 153  8.61718866  7.5161828   Rest
#> 154  7.76866142 11.8979148   Rest
#> 155  7.99315796  0.9695725   Rest
#> 156  1.03063193 12.3131011   Rest
#> 157 14.52084562  7.1119619   Rest
#> 158 13.08619144  9.5896768   Rest
#> 159 12.43258076  7.7261163   Rest
#> 160  4.44566002 12.0270667   Rest
#> 161 13.55255285  2.0720443   Rest
#> 162  4.00241482  8.3378550   Rest
#> 163  4.86036644  3.4414565   Rest
#> 164 13.72844925  2.4865326   Rest
#> 165  2.14623797 10.1008397   Rest
#> 166  0.02482424  8.3084283   Rest
#> 167  9.22527206 13.8728396   Rest
#> 168 10.55504312 12.8560773   Rest
#> 169 10.35180104 12.8139524   Rest
#> 170 -1.85340415  9.6609938   Rest
#> 171  6.92217955  4.9516792   Rest
#> 172 16.63888189  4.4676220   Rest
#> 173 11.50817777  9.7319176   Rest
#> 174  5.18222268  2.0823738   Rest
#> 175  9.87236729 11.5272143   Rest
#> 176  9.64801805  2.2184295   Rest
#> 177  2.22556816  5.5850898   Rest
#> 178  5.99811775 13.7295594   Rest
#> 179  4.46700315  9.0561794   Rest
#> 180  5.97430812  2.5494450   Rest
#> 181  4.05810308  2.5316595   Rest
#> 182 10.58985405 16.5661714   Rest
#> 183  8.98718430  1.8813021   Rest
#> 184  5.66022852 13.2598284   Rest
#> 185  1.94941894  2.6310878   Rest
#> 186  5.13757825  9.8924308   Rest
#> 187  4.48782529  1.6043215   Rest
#> 188  9.26592488  7.8207324   Rest
#> 189  7.37188804  6.4084343   Rest
#> 190 14.50200022 11.1702242   Rest
#> 191  5.60406561  5.7195498   Rest
#> 192  2.19423574 12.5976978   Rest
#> 193  5.64393630  7.1436288   Rest
#> 194  3.16800573 11.0819745   Rest
#> 195  4.01569895 13.3054625   Rest
#> 196  8.24176525  3.5842100   Rest
#> 197  1.03181080  8.7173337   Rest
#> 198  8.45941716 16.5462994   Rest
#> 199  4.60086494  2.7294343   Rest
#> 200 10.51880344 10.7750878   Rest
#> 201 12.59106229 12.1135342   Rest
#> 202 12.67749943  0.8320975   Rest
#> 203  5.27651035 14.4772359   Rest
#> 204 13.26782273  7.7206151   Rest
#> 205 11.30606913  8.5358851   Rest
#> 206 14.28764855  5.4815977   Rest
#> 207 13.73969928  7.1838268   Rest
#> 208 11.93727783 11.6091603   Rest
#> 209  4.95811617 13.2051054   Rest
#> 210 13.02371335  3.4206403   Rest
#> 211 14.51406901  9.2556486   Rest
#> 212  1.26997506 10.7459755   Rest
#> 213  1.42603746 12.1467343   Rest
#> 214  6.26877285  2.9846599   Rest
#> 215 12.97562466  3.2710314   Rest
#> 216  6.52780524  5.4088356   Rest
#> 217  6.41853540  1.2564922   Rest
#> 218 11.59277923  5.5316339   Rest
#> 219 14.38725375  8.5079852   Rest
#> 220  6.76344411  9.6622934   Rest
#> 221 12.27449213  9.8591297   Rest
#> 222 14.99090812  9.4654090   Rest
#> 223 10.55162658  6.8548266   Rest
#> 224  5.41888332 13.9609361   Rest
#> 225  3.08849847  9.2493214   Rest
#> 226  1.58932482 14.3526958   Rest
#> 227  2.51653208  5.9351188   Rest
#> 228 16.12316547 11.1199764   Rest
#> 229 -0.39595432  6.0687481   Rest
#> 230 12.36030857 11.3516251   Rest
#> 231  2.56335219  4.0478481   Rest
#> 232  6.89636576 13.9527915   Rest
#> 233  5.55730142 14.6284317   Rest
#> 234 13.79060208  5.4590771   Rest
#> 235 12.75558145  7.8634827   Rest
#> 236 11.15366659  1.4644848   Rest
#> 237  0.95254965  3.8570922   Rest
#> 238 10.07251286 10.2583913   Rest
#> 239  4.49228542  6.3346267   Rest
#> 240  0.17383369  9.1445522   Rest
#> 241 16.78220580  7.5611308   Rest
#> 242 16.97667554 13.4275537   Rest
#> 243 11.65816007  5.0054494   Rest
#> 244  1.54347335  4.0324129   Rest
#> 245 10.37182201  3.9988695   Rest
#> 246 15.04556072 13.0836537   Rest
#> 247  7.79421745  0.2308630   Rest
#> 248 10.41114892 12.1995166   Rest
#> 249 10.25742125 12.8298245   Rest
#> 250  2.98397604 11.0746737   Rest
#> 251 15.38794575 14.2933882   Rest
#> 252  3.90781057 10.0126547   Rest
#> 253  4.86103471 13.2998218   Rest
#> 254  8.84676931 13.5096311   Rest
#> 255  8.88292332 10.0232280   Rest
#> 256 11.37854391  7.8823307   Rest
#> 257  9.81879078  3.5962800   Rest
#> 258 11.21925958  6.8744945   Rest
#> 259  6.01866737 13.4862379   Rest
#> 260  9.55125476  9.6885599   Rest
#> 261 10.31281483 13.0838621   Rest
#> 262 15.86358082 14.2610422   Rest
#> 263 12.24742170 10.0622683   Rest
#> 264 11.40262072 15.4181650   Rest
#> 265 10.68190628 10.3845142   Rest
#> 266  8.73006206 13.5603088   Rest
#> 267 11.30498131  2.7870470   Rest
#> 268  9.33990068  5.3475774   Rest
#> 269 13.78898887  4.2133649   Rest
#> 270  2.48940675  0.4747108   Rest
#> 271  2.13601831  8.6581698   Rest
#> 272 14.69145329  7.5372173   Rest
#> 273 10.13177000 15.8794387   Rest
#> 274  5.88823550  5.0774556   Rest
#> 275  5.27848987 11.8137493   Rest
#> 276  3.43493675 10.6445671   Rest
#> 277  6.54484533 13.7726503   Rest
#> 278  5.29545098  2.5234812   Rest
#> 279 14.81062382  1.1519489   Rest
#> 280  6.25175521  0.6720570   Rest
#> 281  3.51770513  4.2570222   Rest
#> 282 13.51640352 11.0362814   Rest
#> 283 12.73300225 12.5863435   Rest
#> 284 13.84812513  4.3858712   Rest
#> 285  8.89299405  4.6029814   Rest
#> 286  3.20501744 11.5785863   Rest
#> 287  5.23573456  1.2451252   Rest
#> 288 14.85719581  6.0409182   Rest
#> 289 13.41097220 -2.0876904   Rest
#> 290  6.20004639  2.3535602   Rest
#> 291 10.46582092  2.4244458   Rest
#> 292  4.93605356 11.6826869   Rest
#> 293  7.48542901 15.8028121   Rest
#> 294  6.25886702 16.6083271   Rest
#> 295  5.17450209  2.2391679   Rest
#> 296 11.28708242  1.2119319   Rest
#> 297  3.05390200  5.6649300   Rest
#> 298 15.43147155  1.8163703   Rest
#> 299  6.42879384 12.7616358   Rest
#> 300  7.13261916  9.4258745   Rest
#> 301  7.85240314 14.6152454   Rest
#> 302 11.33537718  9.2223738   Rest
#> 303  2.87756338  2.8277377   Rest
#> 304  8.54279654 13.5717158   Rest
#> 305 15.58109859 10.3602918   Rest
#> 306 16.23886811  8.2922308   Rest
#> 307  6.40223931  3.0068684   Rest
#> 308  7.68518587  8.8270064   Rest
#> 309 12.83988587  1.6751103   Rest
#> 310  0.57278774  7.7493247   Rest
#> 311  6.00585626 12.8091621   Rest
#> 312  1.48855556  5.2423435   Rest
#> 313  3.15502839  6.3344910   Rest
#> 314  5.65040496  5.2031542   Rest
#> 315 12.44920424  6.4022216   Rest
#> 316 10.43444508  2.0348661   Rest
#> 317  7.30733325  3.1335184   Rest
#> 318 12.03386997 14.8821283   Rest
#> 319  6.59210387  0.6178782   Rest
#> 320 11.14913822  4.0377195   Rest
#> 321 12.27778320 10.2546402   Rest
#> 322  5.20699144 12.0794150   Rest
#> 323  3.00340259  3.0319666   Rest
#> 324 16.24483059  3.7198053   Rest
#> 325  1.62070379 10.9511292   Rest
#> 326 11.08720801  1.8744398   Rest
#> 327  2.01366049 -1.2937978   Rest
#> 328  1.96736710 -0.3994929   Rest
#> 329  0.64303222 16.5857731   Rest
#> 330  5.72509712  2.0484504   Rest
#> 331  1.07681829 12.9133739   Rest
#> 332  3.28308676 15.1995058   Rest
#> 333 11.17971526 13.9561599   Rest
#> 334  5.45351946 18.6701999   Rest
#> 335  5.53657538  3.9747124   Rest
#> 336  3.10938213  0.3367976   Rest
#> 337 12.11971653 -0.2344144   Rest
#> 338 14.86711595  1.9376307   Rest
#> 339  5.09483261 14.6266035   Rest
#> 340 12.75375962 10.6768019   Rest
#> 341  4.23348966 11.3868847   Rest
#> 342  0.87463791 11.4526231   Rest
#> 343  5.08921677  1.1412035   Rest
#> 344  2.75316197  7.4261369   Rest
#> 345  8.22027228  4.7099601   Rest
#> 346  9.63573496  6.4239237   Rest
#> 347  9.17924014  9.7308671   Rest
#> 348  8.90884165 16.6383845   Rest
#> 349  6.13799418 12.6731157   Rest
#> 350 10.54768272  4.4893589   Rest
#> 351 13.72774458  7.7727730   Rest
#> 352  6.66236726 12.1129475   Rest
#> 353 13.73879164  7.2805433   Rest
#> 354  4.60618863 12.0088931   Rest
#> 355  9.59317432 14.3381160   Rest
#> 356 16.81999351 13.2159711   Rest
#> 357 13.39412501  1.6461479   Rest
#> 358  1.22011459  5.3917758   Rest
#> 359  5.58848850  3.2277723   Rest
#> 360  5.63579381 11.3363720   Rest
#> 361  7.62901634  9.5050779 Circle
#> 362  6.74686598  5.3294629 Circle
#> 363  7.11428106  8.9152349 Circle
#> 364  5.55417937  6.7300701 Circle
#> 365  9.98263248  6.6198879 Circle
#> 366  6.04420343  7.9005632 Circle
#> 367  9.57927660  7.4238781 Circle
#> 368  8.65456382  7.6388967 Circle
#> 369 10.04246850 10.2686430 Circle
#> 370  8.03486447  7.6339897 Circle
#> 371  6.85750809 11.6081128 Circle
#> 372  9.05521497  7.5625875 Circle
#> 373  4.83781434 10.8338902 Circle
#> 374  6.69415775  9.5549617 Circle
#> 375  7.80473752  6.6555923 Circle
#> 376  9.85632074  6.0995785 Circle
#> 377  8.97842879  9.7574064 Circle
#> 378  6.14231033  6.1438160 Circle
#> 379  7.99482078  4.9954470 Circle
#> 380  7.57616909  6.5344627 Circle
#> 381  8.00314404  6.1077519 Circle
#> 382  8.06433782 10.8587514 Circle
#> 383  8.88704527  7.2594025 Circle
#> 384  7.93928494 11.0751056 Circle
#> 385  9.15426726 11.0027623 Circle
#> 386  8.68979146  5.8414554 Circle
#> 387  5.62468417  9.0684542 Circle
#> 388  8.87555347  8.3113901 Circle
#> 389  5.74186552 10.2686201 Circle
#> 390  6.15906763 11.5932806 Circle
#> 391  7.63639089  9.7090568 Circle
#> 392  7.09201982  7.2247803 Circle
#> 393  8.28524958 10.4473196 Circle
#> 394 10.94834859  9.5379910 Circle
#> 395  7.58118670  6.1120460 Circle
#> 396  6.91012967 10.3133777 Circle
#> 397  8.76281410  9.7776228 Circle
#> 398  8.30375481  6.8695640 Circle
#> 399  7.43926852  9.8014932 Circle
#> 400  9.03713082  6.3954452 Circle
#> 401  9.03939590  5.1453585 Circle
#> 402  8.84286959 10.0286160 Circle
#> 403  9.03809507  8.8312664 Circle
#> 404  9.03277977  5.5334318 Circle
#> 405  8.08774514  5.3463942 Circle
#> 406 11.16015558 12.0241687 Circle
#> 407  8.27087923 10.1136511 Circle
#> 408  7.15927485  7.8572198 Circle
#> 409  9.52816745  7.2161944 Circle
#> 410  5.58208484  9.9563333 Circle
#> 411  5.83084195  7.7416618 Circle
#> 412  5.95320673  6.5594151 Circle
#> 413  7.65938004 10.0767289 Circle
#> 414  7.05564804  6.7371160 Circle
#> 415  9.80493975  9.1788142 Circle
#> 416  8.65669601 11.3645504 Circle
#> 417  6.34264879  8.9640551 Circle
#> 418  9.23966366  8.9935461 Circle
#> 419  8.31466153  9.9134520 Circle
#> 420  6.30866519  7.7886863 Circle
#> 421  6.98605996  8.4767832 Circle
#> 422  5.17763223  8.1009711 Circle
#> 423  6.69877371  9.8243243 Circle
#> 424  4.50542905  7.0820066 Circle
#> 425  8.39620406 10.0889604 Circle
#> 426  5.19826027  6.9938386 Circle
#> 427  8.22562228  7.6220007 Circle
#> 428  4.70707730  5.9950799 Circle
#> 429  5.77549760  7.4965242 Circle
#> 430  8.28522706  8.4908306 Circle
#> 431  7.81288627 10.7215321 Circle
#> 432  5.46972674  8.1667275 Circle
#> 433  7.64792781 10.5460269 Circle
#> 434 10.71191418 10.0186320 Circle
#> 435  9.07586003  7.2484421 Circle
#> 436  9.08581003  9.7771875 Circle
#> 437  7.68381125 10.5973503 Circle
#> 438  7.60467950  7.3369418 Circle
#> 439  8.30753555 10.7925558 Circle
#> 440  8.99361603  7.2366765 Circle
#> 441  9.93726580  8.4895844 Circle
#> 442  7.84653059  9.5151245 Circle
#> 443  7.15764048  5.1160705 Circle
#> 444  9.30406116  6.6846930 Circle
#> 445  6.61141581  7.0448867 Circle
#> 446  4.98717000  7.0753051 Circle
#> 447  7.66081379  7.4220447 Circle
#> 448  8.35292606 10.6783382 Circle
#> 449  7.23469943  6.6466820 Circle
#> 450  6.66234225  9.0875333 Circle
#> 451  5.24343250  8.2006192 Circle
#> 452  9.83181320 10.7918997 Circle
#> 453  8.11544986  6.6416372 Circle
#> 454  6.96413597  4.2478171 Circle
#> 455  9.72099101  8.0803551 Circle
#> 456  8.76051393  8.4253644 Circle
#> 457  7.11822050 11.3313888 Circle
#> 458  7.92061921  8.9502676 Circle
#> 459  7.79522231  5.3746946 Circle
#> 460  6.41976159  8.1807075 Circle
#> 461  6.46403209  8.5079861 Circle
#> 462  8.42594980  8.3907047 Circle
#> 463  8.17901054  8.2219724 Circle
#> 464  6.78209785  8.2185347 Circle
#> 465  8.65641002  8.3496788 Circle
#> 466  8.61406556 10.9291806 Circle
#> 467  9.13398822  6.8124392 Circle
#> 468  7.96315697  8.3636161 Circle
#> 469  8.48414459  4.7290166 Circle
#> 470  5.07973088  6.5050685 Circle
#> 471  7.33484395 10.0423692 Circle
#> 472  4.91735642  5.1141062 Circle
#> 473  8.25952075  8.6415748 Circle
#> 474  9.27916021  8.6171669 Circle
#> 475  8.29473974  9.9123890 Circle
#> 476  7.19506923  8.5710373 Circle
#> 477  4.74218875  9.6721876 Circle
#> 478  6.83161129  9.9573151 Circle
#> 479  8.05602090  5.5480640 Circle
#> 480  8.62051456  6.7861453 Circle
#> 481  7.34128719  5.2454182 Circle
#> 482  9.81522167 11.2982932 Circle
#> 483  8.72905397  9.6521363 Circle
#> 484  4.28926812  8.7770134 Circle
#> 485  4.25310052  7.2321680 Circle
#> 486  6.38837657  9.4133324 Circle
#> 487  8.02339114 10.6720387 Circle
#> 488  7.32092211  6.2999623 Circle
#> 489  4.58956956  9.1823836 Circle
#> 490  8.65510277  7.2467251 Circle
#> 491  7.30865011  9.7003334 Circle
#> 492  8.34579383  5.9925318 Circle
#> 493  6.55880968  7.7731978 Circle
#> 494  5.55256112  9.1202433 Circle
#> 495  7.75268604  9.2569832 Circle
#> 496  5.99549153  6.2405515 Circle
#> 497  5.43135639 11.2917678 Circle
#> 498  5.96425580 10.7390545 Circle
#> 499  9.02350634  9.4271272 Circle
#> 500  8.63779170  6.4793542 Circle
#> 501 11.65130800  8.3883229 Circle
#> 502  5.23370782  6.7666790 Circle
#> 503  4.22010902  7.8871134 Circle
#> 504  8.34782886  8.0948571 Circle
#> 505  8.45762241  7.9207796 Circle
#> 506  5.76828148 10.0990656 Circle
#> 507  9.08705623  8.6658992 Circle
#> 508  6.59143666 10.6274679 Circle
#> 509  9.77291112  6.7747981 Circle
#> 510  8.35289487 10.0996930 Circle
#> 511  8.35908361 11.3314420 Circle
#> 512  6.32411796  6.0792877 Circle
#> 513  6.80925454  6.0888168 Circle
#> 514  6.57920292  9.2028776 Circle
#> 515  4.55242436  6.0852375 Circle
#> 516  8.60794433 10.5641893 Circle
#> 517  8.65670961  9.8527378 Circle
#> 518 10.75313535  9.1104551 Circle
#> 519  8.48773957  8.2975152 Circle
#> 520  7.37220327  8.9600306 Circle
#> 521  8.87866867  7.2564578 Circle
#> 522 10.04754376  6.8739323 Circle
#> 523  8.59642746  5.5534355 Circle
#> 524  5.05536777  9.1445168 Circle
#> 525  8.72559528  9.1075297 Circle
#> 526  8.07207369  9.0614207 Circle
#> 527  8.05707461  9.9373150 Circle
#> 528  9.81274791  9.3184646 Circle
#> 529  8.11234320  8.6460581 Circle
#> 530  8.94420344  9.2275378 Circle
#> 531  9.12389544  9.0167663 Circle
#> 532  6.86061700  9.6229186 Circle
#> 533  8.62159753  7.4421908 Circle
#> 534  9.35038822 10.3385950 Circle
#> 535  6.47622874  9.0866313 Circle
#> 536  6.87709916  7.1326978 Circle
#> 537  6.43017897  8.9344983 Circle
#> 538  8.36245192  5.9014080 Circle
#> 539  6.26878336 10.9473547 Circle
#> 540  8.26768542 10.2279502 Circle
#> 541  7.61615893 10.9175560 Circle
#> 542  7.62593220 10.4045683 Circle
#> 543  5.96922043  7.3026786 Circle
#> 544  6.75664431  7.9628297 Circle
#> 545  6.52178810  9.0331631 Circle
#> 546  7.14635096  8.0228174 Circle
#> 547  7.68610180  9.3277653 Circle
#> 548  9.37284016  9.7927460 Circle
#> 549  7.83267350  8.2441912 Circle
#> 550  5.59202334  6.5184832 Circle
#> 551  6.32998522  8.1481340 Circle
#> 552  7.70894437  8.3951888 Circle
#> 553  8.54241219  9.0506579 Circle
#> 554  6.98558694 10.5573470 Circle
#> 555  9.72687818 10.4243600 Circle
#> 556  8.21231337  7.0984808 Circle
#> 557  9.67805343 10.1949366 Circle
#> 558  8.06250194 10.5935240 Circle
#> 559  6.15190240  6.5478855 Circle
#> 560  6.56201971  6.3763383 Circle
#> 561  6.92295156  9.3481618 Circle
#> 562  6.54065189  9.0287929 Circle
#> 563  7.64114986 10.5405050 Circle
#> 564  7.03875848  8.5911484 Circle
#> 565  5.31623572  6.9390803 Circle
#> 566  9.31488784 10.1593493 Circle
#> 567  7.43397654  9.2329590 Circle
#> 568  7.97964651  7.8088782 Circle
#> 569  9.74801502  8.8801227 Circle
#> 570  8.36090333  8.7459341 Circle
#> 571  7.42729604  9.7774627 Circle
#> 572  7.90078270  9.1586681 Circle
#> 573  4.96832086  7.2504864 Circle
#> 574  8.53252038  6.6742729 Circle
#> 575  9.96846470  8.8473437 Circle
#> 576  9.02222771  8.8732887 Circle
#> 577  9.06504056  7.7857111 Circle
#> 578 10.63347145  8.9285758 Circle
#> 579  7.45902876  4.4656634 Circle
#> 580  8.80616661  9.7748635 Circle
#> 581  6.51549474  7.4832348 Circle
#> 582  9.06622062  9.3254511 Circle
#> 583  8.89394758  9.9676989 Circle
#> 584  5.74294922  8.0704336 Circle
#> 585  5.69755213  8.5791865 Circle
#> 586  6.88018224  7.0993912 Circle
#> 587  5.38233273  8.3484779 Circle
#> 588 10.14395846 11.1823130 Circle
#> 589  8.37654329  8.4348827 Circle
#> 590  6.47244502  8.2613892 Circle
#> 591  6.01842968  9.5904988 Circle
#> 592  8.25477258  5.4250470 Circle
#> 593  8.15564205  5.7278284 Circle
#> 594  5.99041253  7.7267066 Circle
#> 595  5.96100642  9.4570177 Circle
#> 596  6.72059801  6.3966617 Circle
#> 597  7.34420324  6.2093963 Circle
#> 598  6.31478810  6.7089892 Circle
#> 599  9.71472898  6.5889049 Circle
#> 600  7.48030181  6.5159414 Circle
#> 601 10.20094691  9.1875973 Circle
#> 602  7.92837618 10.0186845 Circle
#> 603  5.30040949  7.9951425 Circle
#> 604  7.54534359  6.8590312 Circle
#> 605  7.19276727  9.4035557 Circle
#> 606  7.35830704  8.7803343 Circle
#> 607 10.24949339  7.1914198 Circle
#> 608  6.63006392  8.5080505 Circle
#> 609  9.78833031 10.2784441 Circle
#> 610  7.72444703  6.1063767 Circle
#> 611  7.61516883 11.2245966 Circle
#> 612  8.17117524  6.3997677 Circle
#> 613  8.10506638  6.6118513 Circle
#> 614  7.30899707  7.2045696 Circle
#> 615  9.52608593  9.2130429 Circle
#> 616  8.14605021  7.2654553 Circle
#> 617  9.54075330 10.7106436 Circle
#> 618  4.19169309  7.5445471 Circle
#> 619 10.06194977  8.5346140 Circle
#> 620  7.90572596  7.4593199 Circle
#> 621  6.30620672  8.2396376 Circle
#> 622  8.60503448  7.1306727 Circle
#> 623  4.95300630  6.2381915 Circle
#> 624  5.18127601  9.1502382 Circle
#> 625  9.50785224  8.9298597 Circle
#> 626  6.23844951  9.1286471 Circle
#> 627  6.26370306  9.9297675 Circle
#> 628  8.24509711 10.2786844 Circle
#> 629  6.88233406 11.9336258 Circle
#> 630  7.26923150  9.8839933 Circle
#> 631  6.56750349  8.9705789 Circle
#> 632  5.38408230  6.3621446 Circle
#> 633  8.74596757  6.8579305 Circle
#> 634  5.27803952  6.2145041 Circle
#> 635  7.60126506 10.0665725 Circle
#> 636 11.39841180  9.0643873 Circle
#> 637  9.11963101  8.0438977 Circle
#> 638  7.38103617  6.9359181 Circle
#> 639  9.02920198  7.5682961 Circle
#> 640  8.59497575  5.5339195 Circle
#> 641  9.09818351  8.3010397 Circle
#> 642  8.87861050 10.3092379 Circle
#> 643  9.59638491  7.6088403 Circle
#> 644  8.85121719  5.9720516 Circle
#> 645 10.35017386  6.5041920 Circle
#> 646  7.45358565  5.4248415 Circle
#> 647  7.79580818 10.7949789 Circle
#> 648  6.57691859  7.7944046 Circle
#> 649  7.27499568  4.1626804 Circle
#> 650  7.49710787  4.8545723 Circle
#> 651  8.86218214  7.1513272 Circle
#> 652  8.69988591  8.8404192 Circle
#> 653  8.00330304  6.1463870 Circle
#> 654  5.73162241  8.9002054 Circle
#> 655 10.53938252  6.3289670 Circle
#> 656  7.92702534  5.8834911 Circle
#> 657  5.96026638  8.0839483 Circle
#> 658  5.36355682  8.5737129 Circle
#> 659  8.54746094  7.7391569 Circle
#> 660  6.59542251 10.1361535 Circle
#> 661  9.17566600  5.6370169 Circle
#> 662  4.99732526  7.4827704 Circle
#> 663  3.50237325  6.7034474 Circle
#> 664 10.13455291 10.2352505 Circle
#> 665  8.09039048  9.4221852 Circle
#> 666  7.31083430  6.0672304 Circle
#> 667  7.52129379  9.4100790 Circle
#> 668  8.87856252  9.5974699 Circle
#> 669 10.98603045  7.8698935 Circle
#> 670  5.57227277  9.1404790 Circle
#> 671  6.94625639  8.4560581 Circle
#> 672  7.31556707  7.7316407 Circle
#> 673  7.20075013  9.3586901 Circle
#> 674  8.44255653  9.2145533 Circle
#> 675  7.12608315  9.5020696 Circle
#> 676  5.47730148  9.2593718 Circle
#> 677  5.55571171  5.8726450 Circle
#> 678  8.98858341  8.0592753 Circle
#> 679  6.99387169  9.5817257 Circle
#> 680  5.11120426  8.0667819 Circle
#> 681  4.80182159  8.3421057 Circle
#> 682  4.98406706  6.9527918 Circle
#> 683  7.56339591  5.6718788 Circle
#> 684  6.66086283  8.7382758 Circle

rose(circle_example[, c("x", "y", "class")], var = "class", over_ratio = 0.8)
#>               x           y  class
#> 1    8.72312737  5.79137646   Rest
#> 2    3.10513906 14.44171134   Rest
#> 3   11.85130353  6.42039406   Rest
#> 4    1.90969267  2.98788984   Rest
#> 5   11.92360361  6.71338102   Rest
#> 6    3.09375587  6.26408308   Rest
#> 7   13.67454854  4.79500642   Rest
#> 8    9.97228475 12.18569682   Rest
#> 9    6.25456337  3.90034858   Rest
#> 10  15.94427276  5.44785340   Rest
#> 11  11.80726383 10.58901955   Rest
#> 12  12.17379270 14.09910740   Rest
#> 13   2.49017462 10.50904942   Rest
#> 14  13.21237551  8.59811627   Rest
#> 15   0.63641933  8.28326339   Rest
#> 16   6.62897057 11.22373820   Rest
#> 17   7.91825574  0.37787555   Rest
#> 18  11.65452494  5.20582037   Rest
#> 19   1.50984786  4.13658246   Rest
#> 20   5.60915088 11.61431583   Rest
#> 21   7.17276661 11.87096942   Rest
#> 22   8.51513943  1.20700445   Rest
#> 23   9.62102081 11.34051133   Rest
#> 24   3.39574347  4.30853932   Rest
#> 25  14.62245789  7.81078422   Rest
#> 26   6.25057384  8.97866208   Rest
#> 27   1.91999000  6.47307851   Rest
#> 28  12.75462220 12.56666104   Rest
#> 29  17.64482475 12.95861049   Rest
#> 30  12.87976693  6.43646966   Rest
#> 31   6.52098656 15.20858509   Rest
#> 32  12.58411550 11.52286844   Rest
#> 33   7.86263430 13.07970181   Rest
#> 34  10.49936474  7.30681879   Rest
#> 35   0.89802254  2.54517140   Rest
#> 36   4.55160001 10.32696099   Rest
#> 37  13.29895185  7.82449394   Rest
#> 38   4.64183901 15.91614447   Rest
#> 39   7.12084547 -0.22151723   Rest
#> 40   0.02659684  3.54000405   Rest
#> 41   5.09330663  5.77185511   Rest
#> 42   0.71828946 10.84461995   Rest
#> 43  14.37221790  4.51547460   Rest
#> 44  12.94205815 10.85510722   Rest
#> 45  13.09522523  2.97709282   Rest
#> 46   6.71750945 13.19132557   Rest
#> 47   5.98679541  0.21649831   Rest
#> 48   5.30963752 14.53266224   Rest
#> 49  10.35259346  5.88828327   Rest
#> 50   1.77893516  1.61753439   Rest
#> 51   9.98537476 17.72383371   Rest
#> 52   2.30072752  9.93858381   Rest
#> 53  14.43424132  6.52011143   Rest
#> 54   9.05566870  5.48517418   Rest
#> 55   6.13548302 13.18546101   Rest
#> 56   0.99002542 12.41617758   Rest
#> 57  12.65726672  5.33638777   Rest
#> 58   6.42297283 15.88468885   Rest
#> 59   4.29257483  5.03759084   Rest
#> 60   1.26305669 10.95136793   Rest
#> 61  11.52175848  7.82270159   Rest
#> 62   5.00553349 20.05467441   Rest
#> 63  10.75878700  9.55148438   Rest
#> 64  10.62874718  9.38375246   Rest
#> 65  10.07245025 13.87203304   Rest
#> 66  15.66504281  7.31720126   Rest
#> 67   4.53877796  3.29698512   Rest
#> 68   1.59462579  4.22235606   Rest
#> 69   6.86193954 10.58853805   Rest
#> 70  13.30909327  9.29432483   Rest
#> 71  12.97909991  1.64394533   Rest
#> 72  13.73543173  4.69389085   Rest
#> 73   3.97241333 12.49263532   Rest
#> 74  11.10652124 15.26769852   Rest
#> 75   1.09155566 -1.12673701   Rest
#> 76   7.47887527 -0.07869004   Rest
#> 77   2.61268232 10.31901014   Rest
#> 78   5.95942257 12.40885675   Rest
#> 79   4.13014179  0.84381130   Rest
#> 80  16.47028442 16.42440326   Rest
#> 81   3.15311287  6.15865099   Rest
#> 82   9.76087160 14.07823037   Rest
#> 83  12.17780681 13.99951768   Rest
#> 84  17.42439291  3.67143472   Rest
#> 85  14.23290748 10.64252467   Rest
#> 86   9.60772220 14.98676195   Rest
#> 87   6.29473144  3.07620214   Rest
#> 88  10.37069037  2.17918819   Rest
#> 89  10.11267759 13.69087758   Rest
#> 90   2.62505913  1.88508861   Rest
#> 91   5.87006817 -0.53311540   Rest
#> 92  13.73633189  5.88031097   Rest
#> 93  11.50418141  9.27274593   Rest
#> 94  14.51221686 11.62083997   Rest
#> 95   2.28536279 12.60287044   Rest
#> 96   7.02444117 12.82938408   Rest
#> 97   3.74225207  1.85617234   Rest
#> 98  16.31204714  4.44160739   Rest
#> 99   2.26055091 11.26099217   Rest
#> 100 -0.13207042  1.87487087   Rest
#> 101 10.39720002  9.40328406   Rest
#> 102 12.97991257 10.28007817   Rest
#> 103 11.26029505  9.85307009   Rest
#> 104  2.62222360  5.84150063   Rest
#> 105  9.26436329 11.72429055   Rest
#> 106  4.35101421  1.05771875   Rest
#> 107 13.55832713  4.48946176   Rest
#> 108 10.31009457  8.51908281   Rest
#> 109  2.92982156  4.40307572   Rest
#> 110  5.33892885 14.44170604   Rest
#> 111  4.87538984 13.06896158   Rest
#> 112  9.95206046  6.07840211   Rest
#> 113 12.58680953  4.57632037   Rest
#> 114  6.90593303  5.49437080   Rest
#> 115 14.36617791  7.34475084   Rest
#> 116 14.06759779 14.91181271   Rest
#> 117 12.57869929  3.57364525   Rest
#> 118  6.77693710 12.40318013   Rest
#> 119  5.14129179  1.26485806   Rest
#> 120 13.64971270  9.87451197   Rest
#> 121 -0.37723281 15.72679260   Rest
#> 122 -0.28075277 10.64618192   Rest
#> 123  3.96554064  2.45082371   Rest
#> 124  5.09734345  8.18701988   Rest
#> 125 12.46437835  3.85508943   Rest
#> 126  6.62696680  1.90621580   Rest
#> 127  5.49151011  3.75537768   Rest
#> 128  8.41361821 12.05930250   Rest
#> 129  8.77163140 13.63913693   Rest
#> 130 11.41172201  3.61549081   Rest
#> 131  6.93590598 15.54842993   Rest
#> 132 16.57855342 11.11171088   Rest
#> 133 12.71143261  9.44184086   Rest
#> 134 12.41205186  1.69994087   Rest
#> 135 12.34847441  3.30089996   Rest
#> 136  0.99011671 10.09914730   Rest
#> 137 15.56999465 11.74029728   Rest
#> 138  3.60828572 14.76886585   Rest
#> 139 13.24503563 13.38945105   Rest
#> 140 10.45887261  8.03030981   Rest
#> 141  3.16984859 12.83995009   Rest
#> 142 15.26463634  4.78956588   Rest
#> 143  4.07692451 13.07218312   Rest
#> 144  3.23820497 14.88093054   Rest
#> 145  5.41520138 11.41640123   Rest
#> 146  3.99282226 15.27055228   Rest
#> 147 13.57011910 14.07801099   Rest
#> 148 11.02524457  5.99907751   Rest
#> 149  8.17211708  5.90385182   Rest
#> 150  7.37270517  3.22175182   Rest
#> 151  6.30596976 13.09065712   Rest
#> 152  4.26741032  1.40358276   Rest
#> 153  7.48912284  4.01039689   Rest
#> 154  2.33264428 14.27726279   Rest
#> 155  7.61690391 -0.85308311   Rest
#> 156  6.59333980  4.45928475   Rest
#> 157  4.79504291 13.44087587   Rest
#> 158 14.86987855  8.52391827   Rest
#> 159 11.67409845 11.58444726   Rest
#> 160 12.29356242  3.52212718   Rest
#> 161  5.92746984 15.54252290   Rest
#> 162  7.70943065 12.90995424   Rest
#> 163  2.42693966  2.56047510   Rest
#> 164 15.69573791  7.41965965   Rest
#> 165  2.78181257 10.54946138   Rest
#> 166  1.50897508  6.94078745   Rest
#> 167  2.77334862 13.64415074   Rest
#> 168  7.44794401  2.64868401   Rest
#> 169  7.76451270  9.98939066   Rest
#> 170  6.84018490 13.73632569   Rest
#> 171  3.37511571 11.36019885   Rest
#> 172 16.03105937  6.98520703   Rest
#> 173 16.43445986  5.63696219   Rest
#> 174 15.02226218  2.10016323   Rest
#> 175  1.73717055 10.51376448   Rest
#> 176 13.77443852  3.38620375   Rest
#> 177 15.37783712  2.71516945   Rest
#> 178 12.94096091  5.65680021   Rest
#> 179 11.66441205 10.23166121   Rest
#> 180 11.52837626 13.18188442   Rest
#> 181  8.67143144 11.49525347   Rest
#> 182  8.61315855 12.02455056   Rest
#> 183 14.02240378  2.42171409   Rest
#> 184  1.93728618  4.40827644   Rest
#> 185  5.16353964 13.41307150   Rest
#> 186 12.80412231  9.69207534   Rest
#> 187  1.19456089  2.23041341   Rest
#> 188  3.94117253 13.83168040   Rest
#> 189 14.83802055  9.45019922   Rest
#> 190 -0.04646854  6.15683535   Rest
#> 191 11.43518708 11.50628618   Rest
#> 192  1.35592915  9.36835096   Rest
#> 193 14.68588716  4.67100688   Rest
#> 194  1.29238247 13.81052742   Rest
#> 195  1.67320753  3.47655663   Rest
#> 196  9.37050984 12.76532375   Rest
#> 197  9.33376792  6.64685193   Rest
#> 198  7.10284513 10.55898218   Rest
#> 199  5.22492143 14.38058731   Rest
#> 200 10.52557999 12.49563636   Rest
#> 201 11.08222492 13.81556539   Rest
#> 202  3.34617390  4.38979807   Rest
#> 203  4.03669639 14.16754328   Rest
#> 204 10.92101101  3.83796215   Rest
#> 205  9.66051165  3.91371476   Rest
#> 206  8.51020334 14.42914294   Rest
#> 207 13.71352779  5.96035383   Rest
#> 208  4.36730300 13.81540676   Rest
#> 209  3.10046555  4.46380390   Rest
#> 210 12.42381618  0.21917744   Rest
#> 211  3.27844541 11.21490048   Rest
#> 212  1.75904375  9.94219000   Rest
#> 213 10.69005105  3.14084775   Rest
#> 214  2.61531771 13.20940174   Rest
#> 215 15.44304072  5.46313459   Rest
#> 216  4.32842519  8.89651891   Rest
#> 217 11.82999680  3.44082561   Rest
#> 218  3.91407395  8.03427969   Rest
#> 219  2.22748551  5.86461670   Rest
#> 220  5.49931857 13.83891153   Rest
#> 221 15.39955849  9.43261242   Rest
#> 222 13.42670646  5.24595436   Rest
#> 223  3.66907305 10.14556118   Rest
#> 224 10.80720071  5.62031526   Rest
#> 225 12.83370179  7.74018142   Rest
#> 226  7.24127439 13.29610746   Rest
#> 227  4.31729384  2.33406641   Rest
#> 228 13.01983169 14.49464726   Rest
#> 229 11.50442360  3.41788216   Rest
#> 230 10.78545784  3.32265025   Rest
#> 231 12.74940020 16.26137692   Rest
#> 232  6.84729253  4.03339884   Rest
#> 233  7.30210569 14.94454935   Rest
#> 234 11.27294257  9.52950023   Rest
#> 235  1.14750459  7.72746181   Rest
#> 236  9.22049464 10.24666854   Rest
#> 237 16.20955595  5.63623337   Rest
#> 238  6.09840352 13.46811000   Rest
#> 239  2.93678065  7.77038544   Rest
#> 240  5.46133737 11.08810374   Rest
#> 241 10.62012776 13.91539624   Rest
#> 242  6.95795618  3.81278611   Rest
#> 243 12.96280212  5.10654118   Rest
#> 244  7.88782631 16.10494177   Rest
#> 245 11.31954604  9.21331063   Rest
#> 246  2.69309024 -0.52230548   Rest
#> 247  2.29500731 10.47888755   Rest
#> 248 10.33890292  9.99320438   Rest
#> 249 12.83697985  5.01429196   Rest
#> 250  9.21224444  6.50405990   Rest
#> 251  5.01111671  3.79820746   Rest
#> 252  4.34661081  6.23533352   Rest
#> 253  4.11838047  3.47663957   Rest
#> 254 11.18812765 15.83151202   Rest
#> 255 13.35140290 15.40252215   Rest
#> 256  3.24545683 12.05606681   Rest
#> 257  2.14018637  1.29929076   Rest
#> 258 13.23590204  8.71716294   Rest
#> 259 12.90125530  6.97139102   Rest
#> 260 10.69197651 11.61840348   Rest
#> 261  8.50658158 14.01184266   Rest
#> 262  7.62591222  0.67380881   Rest
#> 263  3.07899449  4.67191637   Rest
#> 264  4.77262528 10.90255609   Rest
#> 265 -0.43148887 10.35863334   Rest
#> 266  3.65589939  4.95955603   Rest
#> 267  2.46067532 12.46750506   Rest
#> 268  3.53610240 -0.66714311   Rest
#> 269  6.43233519 15.38532691   Rest
#> 270 10.48775754  1.43445964   Rest
#> 271  2.64675898  3.93012121   Rest
#> 272  5.33748042  6.95929703 Circle
#> 273  9.33440148  7.94011246 Circle
#> 274  7.71898762  7.95435983 Circle
#> 275  5.09415336  8.10265594 Circle
#> 276  6.00626358  7.78650360 Circle
#> 277  7.92606535  7.77319456 Circle
#> 278  6.98261908  9.97134737 Circle
#> 279  8.98374447 10.79328882 Circle
#> 280  6.06151168  7.98150388 Circle
#> 281  8.76765323 10.30038179 Circle
#> 282  4.94740659  6.41869058 Circle
#> 283  8.81397198 10.72301859 Circle
#> 284  6.03813267  9.32431191 Circle
#> 285  6.59579202  8.13320098 Circle
#> 286  9.30975777  7.12821515 Circle
#> 287  8.77486985  8.20294197 Circle
#> 288 10.75335588  5.81844013 Circle
#> 289  9.85843201  6.97502468 Circle
#> 290  5.23550261  8.36571248 Circle
#> 291  4.33606769  7.35119960 Circle
#> 292  7.26896953  6.34940075 Circle
#> 293  9.63639515  5.77533330 Circle
#> 294  7.53527029 10.06797177 Circle
#> 295 10.63717576 10.65888309 Circle
#> 296  5.02155291  7.78232211 Circle
#> 297  9.11618146  5.19106698 Circle
#> 298  9.30124659  9.01971133 Circle
#> 299  7.33081375 10.76710766 Circle
#> 300  7.85683958  6.75564512 Circle
#> 301  6.52142337  7.08498337 Circle
#> 302  7.64420132  7.67319243 Circle
#> 303  6.53830163  8.96709599 Circle
#> 304  7.86979218  8.55809418 Circle
#> 305  7.87350152  7.59616395 Circle
#> 306  8.97475318  9.83155086 Circle
#> 307  6.67854480  6.30918912 Circle
#> 308  8.82680962 10.63281497 Circle
#> 309  7.31273510  9.71346665 Circle
#> 310  6.55101655  7.50565822 Circle
#> 311  7.24874051  8.27721424 Circle
#> 312  7.95245694  6.95737597 Circle
#> 313  9.16177342  5.84188682 Circle
#> 314  8.94161988 10.71917015 Circle
#> 315  7.99971943  6.79603729 Circle
#> 316  7.89214968  9.93781391 Circle
#> 317  8.86875718  6.46076619 Circle
#> 318  5.98074295  8.43500807 Circle
#> 319  7.83059779  5.54648635 Circle
#> 320  7.40843013  7.95541196 Circle
#> 321  6.21509305  6.23834472 Circle
#> 322  8.35052090  8.72697256 Circle
#> 323  6.48728371  8.79150779 Circle
#> 324  6.63306678  8.98622050 Circle
#> 325  6.19084816  7.60465105 Circle
#> 326  6.31784348  7.27068336 Circle
#> 327  6.11212176  7.14877785 Circle
#> 328  5.07388799  6.60962869 Circle
#> 329  8.62580164  9.38280468 Circle
#> 330  6.74329705  7.03328288 Circle
#> 331  9.67364162 10.24251783 Circle
#> 332  6.90432989 10.77343649 Circle
#> 333  9.67515211  7.43241426 Circle
#> 334 11.08883642  8.89237202 Circle
#> 335 10.17780290 11.09098009 Circle
#> 336  7.36797408  7.63361489 Circle
#> 337  8.06875274  9.43690380 Circle
#> 338  9.02443588  5.52934080 Circle
#> 339  7.61909187  9.67279318 Circle
#> 340  7.61419335  7.25040129 Circle
#> 341 10.09878219  5.81940493 Circle
#> 342  6.02162792  9.63673929 Circle
#> 343  4.62954459  7.82214517 Circle
#> 344  8.65608393 10.55074564 Circle
#> 345  8.03216808 11.34151457 Circle
#> 346  7.77273439  9.99240202 Circle
#> 347  9.30859541 10.11886650 Circle
#> 348  5.62759607  7.21433461 Circle
#> 349  9.21583151  6.50559115 Circle
#> 350  7.19044131  8.17695995 Circle
#> 351  8.24526248  6.12893419 Circle
#> 352  5.17069022  9.56365292 Circle
#> 353  5.60145830  9.03584580 Circle
#> 354  9.11156789 10.54544618 Circle
#> 355  6.97648545  6.58535153 Circle
#> 356  7.51961978 10.99060654 Circle
#> 357  7.32261737  8.04195228 Circle
#> 358  7.65779339 10.81677397 Circle
#> 359 11.14778769  7.75772147 Circle
#> 360  9.62936808  9.02099069 Circle
#> 361  6.81336248 10.60567138 Circle
#> 362 10.10644071  7.95776246 Circle
#> 363  7.32626577  8.22615782 Circle
#> 364  5.97121760  9.15845222 Circle
#> 365  7.87779159  9.35681356 Circle
#> 366  7.89780936  8.23721581 Circle
#> 367  6.40086183  8.45413959 Circle
#> 368  8.77474586  8.89813723 Circle
#> 369  7.94403610  5.95055238 Circle
#> 370  7.80444988  8.15463414 Circle
#> 371  7.34740776  8.94143364 Circle
#> 372  8.66553776  7.72779708 Circle
#> 373  9.79951568  6.53101940 Circle
#> 374  6.51471657  9.83982156 Circle
#> 375  7.12712647  7.37353902 Circle
#> 376  4.62199635  5.11958261 Circle
#> 377  5.78632553  6.62898518 Circle
#> 378  6.11179569 10.39804431 Circle
#> 379 10.75918783 11.00614070 Circle
#> 380  7.27625003  6.56596021 Circle
#> 381  9.62938400  8.86420980 Circle
#> 382 10.38722990  7.71507873 Circle
#> 383  8.70533295  9.56210966 Circle
#> 384  6.86041471  7.53088326 Circle
#> 385 10.23003303  6.76383998 Circle
#> 386  8.10992204 11.27933174 Circle
#> 387 10.30567379  7.47223139 Circle
#> 388  9.96441244  9.94740340 Circle
#> 389  7.12553435 10.11785940 Circle
#> 390  8.13122846  9.97156567 Circle
#> 391  6.54354162 10.35384682 Circle
#> 392  9.45549455 10.39709094 Circle
#> 393  7.86883851  9.72832934 Circle
#> 394  7.73453762  8.26159022 Circle
#> 395  6.57530081 10.19246734 Circle
#> 396  6.99029380  7.06918958 Circle
#> 397  7.16490148  6.33199249 Circle
#> 398  6.50535476  7.17723480 Circle
#> 399  7.27522223  8.63744143 Circle
#> 400  7.75485796  7.15781953 Circle
#> 401  5.55515348  9.58415928 Circle
#> 402 10.03486773  8.12889995 Circle
#> 403  7.03509921  8.09213591 Circle
#> 404  8.82238300  9.18895271 Circle
#> 405 10.51617979  9.54197379 Circle
#> 406  8.06006780  8.30773660 Circle
#> 407  5.22193722  6.52515888 Circle
#> 408  7.55708507  4.59241114 Circle
#> 409  6.90390000 11.34695616 Circle
#> 410  9.76685427  9.90404041 Circle
#> 411  8.84740144  7.60840650 Circle
#> 412  6.07822296  7.99848007 Circle
#> 413  6.65107777  6.65794932 Circle
#> 414 10.53833614  6.40926915 Circle
#> 415  8.46552415  6.13554432 Circle
#> 416  6.70535503  7.78009717 Circle
#> 417  3.85086640  5.86583272 Circle
#> 418  8.37978895  9.07350640 Circle
#> 419  6.23787955  5.16453573 Circle
#> 420  5.37811940  7.14199810 Circle
#> 421  5.62105361  7.34994102 Circle
#> 422  9.39105902  9.32915410 Circle
#> 423  8.77021893  9.30122951 Circle
#> 424  6.73636801  7.53657290 Circle
#> 425  7.15452454  9.20909256 Circle
#> 426  5.92410437  8.96410718 Circle
#> 427  4.78410898  7.17380943 Circle
#> 428  6.29447527  7.03308086 Circle
#> 429  7.03034598  5.49718743 Circle
#> 430 11.03113269 11.18073466 Circle
#> 431  8.27735230  7.20141671 Circle
#> 432  6.57905934  8.26837223 Circle
#> 433  6.22629467  5.07067577 Circle
#> 434  5.94871162  7.60993243 Circle
#> 435 10.30309405 10.27274708 Circle
#> 436  5.33349388  7.26051636 Circle
#> 437  5.50544775  7.93637786 Circle
#> 438  9.91172959  6.94564173 Circle
#> 439 11.25125003  9.09085881 Circle
#> 440  8.03988879  9.08795054 Circle
#> 441  9.74509750 11.45676420 Circle
#> 442  8.76970079  9.67180604 Circle
#> 443  5.29965906  6.24726845 Circle
#> 444  9.68746213  6.06384142 Circle
#> 445  6.54901326  9.64134168 Circle
#> 446  7.50797970  6.10343575 Circle
#> 447  6.59000148  9.63110731 Circle
#> 448 10.70491508  8.70874350 Circle
#> 449  5.48468690  8.07708835 Circle
#> 450  6.34611064  9.05268710 Circle
#> 451  8.46459742  5.96941330 Circle
#> 452  9.24719110  8.79322227 Circle
#> 453  9.56089375  6.75835928 Circle
#> 454  8.29880515 11.72836803 Circle
#> 455  5.90016741  6.70573544 Circle
#> 456  9.74872006  8.97268799 Circle
#> 457  8.32642013  6.77873242 Circle
#> 458  9.86441668  7.42344159 Circle
#> 459  7.27628609  9.10321804 Circle
#> 460  8.04327192 10.73962729 Circle
#> 461  6.57654012  9.22187533 Circle
#> 462  7.17623219  6.14963919 Circle
#> 463  5.86351431  9.64559767 Circle
#> 464  9.39457872  6.21493862 Circle
#> 465  7.64164455  5.76992367 Circle
#> 466  6.99388767  9.16508398 Circle
#> 467  7.34047764  7.61473056 Circle
#> 468  5.11311699  9.23203150 Circle
#> 469  9.03792662  8.16181249 Circle
#> 470  6.15759186  8.85760324 Circle
#> 471  7.72494009  9.52751630 Circle
#> 472  7.46460345  9.23556668 Circle
#> 473  7.43547928 10.91448839 Circle
#> 474  7.91006362  5.30418427 Circle
#> 475  8.87301814  6.25656422 Circle
#> 476  9.94775655  7.50649490 Circle
#> 477  5.57386141  8.10331168 Circle
#> 478  7.00422007  7.02712893 Circle
#> 479  6.81363865  9.72609913 Circle
#> 480  5.84055632  6.66735683 Circle
#> 481  6.17281036  7.27817398 Circle
#> 482  8.20731306  6.32091381 Circle
#> 483  9.81551677  8.26885481 Circle
#> 484  8.73506267 11.24451860 Circle
#> 485  4.93664152  8.27902716 Circle
#> 486  8.82147532 10.72433750 Circle
#> 487  9.30742392  7.27091561 Circle
#> 488 10.04704214  9.92961942 Circle
#> 489  7.13558888 11.08451574 Circle
#> 490  8.13869441 10.15348772 Circle
#> 491  6.76423170  5.50984942 Circle
#> 492  8.55645814  4.74291166 Circle
#> 493  8.36732365 10.46328197 Circle
#> 494  8.29608222 10.46067913 Circle
#> 495  6.75996117  5.86492579 Circle
#> 496  6.30044564  7.29763695 Circle
#> 497  5.69776628  9.26782967 Circle
#> 498  7.12808714  9.77509793 Circle
#> 499  8.24216410  6.04756673 Circle
#> 500  8.06512566 10.06698757 Circle
#> 501  9.10811017 10.43443430 Circle
#> 502 10.03373927  6.32917491 Circle
#> 503  5.96299831  5.97781208 Circle
#> 504  5.79108977  9.95989200 Circle
#> 505 10.64279687  5.43237370 Circle
#> 506  5.15013324  8.00581522 Circle
#> 507  8.41944235  8.14345632 Circle
#> 508  7.89443409 10.50312885 Circle
#> 509  7.14051378  9.36261829 Circle
#> 510  5.71970142  7.92749749 Circle
#> 511  9.06224723 10.71322314 Circle
#> 512  6.79158976  8.58122454 Circle
#> 513  7.38620405  9.25247973 Circle
#> 514 10.17729430  8.83225526 Circle
#> 515  5.95845900  9.76961037 Circle
#> 516  9.61781959  7.21890723 Circle
#> 517  6.91584446  8.99068319 Circle
#> 518  5.73397061  7.94675519 Circle
#> 519  3.14647420  8.82813728 Circle
#> 520  8.27198354 10.61831857 Circle
#> 521  6.83398658  8.40085173 Circle
#> 522  8.68732859  9.49558272 Circle
#> 523  5.91330496  7.41919696 Circle
#> 524  7.40978241  5.54454679 Circle
#> 525  6.91867911  7.15773611 Circle
#> 526  8.12419725 10.96500357 Circle
#> 527 11.32187482  3.85395454 Circle
#> 528  9.68862031  8.73801493 Circle
#> 529  8.75171677  7.20911551 Circle
#> 530  6.20287326  8.16798466 Circle
#> 531  8.76024577  8.99517288 Circle
#> 532 11.45125889  9.98242925 Circle
#> 533  5.58164988  8.56674759 Circle
#> 534  6.91638673  9.09021926 Circle
#> 535  7.90473946  5.14962850 Circle
#> 536  9.89216880  9.87986524 Circle
#> 537  6.98781384  9.66583789 Circle
#> 538  9.09264593  9.90103995 Circle
#> 539  6.17441273  8.16352416 Circle
#> 540  7.60980320  7.58836804 Circle
#> 541  8.44047263  7.26716419 Circle
#> 542  5.17233339  7.84367689 Circle
#> 543  7.22795867 10.35963302 Circle
#> 544  9.94597632  9.61606236 Circle
#> 545  5.83795562  7.20868814 Circle
#> 546  8.04891656  8.07032354 Circle
#> 547  6.14323051  7.45038566 Circle