ROSE Algorithm

A thin wrapper around ROSE::ROSE() that generates a synthetic balanced sample by enlarging the feature space of minority and majority class examples.

Usage

rose(
  df,
  var,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1
)

Arguments

df

A data.frame or tibble. Must have 1 factor variable with exactly 2 levels and remaining numeric variables.

var

Character, name of variable containing the 2-level factor variable.

over_ratio

A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows as the majority level. See vignette("ratio", package = "themis") for more details.

minority_prop

A numeric value between 0 and 1 for the proportion of synthetic observations from the minority class. Defaults to 0.5, which generates an equal split of minority and majority synthetic observations. This parameter controls the class balance within the synthetic data, while over_ratio controls the total size of the synthetic data.

minority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the minority class. Defaults to 1.

majority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the majority class. Defaults to 1.

Value

A data.frame or tibble, depending on type of df.

Details

The factor variable used to balance around must only have 2 levels.

The ROSE algorithm works by selecting an observation belonging to class k and generating new examples in its neighborhood, which is determined by a smoothing matrix H_k. Smaller values of minority_smoothness and majority_smoothness shrink the entries of H_k, producing tighter neighborhoods. This is a cautious choice when there is a concern that excessively large neighborhoods could blur the boundaries between classes.

This function is a thin wrapper around ROSE::ROSE(). For full details on the underlying implementation, see that function's documentation.

References

Lunardon, N., Menardi, G., and Torelli, N. (2014). ROSE: a Package for Binary Imbalanced Learning. R Journal, 6:79–89.

Menardi, G. and Torelli, N. (2014). Training and assessing classification rules with imbalanced data. Data Mining and Knowledge Discovery, 28:92–122.

See also

Examples

rose(circle_example[, c("x", "y", "class")], var = "class")
#>               x          y  class
#> 1    4.03117759  2.3546472   Rest
#> 2   10.65673525 -0.9825904   Rest
#> 3    4.88121450  1.0988890   Rest
#> 4    0.50858964 16.0065459   Rest
#> 5    1.03557834  8.9282826   Rest
#> 6   13.44741685 12.4850085   Rest
#> 7    1.04096632  8.9728543   Rest
#> 8    4.57256612  4.7586111   Rest
#> 9   13.92232557  8.7763772   Rest
#> 10   7.12984787  1.8936658   Rest
#> 11  12.46274304 11.7988532   Rest
#> 12  14.08726061  8.3594358   Rest
#> 13   3.98860210 11.2418495   Rest
#> 14  11.71719697 15.3383009   Rest
#> 15   5.35241076 13.4596426   Rest
#> 16   0.20033399  1.5449734   Rest
#> 17   6.16054468  2.8063120   Rest
#> 18   8.36489003  8.3986197   Rest
#> 19   1.77407245 11.2693437   Rest
#> 20   4.60360423 11.9146623   Rest
#> 21  -0.08200332  1.4375947   Rest
#> 22   9.61306398 14.7592404   Rest
#> 23  11.50838074  4.2486238   Rest
#> 24   7.13979612 12.3057376   Rest
#> 25   6.68186518 12.0912922   Rest
#> 26   8.96372079 12.5609355   Rest
#> 27   1.07270516 12.6740509   Rest
#> 28   3.33254508  7.6606700   Rest
#> 29  12.51998020  2.9233476   Rest
#> 30  11.10953724 17.1747546   Rest
#> 31   0.86455130  4.3165739   Rest
#> 32  13.58743311  7.3389929   Rest
#> 33   0.66350104 12.0119776   Rest
#> 34   1.14602316  9.0955228   Rest
#> 35   7.83078002 12.3867677   Rest
#> 36  15.12414998 12.3296736   Rest
#> 37  10.68517570 13.1115259   Rest
#> 38  11.44237207 16.0604988   Rest
#> 39   6.51099909 12.1662757   Rest
#> 40   8.11813393 11.0648984   Rest
#> 41  17.00213931  9.1139693   Rest
#> 42  16.08464291 12.4080724   Rest
#> 43   7.42653506  4.8116331   Rest
#> 44   6.00526855 16.3228510   Rest
#> 45  11.21239248  3.0740553   Rest
#> 46   2.40381302 13.5489167   Rest
#> 47  -0.34658682 13.5833682   Rest
#> 48  10.47673513 13.0561619   Rest
#> 49  11.13929191  5.9156751   Rest
#> 50  14.91932818  8.5880202   Rest
#> 51   4.88956598  0.4177354   Rest
#> 52   4.29771711 10.1914863   Rest
#> 53   9.91312692 17.8711841   Rest
#> 54  14.77095775  8.1714694   Rest
#> 55   9.89055128 15.0671714   Rest
#> 56  -0.29489044  2.1788588   Rest
#> 57  13.25502266  5.6465042   Rest
#> 58   9.24223236  6.0824009   Rest
#> 59  13.44079092 10.2305708   Rest
#> 60  14.37054259  7.6387842   Rest
#> 61   9.03740420 12.9912489   Rest
#> 62  14.32821879  8.9097263   Rest
#> 63  16.23595127 14.2691825   Rest
#> 64   1.07473419 10.6569270   Rest
#> 65   1.78647348 11.0377370   Rest
#> 66   1.48579324  5.6054967   Rest
#> 67   8.38073889 12.9113329   Rest
#> 68   4.56387096 10.7304792   Rest
#> 69   4.57648794 14.3970903   Rest
#> 70   7.50028265 10.4801150   Rest
#> 71   4.61334310 12.0032770   Rest
#> 72   5.12353940  1.9057132   Rest
#> 73  16.50464113  7.9415312   Rest
#> 74   7.79635675  4.1034775   Rest
#> 75  12.97570801  1.4552005   Rest
#> 76   5.42192053  1.2143713   Rest
#> 77   5.35874280 12.3014801   Rest
#> 78  -1.03048436  0.7551682   Rest
#> 79   8.53406036  8.2746832   Rest
#> 80  13.93315546 12.7531097   Rest
#> 81   6.79262058 12.1893547   Rest
#> 82   7.95210289  3.6113032   Rest
#> 83  10.92839969  9.1721475   Rest
#> 84   6.11958135  3.5840580   Rest
#> 85  14.97396853 15.0231244   Rest
#> 86  -0.29378080 10.7371017   Rest
#> 87   6.86673924  1.8073151   Rest
#> 88   6.67419628 13.8325889   Rest
#> 89  14.32417248 10.1911924   Rest
#> 90   9.07859582  0.5528688   Rest
#> 91  18.94824756 16.6997092   Rest
#> 92   4.31407109 16.9935398   Rest
#> 93   8.36001052 16.8115691   Rest
#> 94   9.76666147  4.7858177   Rest
#> 95   5.34696999  2.6250222   Rest
#> 96  10.81000290 15.1645001   Rest
#> 97   3.71030414  5.7937074   Rest
#> 98   9.49691970  1.2718170   Rest
#> 99  14.98368804  8.1837518   Rest
#> 100  3.23290144  2.9531398   Rest
#> 101  5.91560178  7.0040160   Rest
#> 102  7.41730049  7.3417308   Rest
#> 103 11.22337386  9.8941530   Rest
#> 104  9.30047234  9.8956517   Rest
#> 105  1.30879691  6.5034482   Rest
#> 106  5.63390586  9.7018752   Rest
#> 107 12.19205291  8.6174255   Rest
#> 108  2.67480329  5.0747594   Rest
#> 109 11.07333259  9.3302068   Rest
#> 110  2.28397966 14.6166673   Rest
#> 111  2.81532949  5.7831955   Rest
#> 112  8.12724363 10.8039685   Rest
#> 113  3.79290774  7.4905759   Rest
#> 114 13.72454870  4.4233947   Rest
#> 115  8.49726319 14.5220435   Rest
#> 116  0.54964986  9.2603434   Rest
#> 117  4.80446731  7.4966168   Rest
#> 118  4.60310225 12.9102808   Rest
#> 119 15.13345825  5.1129349   Rest
#> 120 13.76906947  0.3258199   Rest
#> 121  1.78865699  5.7004193   Rest
#> 122  9.97067478 15.4293628   Rest
#> 123 13.43925649 14.0196421   Rest
#> 124 13.05791610  1.5255141   Rest
#> 125  3.56120863  2.8389789   Rest
#> 126  8.94830817  2.4264890   Rest
#> 127  6.68737225  0.4805506   Rest
#> 128 11.76703340  5.7776106   Rest
#> 129  3.49380769  4.9999319   Rest
#> 130 12.28037639  6.6287412   Rest
#> 131  3.97641684  1.1003651   Rest
#> 132  8.53011869  2.5472700   Rest
#> 133  9.63695553 14.3060498   Rest
#> 134  5.95412931 15.6581865   Rest
#> 135 13.04126124  3.3783299   Rest
#> 136  1.80254920 10.5389705   Rest
#> 137  6.15962181  6.7598462   Rest
#> 138 10.59653544  6.5345135   Rest
#> 139  3.94972106  4.4443190   Rest
#> 140 15.12101346 11.0554452   Rest
#> 141 12.80072269 13.3175478   Rest
#> 142 13.61004260  4.0106497   Rest
#> 143 14.98648318 13.6395582   Rest
#> 144  7.89668604 11.5924310   Rest
#> 145  1.29116285 13.4331771   Rest
#> 146 12.50147559 12.9447669   Rest
#> 147  3.27777017 13.2730680   Rest
#> 148  1.95266400  8.5976017   Rest
#> 149 13.16636955 13.0786912   Rest
#> 150  7.45340083  6.9953604   Rest
#> 151  7.13309807 12.0444378   Rest
#> 152 12.24331739  9.9996965   Rest
#> 153  7.98630270  8.4120183   Rest
#> 154  8.40698376  3.4778867   Rest
#> 155  1.87687530 11.6393327   Rest
#> 156 12.62695439  6.2082873   Rest
#> 157 12.45149318  8.9576677   Rest
#> 158 10.23360348 15.0736417   Rest
#> 159 11.54819762 15.6239306   Rest
#> 160  6.00754001  5.4990440   Rest
#> 161  5.31777031  2.4282640   Rest
#> 162 12.84331518 13.6952417   Rest
#> 163  6.12116636 10.8028543   Rest
#> 164  7.57360657 10.1317622   Rest
#> 165  8.64586979 13.6949439   Rest
#> 166  2.25170136 15.0595103   Rest
#> 167 16.02022846 15.7252319   Rest
#> 168 13.51884349  5.8881988   Rest
#> 169  7.22181015  1.6776413   Rest
#> 170  5.06537160  9.2045970   Rest
#> 171  7.69310485  2.0632420   Rest
#> 172  7.00610767 14.5672991   Rest
#> 173 13.12611827  0.7399264   Rest
#> 174  9.56465972  2.5614039   Rest
#> 175 11.11405982 11.7713217   Rest
#> 176  8.20890487  2.4778575   Rest
#> 177 10.13786809 12.3155456   Rest
#> 178  2.03823319 -2.8556559   Rest
#> 179 11.77174615  7.1631426   Rest
#> 180 15.69058388  9.6659237   Rest
#> 181  7.02208465 18.1880190   Rest
#> 182 11.33708201  2.6111836   Rest
#> 183 12.29901270  5.2790420   Rest
#> 184  7.30707224 12.2372761   Rest
#> 185  2.83612837 11.6308964   Rest
#> 186 16.32301948 12.7874954   Rest
#> 187  7.20510679 10.3087171   Rest
#> 188  9.14857252 12.6827556   Rest
#> 189  0.73302931  1.3509453   Rest
#> 190  4.74576720  5.3007840   Rest
#> 191 11.80795073  9.8944989   Rest
#> 192  1.15746402  8.2602731   Rest
#> 193  3.21187182  8.4797755   Rest
#> 194  3.42212865  2.3858205   Rest
#> 195  4.06109618  8.5655922   Rest
#> 196  3.78918107 17.3957209   Rest
#> 197  5.03432145  9.4529344   Rest
#> 198  7.06558643 15.0162821   Rest
#> 199 15.24070745  6.1447431   Rest
#> 200  4.33566663  6.2186860   Rest
#> 201  4.76949726 11.0245565   Rest
#> 202  7.48367868 14.5209681   Rest
#> 203  7.19155100  9.3243316   Rest
#> 204  2.43379335 10.3770324   Rest
#> 205 11.52342672 10.5489146   Rest
#> 206  4.86762132  3.2247972   Rest
#> 207  1.14382209  4.4327638   Rest
#> 208 14.62473688  4.8958982   Rest
#> 209  7.52741051 10.4626771   Rest
#> 210 11.74204477 12.5707361   Rest
#> 211 12.51159092 10.1886537   Rest
#> 212 11.49613553  3.9779397   Rest
#> 213  8.88699597 12.4912518   Rest
#> 214  3.20298091 11.2339969   Rest
#> 215  3.57415755  8.3557328   Rest
#> 216  9.23740120  1.1961542   Rest
#> 217  8.77020502 14.9055155   Rest
#> 218 11.53384857  9.2002362   Rest
#> 219  4.80623527  4.3451920   Rest
#> 220  7.71005863  2.8046236   Rest
#> 221 14.24812338  1.4348345   Rest
#> 222  8.51211965 14.1984756   Rest
#> 223  8.43614781  3.3893423   Rest
#> 224  3.78376166 13.3480419   Rest
#> 225  7.86612117  5.1507968   Rest
#> 226 11.88084552  7.2065972   Rest
#> 227  6.61738288 11.3787469   Rest
#> 228 16.95092797  5.7995106   Rest
#> 229  3.78632431  2.0097925   Rest
#> 230  5.32238766  3.2252526   Rest
#> 231 13.19117008 14.0485862   Rest
#> 232 13.05814276  4.4636701   Rest
#> 233  1.57024428  3.1580143   Rest
#> 234 14.69152155 12.5713030   Rest
#> 235 13.64990260 15.8032801   Rest
#> 236 12.29949571  9.7394952   Rest
#> 237  0.46842449 14.6602736   Rest
#> 238  5.40027886  9.3932036   Rest
#> 239 17.01280051  5.9679620   Rest
#> 240  5.43264656 12.9842615   Rest
#> 241 14.63497731  4.9929193   Rest
#> 242  0.44359950  9.9477941   Rest
#> 243  9.88544187  1.3739375   Rest
#> 244  4.40457386 14.8010292   Rest
#> 245 10.01602387 13.5186412   Rest
#> 246 10.44961152 10.4691415   Rest
#> 247  3.75909445  4.3480867   Rest
#> 248  5.33871279  0.4979884   Rest
#> 249  3.43362405 10.6578333   Rest
#> 250 10.74155670 12.9921711   Rest
#> 251 12.94482507 -0.8616697   Rest
#> 252 11.74233041  1.2033198   Rest
#> 253  3.99322611  9.4433851   Rest
#> 254 10.88235031 15.6112407   Rest
#> 255 13.50893799  6.0799196   Rest
#> 256 13.68709985  3.9809147   Rest
#> 257  9.06243011  5.0906304   Rest
#> 258  3.60842971  7.7663495   Rest
#> 259  3.84073245  5.9589819   Rest
#> 260  8.13766151 13.6119032   Rest
#> 261  6.79609855  1.6397447   Rest
#> 262  9.22288223 11.9569895   Rest
#> 263  2.80335117  8.4221474   Rest
#> 264  3.99680048  9.1389905   Rest
#> 265  8.17421979 14.7741660   Rest
#> 266 10.59898931  4.2620732   Rest
#> 267  5.05238333  2.5966191   Rest
#> 268 15.05010008  4.7672981   Rest
#> 269 16.80538119  3.3815411   Rest
#> 270  4.24654352  4.9664959   Rest
#> 271  7.07405262  0.8349615   Rest
#> 272  8.77941799 15.8408519   Rest
#> 273  9.59162233 11.9157488   Rest
#> 274  7.34754275  3.8051747   Rest
#> 275  9.81815597 13.2730624   Rest
#> 276 15.23217089  0.9677642   Rest
#> 277 12.24465892 14.8589919   Rest
#> 278  7.29591340 15.2130732   Rest
#> 279  6.62623181 11.4169757   Rest
#> 280  2.58725486  4.2436327   Rest
#> 281  5.78271013 -0.3789287   Rest
#> 282 11.95478666  8.2318644   Rest
#> 283  8.62650172 14.2120201   Rest
#> 284  0.96109231  8.3833157   Rest
#> 285  2.99868354  4.5081521   Rest
#> 286 10.44856589 11.6197456   Rest
#> 287 11.48041220  3.8953822   Rest
#> 288  6.70128223 11.8107702   Rest
#> 289 13.30532625  6.0237480   Rest
#> 290  4.19947772 12.8962691   Rest
#> 291 10.20122769  4.0116482   Rest
#> 292  6.65986458 10.0313498   Rest
#> 293 15.87958798  3.1721512   Rest
#> 294  0.42384244 10.1037486   Rest
#> 295 10.05828379 11.1129749   Rest
#> 296 11.55119290 16.0977911   Rest
#> 297  5.19509159 10.4010231   Rest
#> 298  5.24215798  1.3115484   Rest
#> 299 10.23848133 16.7246720   Rest
#> 300  4.68183639 12.9341420   Rest
#> 301 13.50501834 10.4066146   Rest
#> 302  6.83146449 13.5561078   Rest
#> 303 10.38845194  5.6127809   Rest
#> 304  6.97312295  2.2530368   Rest
#> 305 14.89665830  4.0046583   Rest
#> 306 10.86952992  4.9477526   Rest
#> 307 14.55185353  5.9681151   Rest
#> 308 15.11682085  5.5007397   Rest
#> 309  7.14962685  1.3668540   Rest
#> 310  2.08788384 11.4944458   Rest
#> 311 -0.95401732 12.6057310   Rest
#> 312  9.79166238 10.2414697   Rest
#> 313  4.75415456  1.5425557   Rest
#> 314 13.17508901 10.5992705   Rest
#> 315  2.69639494  3.2449553   Rest
#> 316  6.66714255 12.2452561   Rest
#> 317  4.37510690  2.2003667   Rest
#> 318 14.18664984 11.6019948   Rest
#> 319 16.66635641  7.8645617   Rest
#> 320 16.00975781 14.0632592   Rest
#> 321  6.99307665 12.1307445   Rest
#> 322  8.01627150 16.4558390   Rest
#> 323  3.00872400  4.9123677   Rest
#> 324  2.83226259 11.3408190   Rest
#> 325 14.10621645 11.0890410   Rest
#> 326 15.35455138  8.4490405   Rest
#> 327  3.76745538  9.2371145   Rest
#> 328  9.95976203  8.3107798   Rest
#> 329 -0.01691955 10.7963131   Rest
#> 330  4.91136735  7.8159013   Rest
#> 331  4.09669826  7.0732909   Rest
#> 332 13.27346162 13.7335824   Rest
#> 333  7.09049857 12.3655019   Rest
#> 334  4.14335249  3.7573993   Rest
#> 335  4.51506164 12.5036327   Rest
#> 336  5.29563328 13.9027996   Rest
#> 337  6.31408101 10.5750322   Rest
#> 338 12.51331597 17.5927195   Rest
#> 339 12.15919561 11.6127943   Rest
#> 340  5.02933843 15.5970886   Rest
#> 341  4.47178670 13.1486934   Rest
#> 342  9.75671734  5.4717492   Rest
#> 343  6.17471950  2.7544652   Rest
#> 344 13.70081451  8.4387541   Rest
#> 345  7.56238237  7.4150520 Circle
#> 346  5.02775069  8.0681012 Circle
#> 347  6.47475728 10.0937889 Circle
#> 348  7.99506446  8.7520703 Circle
#> 349  7.06272015  5.6039306 Circle
#> 350 10.09144750  8.9425519 Circle
#> 351  9.51502935 10.0055136 Circle
#> 352  8.41475257  9.0824001 Circle
#> 353  4.78135436  8.3657552 Circle
#> 354  7.96647808 11.2316489 Circle
#> 355 10.80828514  6.6252551 Circle
#> 356  8.88311833  7.7063915 Circle
#> 357  9.00235529  7.6918909 Circle
#> 358  6.09578869  8.3365267 Circle
#> 359  6.65656730  6.4773382 Circle
#> 360  6.74341575  9.8951495 Circle
#> 361  6.93808399  9.2024169 Circle
#> 362  9.01184914 10.9798796 Circle
#> 363  7.29080374  9.0396273 Circle
#> 364  8.86829588  6.1567072 Circle
#> 365  5.27137764  8.2110675 Circle
#> 366  6.34979158  7.5702455 Circle
#> 367  7.03009100  9.3733227 Circle
#> 368  7.90038714  5.8489540 Circle
#> 369 10.47420975  7.8885355 Circle
#> 370  7.09685017  6.1974862 Circle
#> 371  7.51322110  7.1869072 Circle
#> 372 10.72080327  8.8319651 Circle
#> 373  5.84222906  9.4006791 Circle
#> 374  5.99284620  9.5294414 Circle
#> 375  7.49552769  7.4582081 Circle
#> 376  5.11763749  6.9533031 Circle
#> 377  8.04078899  8.0874989 Circle
#> 378  8.58744141  5.6302988 Circle
#> 379 10.47763092  8.8368402 Circle
#> 380  7.60825557  7.9554778 Circle
#> 381 10.04424902  6.7151385 Circle
#> 382  8.70518930  7.2928539 Circle
#> 383  9.48522971  7.4456972 Circle
#> 384  6.26249137  8.7623213 Circle
#> 385  7.86218830  8.7008554 Circle
#> 386  9.83947619 10.9392289 Circle
#> 387  8.61570652  5.7633405 Circle
#> 388  6.06351098  8.5302820 Circle
#> 389 10.51369931  8.0828481 Circle
#> 390  6.09021601  7.3848171 Circle
#> 391  9.00326760  6.8115654 Circle
#> 392  7.92713334  4.6866765 Circle
#> 393  4.89472125  8.2033167 Circle
#> 394  5.06799147  6.7043790 Circle
#> 395  8.95385471  9.9926657 Circle
#> 396  8.75842797 12.2306689 Circle
#> 397  9.17882455 10.1010441 Circle
#> 398  8.03619275  8.9781372 Circle
#> 399  5.58530004  7.3460522 Circle
#> 400  8.08696425  7.9210457 Circle
#> 401  6.76705649  8.8299276 Circle
#> 402  8.65953483  6.7779894 Circle
#> 403  5.60934352  8.7971909 Circle
#> 404  7.82266245  4.3639931 Circle
#> 405  7.70913387  6.9921712 Circle
#> 406  5.78034841  9.1086738 Circle
#> 407  9.17857220  5.9122652 Circle
#> 408  8.61131239 10.3932703 Circle
#> 409  6.94612734  7.7002874 Circle
#> 410  9.07702997 10.4266895 Circle
#> 411  9.28714557  6.4250719 Circle
#> 412  6.43021627  8.0087602 Circle
#> 413  8.39624560  5.8420658 Circle
#> 414  6.64022271  9.5217798 Circle
#> 415  6.90784878  7.9029739 Circle
#> 416  8.61873025  5.5415198 Circle
#> 417 10.01167546  9.4935851 Circle
#> 418  9.98087520 10.5064460 Circle
#> 419  8.57968299  8.8216137 Circle
#> 420 10.31371294 12.1794675 Circle
#> 421  8.51902028  8.9532508 Circle
#> 422  8.52017652  7.3151104 Circle
#> 423  7.68906353  6.0294117 Circle
#> 424  7.73492977  7.4430164 Circle
#> 425  9.45782876  8.0183184 Circle
#> 426  5.30208499  9.2659125 Circle
#> 427  9.56131062  7.8426079 Circle
#> 428  5.66848526  8.9840283 Circle
#> 429  5.83411576  9.6469304 Circle
#> 430  5.11446375  6.7937313 Circle
#> 431  6.28605627  9.4918468 Circle
#> 432  8.15322614  9.0462309 Circle
#> 433  6.91636048  9.0281984 Circle
#> 434  7.76801402  9.0867896 Circle
#> 435 10.23941260  9.9175158 Circle
#> 436  7.02610117  8.3887941 Circle
#> 437  7.94746085  8.9526012 Circle
#> 438  9.09663833  9.8715318 Circle
#> 439  6.33852041  9.4210279 Circle
#> 440  5.39708389  6.1584412 Circle
#> 441  9.05472256  8.5287828 Circle
#> 442  7.70418143  4.2720640 Circle
#> 443  7.81364898  9.2384195 Circle
#> 444  6.42194929  5.3183810 Circle
#> 445 10.52776854  8.0342487 Circle
#> 446  7.52600905  8.4071350 Circle
#> 447  9.71153322  9.0714979 Circle
#> 448 10.82736143  7.0913119 Circle
#> 449  7.26418654  6.6850979 Circle
#> 450  7.45627079  6.8504393 Circle
#> 451  7.17196681  5.7101747 Circle
#> 452  5.08317903  8.4687227 Circle
#> 453  7.41062566 10.5117508 Circle
#> 454  6.04016578  7.5563505 Circle
#> 455  7.37971451  7.0982836 Circle
#> 456  8.11351583  9.4290724 Circle
#> 457  7.09946411  9.8204960 Circle
#> 458  7.87536521  5.9196511 Circle
#> 459 11.04496983 10.4880095 Circle
#> 460  7.81220897  6.7691386 Circle
#> 461  4.34484714  6.3368739 Circle
#> 462  7.35376487 11.5857808 Circle
#> 463  7.54729602  7.1835050 Circle
#> 464  5.84756830  8.6764527 Circle
#> 465  7.75892988 11.9132127 Circle
#> 466  7.87569798  6.5474505 Circle
#> 467 10.18073437  9.3118540 Circle
#> 468  5.97309809  6.1613049 Circle
#> 469  8.95761786  7.5780717 Circle
#> 470  6.48507978  6.8492250 Circle
#> 471  6.23579289  7.6933924 Circle
#> 472 10.59926437 10.1040207 Circle
#> 473  6.55311642  9.1393299 Circle
#> 474  6.74341207  9.4849693 Circle
#> 475  6.46352226  6.9022148 Circle
#> 476  6.84707888  6.8102395 Circle
#> 477  8.65364308  7.9388069 Circle
#> 478  7.93728249  8.3940150 Circle
#> 479  6.18659931  7.4717601 Circle
#> 480  6.03131292  6.8288366 Circle
#> 481  9.13406662  7.7524528 Circle
#> 482  4.68484897  7.6792750 Circle
#> 483  6.94752126  8.3558604 Circle
#> 484  5.98380779  6.7924003 Circle
#> 485 11.11124368  8.2156975 Circle
#> 486  7.56144395  8.6339063 Circle
#> 487  6.60402892  6.0649312 Circle
#> 488  8.09590818  9.4346751 Circle
#> 489  5.59570106  9.6814619 Circle
#> 490  8.34354089  6.7201223 Circle
#> 491  6.82829166  7.3473985 Circle
#> 492  6.32220583  8.9531865 Circle
#> 493  6.23049847  5.7494261 Circle
#> 494  6.34048846  6.8873656 Circle
#> 495  9.40444813  9.3242882 Circle
#> 496  7.46974280 10.1500824 Circle
#> 497  8.88140814 10.3805935 Circle
#> 498  8.47973981  4.9205662 Circle
#> 499  8.06464292  6.8244369 Circle
#> 500  6.81374513 10.2554474 Circle
#> 501  5.94449910  8.1615291 Circle
#> 502  8.35594906  5.3457724 Circle
#> 503 10.21332111  7.4702866 Circle
#> 504  7.17359438  8.5690774 Circle
#> 505  5.91147276  7.6785286 Circle
#> 506  8.83075892  6.5603516 Circle
#> 507  7.96375234  5.8142973 Circle
#> 508  8.81218460 10.1689482 Circle
#> 509  6.16060727  6.7569812 Circle
#> 510 10.70184127  8.9841751 Circle
#> 511  9.67646444  7.1419389 Circle
#> 512  5.82536480  6.6065780 Circle
#> 513  6.72276048  6.4818010 Circle
#> 514 10.20083156  9.2883020 Circle
#> 515  7.92176920  7.1102883 Circle
#> 516  7.89999358  9.4053918 Circle
#> 517  6.59042858 10.7682253 Circle
#> 518  7.46929659 10.1233011 Circle
#> 519  9.88080671  7.5687980 Circle
#> 520  6.35113943  6.4170327 Circle
#> 521  6.80592120  9.3147406 Circle
#> 522  8.62830585 10.3837133 Circle
#> 523 11.39375425  5.9811547 Circle
#> 524  7.39528931 10.8884270 Circle
#> 525  9.94771909  9.8099813 Circle
#> 526  7.48162075  8.0583504 Circle
#> 527  8.05137964  6.3817104 Circle
#> 528  9.55985828 10.9162690 Circle
#> 529  6.57957537 11.1040162 Circle
#> 530  8.56675397  7.9145162 Circle
#> 531 10.05370052 10.4776845 Circle
#> 532  4.40625771  7.6335235 Circle
#> 533  6.28323214  9.6343414 Circle
#> 534  8.24374758  8.7774695 Circle
#> 535  8.96073910  8.7569742 Circle
#> 536  6.35952663  9.0052984 Circle
#> 537  3.98353249  6.0105440 Circle
#> 538  9.35906721  9.4279679 Circle
#> 539  8.35631877  6.6755276 Circle
#> 540  8.07237539 11.0536800 Circle
#> 541  6.41498002  7.3142571 Circle
#> 542 10.38381155  8.8032247 Circle
#> 543  9.80312339  9.6800112 Circle
#> 544  8.01595925  8.2470681 Circle
#> 545  6.70993653 10.4626903 Circle
#> 546  3.54131170  7.2143820 Circle
#> 547  7.98980697  8.4623236 Circle
#> 548 10.68405397 10.0404590 Circle
#> 549  6.94630563  6.4497140 Circle
#> 550  5.72628874  6.3451927 Circle
#> 551  8.58199832  5.7253370 Circle
#> 552 10.25604457  7.8660841 Circle
#> 553  7.98148790  9.1278079 Circle
#> 554 11.50546532  9.2358678 Circle
#> 555  7.04460782  9.7876295 Circle
#> 556  8.78302098  7.5965220 Circle
#> 557  8.60602569  9.1225121 Circle
#> 558  8.74759705  5.1969768 Circle
#> 559  9.14062448  7.4128234 Circle
#> 560  7.16387922  7.4025056 Circle
#> 561  7.43117497  4.1463695 Circle
#> 562  9.64262664  7.8681174 Circle
#> 563  6.11671024  8.7783393 Circle
#> 564  9.79503363  6.4791421 Circle
#> 565  8.69469305  9.8317457 Circle
#> 566 10.36734538  7.6586636 Circle
#> 567  4.99856575  9.4490475 Circle
#> 568  6.23153782  7.0817484 Circle
#> 569  6.52419903  9.1793740 Circle
#> 570  5.89963029  8.0499732 Circle
#> 571  8.62184190  6.4592141 Circle
#> 572  8.59061318  9.6332547 Circle
#> 573  8.38299516  9.6422154 Circle
#> 574  6.66666799  6.8765025 Circle
#> 575  8.10871939  8.7865462 Circle
#> 576  7.78321496  8.7720251 Circle
#> 577  7.97443963  6.0636189 Circle
#> 578  5.41515690  9.7227083 Circle
#> 579  8.56298942  7.3781055 Circle
#> 580  9.44692562 10.0389041 Circle
#> 581 10.15554887  9.6289555 Circle
#> 582  6.20043266  8.4716802 Circle
#> 583  9.15291734  5.8654444 Circle
#> 584  7.50997972  9.3827016 Circle
#> 585  8.96126815  9.6892384 Circle
#> 586  6.04238663  9.6563538 Circle
#> 587  8.27931853 10.8981135 Circle
#> 588  9.15725807  7.6430293 Circle
#> 589  8.84873126  6.7353775 Circle
#> 590  6.66520737  7.5391631 Circle
#> 591  8.43193181  6.3525025 Circle
#> 592 11.56914619  6.9563547 Circle
#> 593 10.09123026  7.6737822 Circle
#> 594  7.42540744  7.5072940 Circle
#> 595  9.51619231  7.3958467 Circle
#> 596  5.20775789  6.2076943 Circle
#> 597  8.40658473  9.4245524 Circle
#> 598  5.48405537  6.7537970 Circle
#> 599  7.30629464  7.8077625 Circle
#> 600  9.96512214  7.6587937 Circle
#> 601  9.90816439  8.1290393 Circle
#> 602  6.24725438  9.3345203 Circle
#> 603  8.25820732  9.3041402 Circle
#> 604  7.52583623  6.0544308 Circle
#> 605  6.27413593  7.0583215 Circle
#> 606  6.86247868  9.4124106 Circle
#> 607  8.70872324  9.4040787 Circle
#> 608  6.94486025  5.2893592 Circle
#> 609  4.57323515  7.0650310 Circle
#> 610  7.27403961 10.3565577 Circle
#> 611  6.30452631  6.0027943 Circle
#> 612  7.08391262 10.8965521 Circle
#> 613  9.62344801  9.4775048 Circle
#> 614  7.73878939  6.5048068 Circle
#> 615  6.74564194 10.2556677 Circle
#> 616  8.21547996 10.1899844 Circle
#> 617  7.19978115  6.8512815 Circle
#> 618  9.42588735  9.7125292 Circle
#> 619  9.40562766  7.7769645 Circle
#> 620  6.82797785  8.8487456 Circle
#> 621  4.46829674  7.6845372 Circle
#> 622 10.27995037 11.0874812 Circle
#> 623  8.68219579 10.0097860 Circle
#> 624  8.75377461  9.8057219 Circle
#> 625  7.42413024  8.3254880 Circle
#> 626  9.63925708  7.2116189 Circle
#> 627  9.82340562  6.2478110 Circle
#> 628  4.41267806  5.5332576 Circle
#> 629  7.19297838  9.1141325 Circle
#> 630  6.50370295  6.2454931 Circle
#> 631  8.63311306  6.2473340 Circle
#> 632  7.80222504  8.8180500 Circle
#> 633  6.37864792  8.7816402 Circle
#> 634  6.95990439  6.6667737 Circle
#> 635  8.12506732  4.0367375 Circle
#> 636  6.97614028  9.7926501 Circle
#> 637  6.96780465  9.0027485 Circle
#> 638  8.36064339  9.6719655 Circle
#> 639  7.49446212 11.0966442 Circle
#> 640  5.70597967  9.5656419 Circle
#> 641 10.77780102  7.9987422 Circle
#> 642  8.68529669  6.3530283 Circle
#> 643  7.76484146  9.7040788 Circle
#> 644  8.23430885  5.5220322 Circle
#> 645  4.37431899  8.9859946 Circle
#> 646  6.81469358  8.1395330 Circle
#> 647  6.84259095  9.3119571 Circle
#> 648  5.36657724  8.3885120 Circle
#> 649  9.16232491 10.2674966 Circle
#> 650  8.64995304 10.0996110 Circle
#> 651  8.08850800  9.7853559 Circle
#> 652 10.08817922  6.2636231 Circle
#> 653  7.91495072  9.5114731 Circle
#> 654 10.50632494  9.6478743 Circle
#> 655 11.23171387  8.6045828 Circle
#> 656  7.30969215 10.4135567 Circle
#> 657  6.11931179  8.5802843 Circle
#> 658  9.24579508  8.1615413 Circle
#> 659 10.90447182  8.7226919 Circle
#> 660  7.46904877  8.6161491 Circle
#> 661  9.44307364  5.9027253 Circle
#> 662  6.94449325  5.4913224 Circle
#> 663  7.91266323  7.2431861 Circle
#> 664  7.75214977 11.7388757 Circle
#> 665  8.75471023  6.0496081 Circle
#> 666  7.15755722 10.1632540 Circle
#> 667  6.84418630  6.1182587 Circle
#> 668  8.96950884  7.4412605 Circle
#> 669  6.60860695  6.9540381 Circle
#> 670  9.03419970  7.4527669 Circle
#> 671  8.84184607  7.9696797 Circle
#> 672 10.69526778 11.2979987 Circle
#> 673  7.31227283  5.7693843 Circle
#> 674  7.40317455 11.3672400 Circle
#> 675  7.93187172  9.3430930 Circle
#> 676  6.52043955 10.1882403 Circle
#> 677  5.29928742  6.6821418 Circle
#> 678  7.96375218  7.2846223 Circle
#> 679  9.66023176  6.0620444 Circle
#> 680  9.23769811  8.9493289 Circle
#> 681  5.82308880  6.1964138 Circle
#> 682  7.15431271  5.4318147 Circle
#> 683  8.27226680  6.4821720 Circle
#> 684  8.73017233  6.2993412 Circle

rose(circle_example[, c("x", "y", "class")], var = "class", over_ratio = 0.8)
#>              x           y  class
#> 1   -1.5933079 12.28299115   Rest
#> 2   14.0267151  7.58260828   Rest
#> 3    6.5833137  3.86135731   Rest
#> 4    8.0013388  9.60739066   Rest
#> 5    3.2561877  2.85083690   Rest
#> 6    2.4638271  7.57319320   Rest
#> 7    8.7721954  0.08045149   Rest
#> 8   11.5228946 10.46628421   Rest
#> 9    6.3669828  2.31016059   Rest
#> 10   4.4536549  2.88387604   Rest
#> 11   9.3570995  8.80902650   Rest
#> 12   6.3574610 10.73831782   Rest
#> 13  11.3867943  4.41824723   Rest
#> 14  12.1126639  9.49361986   Rest
#> 15   1.2207380 10.91033959   Rest
#> 16   3.6117118 10.02125313   Rest
#> 17  12.2625769 15.68151681   Rest
#> 18   1.0065850 16.12997290   Rest
#> 19  12.5868545 10.40577858   Rest
#> 20   3.1044481  3.17291140   Rest
#> 21  10.0746946 12.45900387   Rest
#> 22   0.7535192  5.73103446   Rest
#> 23  12.1774478  4.15926352   Rest
#> 24   1.3304981  5.91150934   Rest
#> 25   3.3192934  6.49775199   Rest
#> 26   9.2712081 11.37119991   Rest
#> 27  12.9353110  0.11205381   Rest
#> 28  12.1972807 12.87978570   Rest
#> 29   4.0637905  2.91254791   Rest
#> 30   2.5713327  6.07594798   Rest
#> 31  15.0151190 10.91368060   Rest
#> 32   3.8578673 12.11179859   Rest
#> 33  10.3150758  5.40589807   Rest
#> 34  11.9911424  9.70171235   Rest
#> 35   3.8965514 10.74874443   Rest
#> 36   3.2509354  6.52724064   Rest
#> 37  12.8439959  5.11023704   Rest
#> 38  11.7153491  2.43614413   Rest
#> 39  15.8391193  9.17831805   Rest
#> 40   2.3386400  5.14952663   Rest
#> 41  10.1266848  3.61652188   Rest
#> 42  13.6160723  9.37845004   Rest
#> 43  10.2340311 11.51356859   Rest
#> 44   5.8435008 14.40868299   Rest
#> 45   6.4844199 13.33764738   Rest
#> 46   9.4161651 13.35470298   Rest
#> 47  12.5023810  5.93588633   Rest
#> 48   2.7681415  5.22790732   Rest
#> 49   2.8350106  6.90562754   Rest
#> 50   9.1696747 14.95060215   Rest
#> 51   8.4143574 12.60009389   Rest
#> 52   6.7369028 -1.18327742   Rest
#> 53   5.0199071  2.27663040   Rest
#> 54   2.8293698  7.61634698   Rest
#> 55   5.6601297  3.67576207   Rest
#> 56   6.0103076 12.30951981   Rest
#> 57  12.0623514 10.20328029   Rest
#> 58  11.2451693  6.91373080   Rest
#> 59   9.1513268 12.40871375   Rest
#> 60  10.3078974  9.17532959   Rest
#> 61   4.2308615 12.54042759   Rest
#> 62  10.6751628 15.39384408   Rest
#> 63  15.7747640  6.29859896   Rest
#> 64   6.3162922  8.07030601   Rest
#> 65  13.8558023 13.38350995   Rest
#> 66   3.6653405 15.96286385   Rest
#> 67   9.9045820  4.64200862   Rest
#> 68  13.0167037 13.93129411   Rest
#> 69   0.5176676 12.27285217   Rest
#> 70  12.3893361  4.65243658   Rest
#> 71  13.2796080  8.24048639   Rest
#> 72   3.7146009  0.95380340   Rest
#> 73   8.5668736 10.62499971   Rest
#> 74  15.3163790 10.78034492   Rest
#> 75  11.0795048  5.84689991   Rest
#> 76   9.4390439  3.87910974   Rest
#> 77   7.0883445 11.15516342   Rest
#> 78  11.7738849  5.69535094   Rest
#> 79   3.3167243  2.23164012   Rest
#> 80  15.0201817  0.49125618   Rest
#> 81  13.9711941  9.25587773   Rest
#> 82   4.4270428  9.42207994   Rest
#> 83   6.2166390 12.03545364   Rest
#> 84  11.8731375 13.36280285   Rest
#> 85   6.6071117 14.62350452   Rest
#> 86  12.5820390 15.81915900   Rest
#> 87  16.4591584  4.23333169   Rest
#> 88  13.4957687 14.13376724   Rest
#> 89   1.7497425  6.62833574   Rest
#> 90   5.3516169  9.55235339   Rest
#> 91   7.5407060 12.62961290   Rest
#> 92  12.7242550 10.05298259   Rest
#> 93   3.6171396  5.96328885   Rest
#> 94  10.2320574  9.29723331   Rest
#> 95  12.7624538 12.75262726   Rest
#> 96   1.8465326  0.12200595   Rest
#> 97  14.0320856 11.28783922   Rest
#> 98  15.1630479 11.81781448   Rest
#> 99  14.1875493 11.19126007   Rest
#> 100  7.4626425 10.47005282   Rest
#> 101  9.6005800 14.68428540   Rest
#> 102  4.7004939 15.26149636   Rest
#> 103  5.2781991  7.05916182   Rest
#> 104  9.6286027 11.40655524   Rest
#> 105 11.3341332  9.66353911   Rest
#> 106 14.0943663  7.31126557   Rest
#> 107  1.7695268  9.32866170   Rest
#> 108  0.8113048  7.97466692   Rest
#> 109  2.5989617  1.42356837   Rest
#> 110  8.5809255  3.28881354   Rest
#> 111 15.7024406 12.41361185   Rest
#> 112  2.3507842  0.92139584   Rest
#> 113 14.5088183  7.65262578   Rest
#> 114  4.1290275  6.88243598   Rest
#> 115 16.8153350  6.28410472   Rest
#> 116 11.9763317 10.48362121   Rest
#> 117 13.3663980  3.36910945   Rest
#> 118  2.2046128  7.24846029   Rest
#> 119 14.5356830 10.03069901   Rest
#> 120  4.1101707  3.04538416   Rest
#> 121 11.0506037  2.76741198   Rest
#> 122 15.0415077  9.23259894   Rest
#> 123  4.0715527 16.27547138   Rest
#> 124  3.3032596  9.34260335   Rest
#> 125 16.2367806  3.87733686   Rest
#> 126  9.4232770  9.95418637   Rest
#> 127 11.7086838  5.23246404   Rest
#> 128 -0.1705898 12.92130871   Rest
#> 129 11.6668514  6.69061940   Rest
#> 130 14.1917347  6.72390137   Rest
#> 131  1.9168163  7.43730989   Rest
#> 132  8.2351089 14.92561150   Rest
#> 133  5.5712100  0.47382941   Rest
#> 134  6.8708694 13.63014699   Rest
#> 135 11.4679085 15.05169343   Rest
#> 136  8.2837494 16.56163135   Rest
#> 137 10.6580997 10.66538237   Rest
#> 138  7.0088288 13.97849164   Rest
#> 139  6.4009326  3.59322420   Rest
#> 140  7.8834095  1.21344358   Rest
#> 141  3.2383249  4.32006647   Rest
#> 142  7.2991661 14.87845462   Rest
#> 143  3.9389185 15.62531696   Rest
#> 144  5.3196784 11.65213584   Rest
#> 145 12.1939661  6.00763575   Rest
#> 146 12.9431474 13.15025909   Rest
#> 147  0.4328864 15.27593266   Rest
#> 148  0.7013893  7.29049016   Rest
#> 149  0.7533658  6.40704739   Rest
#> 150  9.9509780  7.78617316   Rest
#> 151  9.6648413 10.94920173   Rest
#> 152  1.0080293  8.23542816   Rest
#> 153 14.0644628  9.74557238   Rest
#> 154  3.8567650  1.47491792   Rest
#> 155 14.8470311 13.61677999   Rest
#> 156 14.4922036 14.69269807   Rest
#> 157 14.9232491 15.44152565   Rest
#> 158  9.3948802 14.53085918   Rest
#> 159  5.3968643  4.67854705   Rest
#> 160  3.6587441  9.83411660   Rest
#> 161  4.2896899  6.33031929   Rest
#> 162 15.3215999  7.15452082   Rest
#> 163 11.5915208  5.89182581   Rest
#> 164 13.8921334  3.51801206   Rest
#> 165  6.7403740 11.83513892   Rest
#> 166  6.1594062 13.36636646   Rest
#> 167  6.6959363 13.05356050   Rest
#> 168  1.0760595  8.30920559   Rest
#> 169  7.1469602  3.95718063   Rest
#> 170  8.9888205 11.56282810   Rest
#> 171  5.4545825 13.62482709   Rest
#> 172  4.4684132 16.28136575   Rest
#> 173  9.8852603  2.73286114   Rest
#> 174  1.3547196  1.09502215   Rest
#> 175  4.3299422 12.76952763   Rest
#> 176  6.3394421 11.55657297   Rest
#> 177 10.7338330 11.80899055   Rest
#> 178 10.4472547 11.55255955   Rest
#> 179  6.0370840  7.87590803   Rest
#> 180  3.4762883  6.48275361   Rest
#> 181  5.4521633 -0.65206696   Rest
#> 182  6.3580215  6.94502389   Rest
#> 183  3.5161349  7.68235965   Rest
#> 184  2.1710427  4.67639162   Rest
#> 185 11.9706908  4.50299141   Rest
#> 186  2.4081244  6.72132988   Rest
#> 187  8.5925329 15.97326610   Rest
#> 188 11.4297760 11.39629719   Rest
#> 189 14.3045901  8.87412015   Rest
#> 190  8.5988713 14.28732061   Rest
#> 191  5.8913205 12.02303809   Rest
#> 192  1.9271827  6.39501945   Rest
#> 193 -1.0947932 15.87465748   Rest
#> 194  7.0874776  8.87326863   Rest
#> 195 14.0565352  7.11713746   Rest
#> 196 12.8383856 10.91108542   Rest
#> 197  7.3225551 15.54965096   Rest
#> 198 10.9902267 -0.89653649   Rest
#> 199  1.9873632  4.62047878   Rest
#> 200  4.4975820  7.79734260   Rest
#> 201  0.4540496 12.44922327   Rest
#> 202  5.9255865  9.37686590   Rest
#> 203 12.0457762  3.12821944   Rest
#> 204 14.8206408  8.48244526   Rest
#> 205 12.6629381  3.25414790   Rest
#> 206 13.5788070  6.17135534   Rest
#> 207  2.6003064  8.72764274   Rest
#> 208 14.4558573 14.40046999   Rest
#> 209  7.6135255 14.98646766   Rest
#> 210 11.1626826 15.09095021   Rest
#> 211  4.6520024 14.58250157   Rest
#> 212  0.9379034  9.44563008   Rest
#> 213  2.1778872  3.42090081   Rest
#> 214  6.0709375 10.23770120   Rest
#> 215 11.4917958 10.43302807   Rest
#> 216  3.7514126 11.22489131   Rest
#> 217  3.1506284 10.49822175   Rest
#> 218  3.8466786  8.36992655   Rest
#> 219  8.3271639  9.76068751   Rest
#> 220  2.7289681  0.80139094   Rest
#> 221 10.3622456  9.38375246   Rest
#> 222 11.9824178  5.71693426   Rest
#> 223  6.4227250 13.06125495   Rest
#> 224 14.4449621  4.46887621   Rest
#> 225  7.3260731  6.24518282   Rest
#> 226 11.2639938  7.77443646   Rest
#> 227 -1.2505319 13.28369650   Rest
#> 228  4.1889418  7.29860593   Rest
#> 229  5.7167562 11.29333674   Rest
#> 230  1.8180998  2.77106520   Rest
#> 231  6.9839666 15.87149015   Rest
#> 232  7.4735512 11.00281756   Rest
#> 233  1.8013684  4.01552931   Rest
#> 234  2.2728333  1.62196299   Rest
#> 235  5.4891377 14.60106555   Rest
#> 236  4.2896795 11.69901957   Rest
#> 237  1.4867707 15.09074521   Rest
#> 238  5.8454947  9.49890642   Rest
#> 239 17.8988528 14.81624530   Rest
#> 240  9.3703675  4.59147812   Rest
#> 241  4.9690801  2.98806679   Rest
#> 242  7.3074607 13.85208159   Rest
#> 243 15.0770368  5.00532980   Rest
#> 244  1.2252050  1.28698480   Rest
#> 245 12.7274617  4.51699398   Rest
#> 246  3.0898866  5.64628455   Rest
#> 247 17.2628203  1.68054317   Rest
#> 248 11.8504695 11.83257110   Rest
#> 249  2.7674247 -0.76680271   Rest
#> 250 11.9231613  4.71501171   Rest
#> 251 12.0606299 12.57736879   Rest
#> 252 14.0546221 10.27378507   Rest
#> 253  5.2789034 10.62693479   Rest
#> 254 13.2592625  6.21962184   Rest
#> 255  1.3726852  5.98354332   Rest
#> 256  4.0031325 12.88355207   Rest
#> 257 11.0611806  1.15772427   Rest
#> 258 12.1197798  6.89547959   Rest
#> 259  0.3518476  3.96820844   Rest
#> 260  8.3570469 14.54070358   Rest
#> 261  6.0966292 12.43814360   Rest
#> 262  7.3559181  1.43878495   Rest
#> 263 10.3008157 13.66469779   Rest
#> 264 -0.6494853  7.86164134   Rest
#> 265 15.2066723  6.93296010   Rest
#> 266  5.6542456 11.25334536   Rest
#> 267  4.6878701  7.34117231   Rest
#> 268 10.1748083  9.15464219   Rest
#> 269 16.6134901 10.95793751   Rest
#> 270 14.3747871  4.01425471   Rest
#> 271  8.3718847 14.44250710   Rest
#> 272  9.0876181 10.14214688   Rest
#> 273  5.2306446 13.66085537   Rest
#> 274 10.2614591  4.24798139   Rest
#> 275 -0.7054117  0.99342745   Rest
#> 276  2.2841494  7.18961781   Rest
#> 277 12.4383655 10.05643116   Rest
#> 278  4.3004578 15.97567526   Rest
#> 279  9.3509648  0.26654210   Rest
#> 280  1.1973801  1.18058655   Rest
#> 281  0.2044423 11.80377192   Rest
#> 282  1.9863224  9.75894915   Rest
#> 283 11.4171652  4.46799478   Rest
#> 284  8.8871307 10.63164454   Rest
#> 285 11.8240900  2.80632395   Rest
#> 286  5.2166950 12.49952996   Rest
#> 287  6.6939905  1.12828873   Rest
#> 288  8.7125500  6.78620191 Circle
#> 289  6.3939140  7.99379166 Circle
#> 290  6.7162494  9.07298966 Circle
#> 291  9.1750332  9.28216549 Circle
#> 292  5.1163586  6.79877974 Circle
#> 293  6.1281802  5.95926183 Circle
#> 294 10.9430318  9.66977034 Circle
#> 295  6.3820920  7.41733446 Circle
#> 296  7.8229482 10.16071676 Circle
#> 297  6.7850588  7.46038120 Circle
#> 298  7.9797345  6.26432656 Circle
#> 299  9.4043429  8.31116964 Circle
#> 300  8.8511100  6.30838715 Circle
#> 301  7.5044798 10.23826245 Circle
#> 302  8.8351518  8.15477382 Circle
#> 303 10.3654720  7.15008483 Circle
#> 304  8.3816232  8.25061153 Circle
#> 305  9.5102698  6.23331947 Circle
#> 306  7.0728040  7.74252608 Circle
#> 307  6.8340178 10.23449621 Circle
#> 308  8.2239433  6.99285956 Circle
#> 309  8.3388608  6.72530906 Circle
#> 310  9.1654084  7.38730769 Circle
#> 311 10.2145031 10.49141362 Circle
#> 312  8.8546508  8.07028738 Circle
#> 313 11.2314536  8.73287878 Circle
#> 314  5.2622311  7.68138374 Circle
#> 315  8.1878792  9.53964796 Circle
#> 316  6.3828037  6.10122942 Circle
#> 317  9.0492145  9.12253182 Circle
#> 318  7.0615794  8.51718870 Circle
#> 319  9.3446504  7.38333017 Circle
#> 320  7.9124392 11.09403548 Circle
#> 321  5.3123091  6.43259488 Circle
#> 322  9.3423849 11.06638912 Circle
#> 323 10.0147113  5.06707276 Circle
#> 324  6.9179242  8.73825078 Circle
#> 325  6.0852806  9.42573240 Circle
#> 326  6.3206254  6.37619875 Circle
#> 327  6.7953562  9.49612304 Circle
#> 328  7.4728223  4.70368932 Circle
#> 329  9.5730926  9.54704005 Circle
#> 330  8.9564010  7.59130231 Circle
#> 331  8.2536020 10.43595727 Circle
#> 332  8.2706992  8.82402154 Circle
#> 333  8.5921517  9.34111945 Circle
#> 334  7.8320739  6.44953426 Circle
#> 335 10.2305849  8.22277228 Circle
#> 336  8.9397849 11.03455871 Circle
#> 337  8.7151013 10.29521174 Circle
#> 338  8.0337502  8.56207641 Circle
#> 339  5.6079525  6.46859969 Circle
#> 340  9.4056566  9.74695048 Circle
#> 341  5.9166354  9.26316803 Circle
#> 342  6.4561708  6.13965444 Circle
#> 343  8.5177463  9.57583864 Circle
#> 344  9.4313693  6.44873217 Circle
#> 345  8.9964256 11.57179523 Circle
#> 346  4.9351184  8.05229282 Circle
#> 347  7.9851031  7.79023031 Circle
#> 348  7.0896598  8.08313238 Circle
#> 349  7.7438245  7.58009511 Circle
#> 350  6.1968550  7.31253857 Circle
#> 351  9.6517718  5.70409080 Circle
#> 352  3.7825201  6.94016613 Circle
#> 353  7.7215560 10.19118748 Circle
#> 354  9.1011537  9.89412600 Circle
#> 355  7.1215565  9.84058604 Circle
#> 356  9.7279792  7.74458475 Circle
#> 357  8.0090307  7.36095225 Circle
#> 358  8.7533575  7.44777283 Circle
#> 359  8.1136742  8.92090298 Circle
#> 360  8.9202213 11.07214764 Circle
#> 361  7.5539901  6.62075856 Circle
#> 362  6.9121490  8.37230025 Circle
#> 363 10.6644039  6.88863948 Circle
#> 364  7.0333262  7.28643255 Circle
#> 365  8.1844093  9.18837029 Circle
#> 366  7.5778329  6.43562320 Circle
#> 367  8.8731382 10.66660295 Circle
#> 368  6.7239365  7.76971495 Circle
#> 369  6.0576786  6.15269985 Circle
#> 370  7.9619220  6.77820129 Circle
#> 371  8.1317599 10.10626577 Circle
#> 372  7.5241995  9.38227840 Circle
#> 373  9.9397221  8.31516553 Circle
#> 374 10.0595388  9.12392279 Circle
#> 375  6.9764556 10.42374247 Circle
#> 376  6.4153631  7.73153767 Circle
#> 377  9.9457017  7.99579090 Circle
#> 378  7.4733610  9.01792600 Circle
#> 379 10.4799622  8.12429827 Circle
#> 380  7.1280759  7.84614761 Circle
#> 381  9.5264639  6.92779752 Circle
#> 382  8.5245213  9.10262203 Circle
#> 383  6.4938040  6.39481739 Circle
#> 384  9.2034051  9.41699851 Circle
#> 385  6.9085351  7.89234972 Circle
#> 386 10.0358070  8.44027929 Circle
#> 387  8.9818646  7.23374708 Circle
#> 388  8.4564112  7.61978458 Circle
#> 389  7.6649368  9.30982578 Circle
#> 390 11.0808739  7.10039252 Circle
#> 391  7.7050447  9.14788724 Circle
#> 392  6.3224103  6.35216850 Circle
#> 393  4.2819477  8.10384816 Circle
#> 394  8.4192145  9.20767260 Circle
#> 395  8.0046495 10.51579830 Circle
#> 396  9.2618346  7.06217404 Circle
#> 397  4.2700030  8.45510184 Circle
#> 398  9.3852536  7.92721711 Circle
#> 399  8.8878876  9.32974940 Circle
#> 400  5.5195544  7.00241529 Circle
#> 401  7.1178355  7.68732771 Circle
#> 402  7.9969257 10.95713220 Circle
#> 403  4.4380292  8.29198570 Circle
#> 404  5.2993918 10.55737385 Circle
#> 405  9.6351806  7.55744880 Circle
#> 406  9.6189224  8.26432415 Circle
#> 407  9.7362375  6.63883432 Circle
#> 408  6.2716360  7.73456691 Circle
#> 409  6.8100851  4.98310254 Circle
#> 410  8.5200653  5.51077464 Circle
#> 411  8.8537065  7.11991120 Circle
#> 412  9.5574276  8.41659880 Circle
#> 413  6.8944388  9.65869405 Circle
#> 414  5.4165030  7.02198547 Circle
#> 415  5.6144150  5.69479998 Circle
#> 416  8.4626289  3.51422579 Circle
#> 417  9.7346582  7.26466667 Circle
#> 418  6.5740532  7.74302227 Circle
#> 419  6.1042886 10.33835545 Circle
#> 420  7.8642994  6.42695244 Circle
#> 421  5.5682464  8.88691774 Circle
#> 422 10.1377797  5.86339187 Circle
#> 423  7.9728722  4.19208537 Circle
#> 424  6.9565938  9.89956795 Circle
#> 425  6.2809445 10.56491021 Circle
#> 426  9.4397013  9.91583105 Circle
#> 427  6.8437435  8.45039447 Circle
#> 428  9.5388036 10.47329353 Circle
#> 429  5.7069481  8.20035440 Circle
#> 430  7.1243099  6.87331229 Circle
#> 431  7.0763037 10.82055188 Circle
#> 432  7.3017124  9.64266509 Circle
#> 433  6.5570071  7.38660918 Circle
#> 434  7.0495621  8.86170890 Circle
#> 435  6.9666007 10.85411999 Circle
#> 436 10.3812658  5.86769386 Circle
#> 437  6.5078144 10.70878906 Circle
#> 438 10.2201248  9.08925903 Circle
#> 439  6.4058921  8.08399245 Circle
#> 440  9.6699388  5.51439777 Circle
#> 441 10.1343123  6.45711677 Circle
#> 442  6.8637612  8.23747611 Circle
#> 443  9.0771128  5.89527480 Circle
#> 444  5.7017488  7.33723322 Circle
#> 445 11.2927314  9.03498534 Circle
#> 446  5.4595669  8.62569903 Circle
#> 447  6.8708160  9.46947471 Circle
#> 448  4.9449537  9.21472014 Circle
#> 449  8.4794515  9.06288238 Circle
#> 450  9.3050868  9.43615585 Circle
#> 451  7.4951381  9.88579462 Circle
#> 452  8.8854041 10.41589147 Circle
#> 453  9.0567967 10.51579512 Circle
#> 454  7.8114598  9.45564704 Circle
#> 455  6.6606576  6.44357918 Circle
#> 456  5.5890368  8.75844561 Circle
#> 457  6.3541278  9.52914943 Circle
#> 458  7.6433153 11.91237552 Circle
#> 459  4.4810245  6.75398243 Circle
#> 460  9.5113007 10.69688815 Circle
#> 461 11.0111310  8.30781917 Circle
#> 462  7.8735015  7.02614832 Circle
#> 463  6.9948937  7.03600451 Circle
#> 464  6.3595519  8.05104969 Circle
#> 465  6.4417108  6.67948881 Circle
#> 466  6.6770882  6.21391264 Circle
#> 467  6.5510166  5.73688718 Circle
#> 468 10.1343924  7.53358894 Circle
#> 469  7.1932820 11.15891170 Circle
#> 470  5.7568195  7.28479205 Circle
#> 471  8.5106384  9.51193418 Circle
#> 472  8.3849570  7.05753977 Circle
#> 473  9.6894677  7.65155375 Circle
#> 474  7.7616906 11.10764304 Circle
#> 475  5.9807430  8.79048835 Circle
#> 476  9.4527898 11.48685098 Circle
#> 477  6.8567976 10.24455121 Circle
#> 478 10.6551510  8.73732004 Circle
#> 479  6.8030165  9.75759351 Circle
#> 480  8.4157789  6.93955623 Circle
#> 481  5.9274010  7.97816629 Circle
#> 482  7.8608287  7.72622835 Circle
#> 483  6.3178435  8.00244552 Circle
#> 484  5.1392106  6.29286619 Circle
#> 485  9.7697160  9.06017953 Circle
#> 486  9.8582021  7.07244131 Circle
#> 487  7.5350118 11.07296983 Circle
#> 488  5.6383979  7.77079061 Circle
#> 489 10.5233298  9.15020419 Circle
#> 490  5.9985538  8.70226827 Circle
#> 491  8.3687261  8.12592971 Circle
#> 492  8.4318063  8.40436489 Circle
#> 493  7.1298868  8.93566597 Circle
#> 494  8.2715015  4.44023160 Circle
#> 495 10.6981300  6.67752258 Circle
#> 496  9.5312344 10.88790872 Circle
#> 497  5.7529187  9.35168666 Circle
#> 498  8.8275073  8.45661897 Circle
#> 499  7.5775151  6.96688873 Circle
#> 500  7.1540111  9.23353298 Circle
#> 501  9.1205121 10.30337729 Circle
#> 502  8.3253499  6.35995500 Circle
#> 503  5.4001146  8.67389236 Circle
#> 504  9.7811688  8.57569741 Circle
#> 505  5.7439409  9.18608293 Circle
#> 506  8.2835655  6.35080487 Circle
#> 507  6.6597671 10.08407162 Circle
#> 508  8.1806654 11.00233942 Circle
#> 509  9.8798508  8.57288155 Circle
#> 510  7.3981359 10.57723167 Circle
#> 511  5.9345860  8.49238776 Circle
#> 512  5.9284747  8.23110032 Circle
#> 513  8.9148260 11.36240427 Circle
#> 514  9.1045569  7.56389509 Circle
#> 515  6.2318960 10.17341454 Circle
#> 516  8.1009557  8.83122082 Circle
#> 517  8.7470124 12.06693288 Circle
#> 518  7.6255791 10.87879667 Circle
#> 519  9.9805297 11.66612897 Circle
#> 520  9.5663087  6.82749667 Circle
#> 521  9.6558602 11.38923534 Circle
#> 522  8.6125654  9.27848196 Circle
#> 523  8.3074480  7.18414091 Circle
#> 524  7.9245727 10.15229650 Circle
#> 525  5.5904765  8.17924019 Circle
#> 526  7.1824533  8.12954914 Circle
#> 527  5.1633708  6.43944450 Circle
#> 528  8.1841482  6.58535153 Circle
#> 529  7.4395325  7.12302212 Circle
#> 530  6.3574357  6.36924331 Circle
#> 531  4.7729396  9.40072790 Circle
#> 532  7.7233417 10.08105390 Circle
#> 533  9.5600535  8.07839378 Circle
#> 534  6.9539684 10.60567138 Circle
#> 535  6.1174421  7.14510018 Circle
#> 536  9.0652693  8.90721472 Circle
#> 537  6.4707519  6.25663095 Circle
#> 538  8.1234312  5.64724567 Circle
#> 539  9.5795061  6.49284783 Circle
#> 540  7.0415844 10.05602940 Circle
#> 541  9.2330345  7.29330932 Circle
#> 542  7.5048621  7.59846584 Circle
#> 543  6.6793293  9.45957861 Circle
#> 544  8.1354718  7.10121648 Circle
#> 545  7.2558741  9.33578582 Circle
#> 546  8.9056406  4.83055918 Circle
#> 547  6.7529162  7.04886341 Circle