ROSE Algorithm

A thin wrapper around ROSE::ROSE() that generates a synthetic balanced sample by enlarging the feature space of minority and majority class examples.

Usage

rose(
  df,
  var,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1
)

Arguments

df

A data.frame or tibble. Must have 1 factor variable with exactly 2 levels and remaining numeric variables.

var

Character, name of variable containing the 2-level factor variable.

over_ratio

A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows as the majority level. See vignette("ratio", package = "themis") for more details.

minority_prop

A numeric value between 0 and 1 for the proportion of synthetic observations from the minority class. Defaults to 0.5, which generates an equal split of minority and majority synthetic observations. This parameter controls the class balance within the synthetic data, while over_ratio controls the total size of the synthetic data.

minority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the minority class. Defaults to 1.

majority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the majority class. Defaults to 1.

Value

A data.frame or tibble, depending on type of df.

Details

The factor variable used to balance around must only have 2 levels.

The ROSE algorithm works by selecting an observation belonging to class k and generating new examples in its neighborhood, which is determined by a smoothing matrix H_k. Smaller values of minority_smoothness and majority_smoothness shrink the entries of H_k, producing tighter neighborhoods. This is a cautious choice when there is a concern that excessively large neighborhoods could blur the boundaries between classes.

This function is a thin wrapper around ROSE::ROSE(). For full details on the underlying implementation, see that function's documentation.

References

Lunardon, N., Menardi, G., and Torelli, N. (2014). ROSE: a Package for Binary Imbalanced Learning. R Journal, 6:82–92.

Menardi, G. and Torelli, N. (2014). Training and assessing classification rules with imbalanced data. Data Mining and Knowledge Discovery, 28:92–122.

See also

Examples

rose(circle_example[, c("x", "y", "class")], var = "class")
#>              x           y  class
#> 1   10.5562093  9.54506933   Rest
#> 2    4.2483015  8.69310004   Rest
#> 3    2.8546038  7.00729898   Rest
#> 4    1.3189180  1.86396168   Rest
#> 5    3.1988587 11.90123669   Rest
#> 6   10.8415881  4.34145402   Rest
#> 7    5.9585034 12.95886599   Rest
#> 8    5.9546851 13.59844483   Rest
#> 9   -0.6591977  4.30494688   Rest
#> 10   2.8301167  1.26053615   Rest
#> 11   1.8378504  2.97558449   Rest
#> 12   5.4142273  3.59054473   Rest
#> 13  14.5374653  2.17680343   Rest
#> 14   4.3435006  2.43556000   Rest
#> 15  11.9911102 11.73297345   Rest
#> 16   8.4976869 12.06775769   Rest
#> 17   3.9055080  2.62322105   Rest
#> 18   8.3146327  6.36197769   Rest
#> 19  12.8891019  0.60833981   Rest
#> 20   7.4586326  4.81840853   Rest
#> 21   7.1496648 16.98041369   Rest
#> 22   3.3842632 11.12717219   Rest
#> 23  18.0934727 11.01229460   Rest
#> 24   9.2092165 15.55412732   Rest
#> 25  12.1560434  1.89865376   Rest
#> 26   4.8787142 13.96444414   Rest
#> 27  14.9794633  8.13750338   Rest
#> 28  17.9056487  0.74319406   Rest
#> 29   5.1571141  4.73585472   Rest
#> 30   3.6937974  0.67034035   Rest
#> 31  12.2395817  8.31897172   Rest
#> 32   3.5154331 10.11884887   Rest
#> 33  10.6521694 11.27243101   Rest
#> 34   3.9438844  9.24195452   Rest
#> 35  -1.1521650  4.31872420   Rest
#> 36  12.9520289  2.39325721   Rest
#> 37   2.6574239  1.46667285   Rest
#> 38  11.7507985 17.81552681   Rest
#> 39  10.2534642 11.66735996   Rest
#> 40   8.9037005  1.56034751   Rest
#> 41   3.4492882 13.81492339   Rest
#> 42  -1.5182579  3.45779006   Rest
#> 43   8.1140569  4.06801602   Rest
#> 44   9.0372766  9.14929936   Rest
#> 45   5.8343088 12.90583673   Rest
#> 46  10.7738187 16.01285714   Rest
#> 47   0.7589551  5.18584461   Rest
#> 48   1.1615379 12.85585513   Rest
#> 49   1.0282659  5.62235562   Rest
#> 50   5.9050244  9.56485282   Rest
#> 51  10.2201070  2.99126044   Rest
#> 52   3.4805720 11.64244689   Rest
#> 53  10.1645177 15.52547283   Rest
#> 54   6.9147109  9.66696905   Rest
#> 55   6.5589400 14.79787310   Rest
#> 56  12.6614962  1.08715108   Rest
#> 57   7.9013105 11.95006545   Rest
#> 58  17.6545455 13.97964174   Rest
#> 59  13.2511133  7.11653027   Rest
#> 60   8.9128326 15.37752597   Rest
#> 61   5.1620973 15.90854583   Rest
#> 62   8.2894685 13.69469808   Rest
#> 63  11.4958461 11.92728553   Rest
#> 64  11.8735409 12.96599394   Rest
#> 65   4.0526006 10.47603769   Rest
#> 66   6.7781283 15.04420311   Rest
#> 67   8.4137509  3.30995730   Rest
#> 68   3.9898019  3.49152323   Rest
#> 69  12.3260719 10.42034553   Rest
#> 70   7.8770038 10.89127115   Rest
#> 71   4.2195178 14.35875659   Rest
#> 72  11.0941628  4.57203989   Rest
#> 73   2.9913327  6.18602178   Rest
#> 74   1.3842897 -1.25980868   Rest
#> 75   2.9778417  3.70304264   Rest
#> 76  13.8476747  0.12972380   Rest
#> 77   8.6680884 15.94466570   Rest
#> 78  11.3498605  7.86070997   Rest
#> 79   7.8797069  4.26999918   Rest
#> 80   5.0161227 11.72373410   Rest
#> 81   9.1688046  2.99785893   Rest
#> 82   7.9424323  1.71136914   Rest
#> 83   4.5986731 12.70143977   Rest
#> 84   7.1129619 15.58936143   Rest
#> 85   6.1768922  4.40430226   Rest
#> 86   1.2507101 12.71325377   Rest
#> 87  11.4031794  2.88864322   Rest
#> 88  10.1596895  6.66981557   Rest
#> 89  14.2636951  1.78429310   Rest
#> 90   3.1624062 11.39245684   Rest
#> 91   8.7800866  0.24679649   Rest
#> 92  13.1093255 -0.80639181   Rest
#> 93   9.0248736 13.31780180   Rest
#> 94  11.2695343  9.02498422   Rest
#> 95   7.8812742 15.35098735   Rest
#> 96   6.1890713 10.68148364   Rest
#> 97   0.8987138  2.51564531   Rest
#> 98  15.4697884 11.51093163   Rest
#> 99   8.9955489 13.89284227   Rest
#> 100  3.9778059  1.22416956   Rest
#> 101  3.9126510  9.44250267   Rest
#> 102  5.0166426  6.46770348   Rest
#> 103 15.2562531  6.57465427   Rest
#> 104  6.3152293  6.30705759   Rest
#> 105  1.0240373  3.67988539   Rest
#> 106  3.5812891  2.99616620   Rest
#> 107  6.6325866  0.49246783   Rest
#> 108 14.7843472  6.64437857   Rest
#> 109 11.4699233  2.93392281   Rest
#> 110 13.2362772  2.83636219   Rest
#> 111 12.0683058 12.84841810   Rest
#> 112  8.0345560  6.10300618   Rest
#> 113 14.5481911  4.96062004   Rest
#> 114  5.7438172  6.24038634   Rest
#> 115  2.7099814  1.58746194   Rest
#> 116 12.7402573  2.42072624   Rest
#> 117  7.6591263  0.51346128   Rest
#> 118  3.9406822  8.71890688   Rest
#> 119 10.9270207 12.26759546   Rest
#> 120 14.5019244  4.49241619   Rest
#> 121  5.4995742 -0.06040412   Rest
#> 122 16.3225550 13.70009660   Rest
#> 123  5.2236292 13.59576943   Rest
#> 124  5.4912698  7.86010257   Rest
#> 125  7.5355963 11.89966584   Rest
#> 126  8.6706121 12.74917605   Rest
#> 127 12.7364575  6.15162338   Rest
#> 128  3.0526473  5.76264292   Rest
#> 129  2.8453145  7.35696397   Rest
#> 130 15.0510559 15.81362617   Rest
#> 131  4.1351252  7.52710856   Rest
#> 132  1.7024441  4.61183681   Rest
#> 133  8.6046983 14.78157679   Rest
#> 134  8.4607975 12.20850283   Rest
#> 135  4.2035389 15.95647274   Rest
#> 136  3.2772039  3.52561990   Rest
#> 137 12.8768010  3.36112166   Rest
#> 138 10.5139161 15.30746236   Rest
#> 139  8.1185519 13.91388609   Rest
#> 140  8.7331490  2.91476788   Rest
#> 141  9.2796392  9.28071297   Rest
#> 142  2.4178422 11.24630288   Rest
#> 143  8.9931665 14.26722027   Rest
#> 144  8.6620273 15.24896750   Rest
#> 145  3.6147939  1.63256906   Rest
#> 146  3.0815860 10.24764780   Rest
#> 147 17.5199969 16.72779178   Rest
#> 148  2.0441848  0.91756999   Rest
#> 149  0.6065012  0.32843927   Rest
#> 150 12.5185804  2.32234146   Rest
#> 151  1.7990144  7.60439288   Rest
#> 152 -0.1187861  2.92229492   Rest
#> 153  4.7177156 14.07417340   Rest
#> 154 12.0861577 10.77022403   Rest
#> 155 13.6106933  7.62203926   Rest
#> 156 10.3686375  5.10920400   Rest
#> 157 13.3350840 15.37643138   Rest
#> 158  5.0222180 13.12473669   Rest
#> 159 12.1590534  9.55389255   Rest
#> 160 13.4489865 14.09892954   Rest
#> 161  4.4577854 10.34745204   Rest
#> 162  4.7146334 12.39141545   Rest
#> 163 10.2583710 10.36128752   Rest
#> 164 14.4836440  2.47194776   Rest
#> 165  5.3458955  1.72988909   Rest
#> 166  9.5687266  7.17669772   Rest
#> 167 -1.4754921 13.48030716   Rest
#> 168  8.7146848 11.61564185   Rest
#> 169  5.4968908  2.04672237   Rest
#> 170  9.0117297 11.08836558   Rest
#> 171 11.9523896 14.13290951   Rest
#> 172 10.2655337 11.98816941   Rest
#> 173  6.2465545 10.14434416   Rest
#> 174  1.1975158  7.93248156   Rest
#> 175  6.7013944  4.27230960   Rest
#> 176  7.5135623 11.08488139   Rest
#> 177  9.1325910  3.25072096   Rest
#> 178 12.3217339 10.01967437   Rest
#> 179 14.2564284  9.48319928   Rest
#> 180  1.1568334  2.66403131   Rest
#> 181  6.0265105  1.68246813   Rest
#> 182  4.7675335  6.92311495   Rest
#> 183  9.5849765  5.24008063   Rest
#> 184 11.3784627  2.65566965   Rest
#> 185 -0.2356738 18.42265915   Rest
#> 186  6.0852939  1.30405229   Rest
#> 187 11.9204077 14.62212276   Rest
#> 188 13.3715327 11.98340842   Rest
#> 189  3.9680621  2.09799048   Rest
#> 190 13.8507868 13.63021709   Rest
#> 191  6.3388256  1.53799484   Rest
#> 192 15.2617513  3.77632066   Rest
#> 193  2.9523907 12.42372366   Rest
#> 194  3.5197886  3.34039378   Rest
#> 195 13.4014100  7.74243754   Rest
#> 196  4.2737871  4.08651479   Rest
#> 197  6.5480139 13.23359864   Rest
#> 198 13.0716753  1.30443355   Rest
#> 199 11.6189947 12.64139872   Rest
#> 200 -0.8307149  2.00512992   Rest
#> 201  2.3942955  2.98660497   Rest
#> 202 11.0837601  2.32808040   Rest
#> 203  4.9524186  9.87176345   Rest
#> 204  4.1358926  7.10706554   Rest
#> 205  6.5757450 12.27596313   Rest
#> 206 10.3455010  5.88919350   Rest
#> 207  3.3698651 11.41262575   Rest
#> 208  1.4337774  6.00516093   Rest
#> 209  9.2013376  2.01062333   Rest
#> 210 11.2177224  4.61241323   Rest
#> 211  5.2434293  4.07280300   Rest
#> 212 10.9243912  3.07632974   Rest
#> 213  4.1837927  7.57720545   Rest
#> 214 14.5650548 10.15864711   Rest
#> 215  2.6415229  9.30123688   Rest
#> 216  9.8538947  3.99670119   Rest
#> 217 11.1868632  0.24560474   Rest
#> 218 14.1324935 13.97993571   Rest
#> 219 11.2776673 14.59632673   Rest
#> 220 13.6039249 12.59484806   Rest
#> 221 10.0217362  6.58125275   Rest
#> 222 14.1971461  2.69058857   Rest
#> 223 13.7731290  8.91378767   Rest
#> 224 10.9844814  2.95323776   Rest
#> 225 13.3953216  4.06438832   Rest
#> 226  3.8645591  4.36163450   Rest
#> 227 13.9779557  5.15623010   Rest
#> 228  2.6487183  8.09294901   Rest
#> 229  4.6760472 12.88857427   Rest
#> 230 12.9387850 13.60668883   Rest
#> 231  5.9698862  9.54813781   Rest
#> 232 12.5271865 13.24356366   Rest
#> 233  6.6792000  1.35886725   Rest
#> 234 15.5940843  8.79089772   Rest
#> 235 11.0810956  3.57014677   Rest
#> 236 12.6091889 17.31533586   Rest
#> 237  6.7260757  7.88907711   Rest
#> 238  4.1946524  3.23548284   Rest
#> 239 10.4150973 13.28060733   Rest
#> 240  0.1610534  6.44648888   Rest
#> 241  5.9935478 14.57876085   Rest
#> 242 13.3097609  1.71095034   Rest
#> 243  2.5916843 15.35763036   Rest
#> 244  2.7996188 15.40523986   Rest
#> 245  3.3810403  1.49205569   Rest
#> 246  4.6678653 14.76511107   Rest
#> 247  2.3946734 14.25678760   Rest
#> 248  4.9933655 15.33607520   Rest
#> 249  6.6294636  1.66907231   Rest
#> 250  8.3279163 12.51811896   Rest
#> 251 -0.1123828  5.22081505   Rest
#> 252  0.6525841  6.26746622   Rest
#> 253 13.3930632  7.33208149   Rest
#> 254  4.5488511 11.74364344   Rest
#> 255  9.2298661  7.17249842   Rest
#> 256  3.9148795  7.35998190   Rest
#> 257 -0.3577051  3.20098583   Rest
#> 258 11.3182199  8.61208682   Rest
#> 259 -1.2207012  7.96506448   Rest
#> 260  2.4036438  3.36626987   Rest
#> 261  6.7984151 16.44549895   Rest
#> 262  4.4171207 11.08632069   Rest
#> 263  6.9591368 10.50905346   Rest
#> 264 15.5010008 12.00343801   Rest
#> 265  2.5898623 14.47386398   Rest
#> 266 16.2826875  8.17400640   Rest
#> 267  5.2129433 12.88292823   Rest
#> 268  3.9904648  2.94918605   Rest
#> 269  8.3533581  0.41474116   Rest
#> 270 17.0214740 10.71177891   Rest
#> 271  8.1692715  0.44053186   Rest
#> 272 11.2111712  6.16916997   Rest
#> 273  5.3619760 12.50057361   Rest
#> 274  8.3624510 16.13444260   Rest
#> 275 13.5206019 12.94796808   Rest
#> 276 12.0701983 14.23293425   Rest
#> 277  6.8650086  4.00743118   Rest
#> 278  6.6662623  7.93399110   Rest
#> 279 13.4663160  9.49840917   Rest
#> 280 14.8193472  6.96583033   Rest
#> 281 10.4360467  9.40199747   Rest
#> 282 15.1803193  8.70225097   Rest
#> 283 11.9830031  9.02160248   Rest
#> 284  7.1362669 -1.49878858   Rest
#> 285  3.2939914 14.79684476   Rest
#> 286  1.9029607  2.93271059   Rest
#> 287  5.8619622 11.75821164   Rest
#> 288  8.7784006 13.92509286   Rest
#> 289 12.2587667  6.47117012   Rest
#> 290  4.4763720  4.92244679   Rest
#> 291 14.6211177 13.35152423   Rest
#> 292 12.4418256 14.35453901   Rest
#> 293 15.4177092  7.77950249   Rest
#> 294  8.9019582 13.54165039   Rest
#> 295 15.5906343  7.68530620   Rest
#> 296  8.2036606 11.53224035   Rest
#> 297  7.4432491 14.41636102   Rest
#> 298  3.6628649 11.06673948   Rest
#> 299  1.1530460 12.76881672   Rest
#> 300  6.1257731  5.54606372   Rest
#> 301  9.6695409  0.02800136   Rest
#> 302  8.0419929 15.78545155   Rest
#> 303 13.6768355 -0.59049069   Rest
#> 304 14.4591293  4.13210599   Rest
#> 305  1.0927210  3.73001776   Rest
#> 306 12.2946206 12.51711533   Rest
#> 307  7.3835032 10.47394238   Rest
#> 308 14.3292186  4.03133412   Rest
#> 309  6.8114411 12.21957123   Rest
#> 310  2.2343260 10.53008083   Rest
#> 311  6.7016633  3.08337764   Rest
#> 312  5.6750200 12.94292163   Rest
#> 313 14.1342080  0.12270367   Rest
#> 314  7.9977267 10.00822064   Rest
#> 315  1.7887042  4.46448541   Rest
#> 316  4.0970586 10.66539938   Rest
#> 317  8.4317395 14.26831708   Rest
#> 318 13.6283298  5.98518687   Rest
#> 319  1.6936079  9.16857017   Rest
#> 320 13.2507974  4.42181556   Rest
#> 321  1.4829916 15.63777745   Rest
#> 322  1.5506194 16.95740890   Rest
#> 323  6.7012733  2.05521909   Rest
#> 324  4.7146028 11.44934110   Rest
#> 325 12.2925385  9.85437837   Rest
#> 326  1.4953534 14.69864904   Rest
#> 327  7.7932290  3.93753716   Rest
#> 328  6.0674180  3.87980664   Rest
#> 329 12.2690467  4.25171332   Rest
#> 330  9.2855071  1.64029363   Rest
#> 331  1.0711828  6.25310693   Rest
#> 332  7.7058392  2.65009799   Rest
#> 333 -0.8155140  9.75383501   Rest
#> 334 14.9562989  4.91124218   Rest
#> 335  6.7730838  6.60281789   Rest
#> 336  7.1228485 13.03431206   Rest
#> 337  6.7642862  3.37074728   Rest
#> 338  9.2572419 13.44436863   Rest
#> 339 10.7137396  9.96205101   Rest
#> 340  7.8090033  8.58177791 Circle
#> 341  8.4609920  5.48571056 Circle
#> 342  8.2011925  9.95161698 Circle
#> 343  5.5241543  7.65463701 Circle
#> 344  5.4549693  5.30186541 Circle
#> 345  7.4918909  8.71755413 Circle
#> 346  6.3899878  9.23204529 Circle
#> 347  8.2841498  6.14424035 Circle
#> 348  6.6454317  7.07915933 Circle
#> 349  8.5337277  7.12076340 Circle
#> 350  8.4860229  5.34949756 Circle
#> 351  8.7910304  9.38227874 Circle
#> 352  5.3214629  7.68585640 Circle
#> 353  9.1089616  9.62185781 Circle
#> 354  4.5808833  7.78004613 Circle
#> 355  5.7116287  6.13015062 Circle
#> 356  9.3481100  8.59602897 Circle
#> 357  9.4524702 10.54195884 Circle
#> 358  8.3814787  9.73417149 Circle
#> 359  8.5537855  8.23408693 Circle
#> 360  9.0311377  7.52124751 Circle
#> 361  8.0589452  6.47219040 Circle
#> 362  8.1482593  5.54881040 Circle
#> 363 10.9946514  9.14060661 Circle
#> 364  6.4611856  5.63939746 Circle
#> 365  7.3745018 11.72037857 Circle
#> 366  8.1250307 10.82583428 Circle
#> 367  9.7406697  6.01220703 Circle
#> 368  4.3127566  7.90160176 Circle
#> 369  6.2444398  8.08691799 Circle
#> 370  9.0686131  6.29978010 Circle
#> 371  7.0197395  8.27575073 Circle
#> 372  7.9825852  8.51534150 Circle
#> 373  8.9836482  7.10169998 Circle
#> 374  9.1679132  7.34739798 Circle
#> 375 10.6746538  7.99976210 Circle
#> 376  8.1594320 10.65969892 Circle
#> 377  8.7557818  9.25796618 Circle
#> 378  8.9698889  6.73515095 Circle
#> 379  6.5988040  7.34890628 Circle
#> 380  8.9840704  6.04233368 Circle
#> 381  9.0393368  9.53257032 Circle
#> 382  8.2665713  6.00579296 Circle
#> 383  6.3893879  9.10675654 Circle
#> 384  8.2254452  6.87930955 Circle
#> 385  6.9449617  6.26785053 Circle
#> 386 11.2432344  7.18109801 Circle
#> 387  8.3850366  7.14260155 Circle
#> 388  6.3527121  7.45887086 Circle
#> 389  8.8474029  9.69638986 Circle
#> 390 10.9998999  8.53650039 Circle
#> 391  8.5673763  9.46957209 Circle
#> 392  5.7158764  9.93349830 Circle
#> 393  4.6983802  8.08846068 Circle
#> 394  8.3012347 10.50845412 Circle
#> 395  6.0765866  9.82761965 Circle
#> 396  8.4998812  8.24692512 Circle
#> 397  8.5922158  7.55545526 Circle
#> 398 10.1184790 10.89732974 Circle
#> 399  9.6061873 10.62330365 Circle
#> 400  4.1648405  7.74021146 Circle
#> 401  8.5437081  7.28200498 Circle
#> 402  5.9094686  8.57184248 Circle
#> 403  9.0257504  6.71120831 Circle
#> 404 10.5363190  8.96203561 Circle
#> 405  8.3184634  9.58964424 Circle
#> 406  6.0285704  7.94610693 Circle
#> 407 10.0653797  9.81187846 Circle
#> 408  9.2459054  8.99652746 Circle
#> 409  9.6019467 10.05951332 Circle
#> 410  8.8622226  6.75148045 Circle
#> 411  7.1847356  9.80061361 Circle
#> 412 10.4078513 10.19354379 Circle
#> 413  7.5521369  7.94047812 Circle
#> 414  5.8551788  7.94918013 Circle
#> 415  5.8163312  9.01280484 Circle
#> 416  5.8714963  6.36928710 Circle
#> 417  6.8926295  8.30335262 Circle
#> 418  6.0926988  7.77726291 Circle
#> 419  8.1747033  8.28010501 Circle
#> 420  8.4685765  7.09105274 Circle
#> 421  6.7942016  5.92490846 Circle
#> 422  9.1226116  5.72673578 Circle
#> 423  5.5649282  8.04858619 Circle
#> 424  5.3757416  9.22753983 Circle
#> 425  9.6072720  8.92365268 Circle
#> 426 10.3241552  8.36184199 Circle
#> 427  8.0384988  7.35450414 Circle
#> 428  7.0568832 11.18043787 Circle
#> 429  7.8893144 10.22660646 Circle
#> 430  9.2292935  9.17240685 Circle
#> 431  6.5462054  7.71853230 Circle
#> 432  8.5741912  5.97708667 Circle
#> 433 10.8547519 10.86577051 Circle
#> 434  9.8553797  9.28101339 Circle
#> 435  7.2536981  6.38175261 Circle
#> 436  9.4077815  6.23637594 Circle
#> 437  8.8285716  6.04987106 Circle
#> 438  7.8049584 10.32590596 Circle
#> 439  8.1432995  7.02364203 Circle
#> 440  4.8113841 10.60298657 Circle
#> 441  9.4345837  8.21249911 Circle
#> 442  8.5684821  9.21610838 Circle
#> 443  9.4758910  7.09600018 Circle
#> 444  6.6219325  7.20702734 Circle
#> 445  7.5535601  4.96489487 Circle
#> 446  4.7985457  7.22275729 Circle
#> 447  6.0089344  8.42684056 Circle
#> 448  9.0605894 10.10347234 Circle
#> 449  9.3393419  5.92956310 Circle
#> 450  8.9677284  7.26893173 Circle
#> 451  5.7015258  6.72133648 Circle
#> 452  7.2544476  8.58010593 Circle
#> 453  5.8857111  6.44188593 Circle
#> 454  8.9183546 10.21723982 Circle
#> 455  6.5567379  6.30333160 Circle
#> 456  5.5587337  7.43450319 Circle
#> 457  6.0405728 10.30459494 Circle
#> 458  9.5321985 10.00654798 Circle
#> 459  7.0240115  8.99694358 Circle
#> 460 11.1671705  5.71921262 Circle
#> 461  6.7633429  9.16814249 Circle
#> 462  5.7342553  6.23836353 Circle
#> 463  7.3301206  9.80226219 Circle
#> 464  8.2927582  6.40237846 Circle
#> 465  7.3187953  6.76002375 Circle
#> 466  8.5863833  7.48517016 Circle
#> 467  5.4566559  8.00119091 Circle
#> 468  8.2352051 11.01902288 Circle
#> 469  7.5922381  6.03460468 Circle
#> 470  8.3452159  7.74382569 Circle
#> 471  6.6971146  8.76466359 Circle
#> 472  7.7708045 10.43991431 Circle
#> 473  8.2714684  8.70094380 Circle
#> 474  5.0905147  8.33655561 Circle
#> 475  6.6054663  9.32808373 Circle
#> 476 10.3616447  8.04007035 Circle
#> 477  6.1171582  5.29945384 Circle
#> 478  9.4369066  6.98721220 Circle
#> 479 10.1425370  6.55335054 Circle
#> 480  7.6432837  7.57507298 Circle
#> 481  7.0680851  7.46351343 Circle
#> 482 10.2227530  9.83496222 Circle
#> 483  8.9966029  4.92059252 Circle
#> 484 10.4272990  7.85813252 Circle
#> 485  8.0584877  6.77034766 Circle
#> 486  5.1002414  6.52969493 Circle
#> 487  9.1589833  4.63898207 Circle
#> 488  7.0002130  5.57730416 Circle
#> 489  9.3842899  9.92092616 Circle
#> 490  8.8052597  8.00104652 Circle
#> 491  9.4626046  9.49972306 Circle
#> 492 11.2960404  6.38804109 Circle
#> 493  6.3508561  6.09489936 Circle
#> 494  7.2194330  7.96066368 Circle
#> 495 10.4485451  9.05716350 Circle
#> 496  7.2207859  9.72770461 Circle
#> 497  8.0664017  8.29316730 Circle
#> 498 10.1065628  9.04827454 Circle
#> 499 10.2198160  6.99561694 Circle
#> 500  9.1948824  8.51915697 Circle
#> 501  4.8059582  9.39708197 Circle
#> 502  7.5653076  6.97839246 Circle
#> 503  9.4919900  7.73640366 Circle
#> 504  4.3963610  8.23332249 Circle
#> 505  5.8371838  8.21485401 Circle
#> 506  9.3820705 10.33601233 Circle
#> 507  8.5997580  6.59001593 Circle
#> 508  5.6006399  8.68806419 Circle
#> 509  5.2783164  7.73062356 Circle
#> 510  6.3248475  8.33573187 Circle
#> 511  5.8100336  7.80919292 Circle
#> 512  9.1239845 10.79008818 Circle
#> 513  7.5024758 10.33880886 Circle
#> 514  6.6039486 10.72998714 Circle
#> 515  6.7082642  8.43709765 Circle
#> 516  5.2965912  8.65621559 Circle
#> 517  7.9251684  6.37568363 Circle
#> 518  9.2025161  7.71421853 Circle
#> 519  9.2262672  5.90069940 Circle
#> 520  8.9721829  7.74422642 Circle
#> 521  6.7939112  8.29413053 Circle
#> 522  8.9108761  5.75096874 Circle
#> 523  8.7345788  6.67323533 Circle
#> 524  8.4577687  8.15101458 Circle
#> 525  6.3661165  7.38181956 Circle
#> 526  8.1187840  6.13880132 Circle
#> 527  7.0251659  7.13121198 Circle
#> 528  5.9179135  8.06533655 Circle
#> 529  9.4525878  9.47447877 Circle
#> 530 11.1932236  7.35679870 Circle
#> 531  7.0100297  9.27370782 Circle
#> 532  9.9238721  9.17903589 Circle
#> 533  7.6045085  8.43678505 Circle
#> 534 10.3064642 10.34854807 Circle
#> 535  6.4705470  7.83361809 Circle
#> 536  7.0065779  8.60953691 Circle
#> 537  8.5242545  8.23887412 Circle
#> 538  8.2443337 10.12292704 Circle
#> 539  6.7040987  9.99643948 Circle
#> 540  8.1984636  7.72939112 Circle
#> 541 10.1394725 10.39760485 Circle
#> 542 10.2526244  6.72079628 Circle
#> 543  7.0250772  7.38538497 Circle
#> 544  7.0776179  9.81729558 Circle
#> 545 10.8771400  7.84038366 Circle
#> 546  7.3628524  9.49234155 Circle
#> 547  8.0589311  9.54133428 Circle
#> 548  7.1127099  8.80846591 Circle
#> 549  7.7474529  9.44488310 Circle
#> 550  7.1940135  6.10869276 Circle
#> 551  7.5651485  5.11043693 Circle
#> 552 10.5779441 10.77360198 Circle
#> 553  8.6250226  7.81531430 Circle
#> 554  6.6978794  6.65583652 Circle
#> 555 10.2633896  8.52926866 Circle
#> 556  9.3969276  9.96951359 Circle
#> 557  5.8897041  6.71421675 Circle
#> 558  7.7077180  6.70981124 Circle
#> 559  7.2695619  9.56467251 Circle
#> 560  7.1811120  7.27329330 Circle
#> 561  9.3672510  9.89095010 Circle
#> 562  7.8004141  9.46051560 Circle
#> 563  4.1000400  7.88123089 Circle
#> 564  7.2000888  8.84728270 Circle
#> 565  9.7131027 10.99960548 Circle
#> 566  5.5923252  6.81304672 Circle
#> 567  7.3013825  7.21509878 Circle
#> 568  9.4810531  6.40182819 Circle
#> 569  7.2798472  9.46017468 Circle
#> 570  4.4458407  6.39304205 Circle
#> 571  7.9542358 11.36927739 Circle
#> 572  9.8809016  7.99743407 Circle
#> 573  8.5638275  6.72293789 Circle
#> 574  6.3365941 10.03122969 Circle
#> 575  6.3655814  6.83357980 Circle
#> 576  4.5686155  7.07222360 Circle
#> 577  7.7509353 10.82709525 Circle
#> 578  5.9736276 11.31114564 Circle
#> 579  5.7833666  9.39483607 Circle
#> 580  7.8441691  9.08982361 Circle
#> 581  6.8969037  7.85692070 Circle
#> 582  6.2680540  7.57562340 Circle
#> 583  6.8706259  8.57866208 Circle
#> 584  8.2743713  7.08819894 Circle
#> 585  8.6972288  9.49567877 Circle
#> 586  8.8666593  6.24489834 Circle
#> 587  9.0979999  7.48488163 Circle
#> 588  6.6583048  7.13385041 Circle
#> 589  5.9581081  8.10921848 Circle
#> 590  7.3910863  7.59348408 Circle
#> 591  5.5182159 11.11296005 Circle
#> 592  5.6585691  7.73351824 Circle
#> 593  7.4929217  9.66375177 Circle
#> 594  8.3265133  7.93190237 Circle
#> 595  4.8706618  7.02145978 Circle
#> 596  9.8907678  8.52823204 Circle
#> 597  7.5280481  8.04948405 Circle
#> 598  8.9053515  9.07562468 Circle
#> 599  5.3978729  9.30646617 Circle
#> 600  5.3791585  6.28822377 Circle
#> 601  5.6605189  8.75423640 Circle
#> 602  9.1062813  8.20681819 Circle
#> 603 10.0333398 10.13274743 Circle
#> 604  6.7530062  8.05761767 Circle
#> 605 11.1577664  8.12652436 Circle
#> 606  9.7102648  6.17387058 Circle
#> 607  6.0893011  8.52490593 Circle
#> 608  7.3662921  9.62778009 Circle
#> 609  8.8598966  9.34945023 Circle
#> 610  9.9035027  6.66961207 Circle
#> 611  5.8475803  8.97005606 Circle
#> 612  6.4236958  9.46138310 Circle
#> 613  8.7138873  7.40584996 Circle
#> 614  7.0075438  8.01991377 Circle
#> 615  9.1298232 11.51525936 Circle
#> 616  7.5562264 11.78977723 Circle
#> 617  9.1509954  7.97545106 Circle
#> 618  8.7460360  9.69320909 Circle
#> 619 10.7665943  9.69700980 Circle
#> 620  8.7104897  8.94321279 Circle
#> 621  8.2357251  6.70889961 Circle
#> 622  4.6250391  6.01367911 Circle
#> 623  7.3079351 10.18367348 Circle
#> 624  8.0908810  7.69768597 Circle
#> 625  8.6795842  5.18700376 Circle
#> 626  5.4769092  8.63180927 Circle
#> 627  7.9761532  9.09238269 Circle
#> 628  7.9112295  8.86899438 Circle
#> 629  6.0063228  5.27403134 Circle
#> 630  9.4369144  6.52200784 Circle
#> 631 10.3649361  7.60990458 Circle
#> 632  5.2904430  8.32018975 Circle
#> 633 10.2483053  9.73016569 Circle
#> 634 10.9297433  7.63143086 Circle
#> 635  6.6441473  6.82635314 Circle
#> 636 10.8455514  8.04327712 Circle
#> 637  9.6971261  5.56576861 Circle
#> 638  6.6230027  8.61857730 Circle
#> 639  5.0019237  5.85974092 Circle
#> 640  4.9441767  7.27620989 Circle
#> 641  5.8131820 11.33154673 Circle
#> 642  9.6620111  9.60251436 Circle
#> 643  9.4551421  8.93481951 Circle
#> 644  7.7659668  7.46126000 Circle
#> 645  4.5208026  6.95179719 Circle
#> 646  7.1953538 10.19743171 Circle
#> 647  8.9599056  9.76815114 Circle
#> 648  8.7314818  6.05434438 Circle
#> 649  7.6330353  7.45514636 Circle
#> 650  7.2060974  9.44308925 Circle
#> 651  6.9324212  9.07497964 Circle
#> 652 10.9968506  6.35664819 Circle
#> 653  7.5230653  9.08182638 Circle
#> 654  5.9727516  6.13013666 Circle
#> 655 10.2278627  6.00885342 Circle
#> 656  7.5775653  5.51140324 Circle
#> 657  7.1764726  7.56093113 Circle
#> 658  8.7429436  9.20365467 Circle
#> 659  7.7333581  6.66133397 Circle
#> 660  5.0852565  6.93424108 Circle
#> 661  7.7495963  9.78754392 Circle
#> 662  9.1195713  6.52623179 Circle
#> 663 10.8456323  8.22493695 Circle
#> 664  7.1209145  8.19681028 Circle
#> 665  6.2937935  8.45309909 Circle
#> 666  7.6950765  9.36845552 Circle
#> 667  6.5678046  8.43211736 Circle
#> 668  8.0746941 11.44921456 Circle
#> 669  7.7886750  8.33968365 Circle
#> 670  6.4050456  8.65281770 Circle
#> 671 10.3355871  9.34312607 Circle
#> 672  7.8380130  5.75202449 Circle
#> 673  9.1212804 10.56363980 Circle
#> 674  6.7550139  6.76038672 Circle
#> 675  9.5884181  9.89666580 Circle
#> 676  7.7477326  8.28358781 Circle
#> 677  9.5241415  7.52213224 Circle
#> 678 11.0554840  9.86171968 Circle
#> 679  6.1604554  8.23411366 Circle
#> 680  9.7150999  6.53053249 Circle
#> 681  6.2890596  6.98076948 Circle
#> 682  8.9395772 10.30211397 Circle
#> 683  7.9903582  9.87887092 Circle
#> 684  7.3689851  7.21589681 Circle

rose(circle_example[, c("x", "y", "class")], var = "class", over_ratio = 0.8)
#>                x           y  class
#> 1    6.019889452  4.83432935   Rest
#> 2    7.014456532 14.42162839   Rest
#> 3   12.354142661 10.82393440   Rest
#> 4    1.942391744  9.59048703   Rest
#> 5    3.188168922  6.60552519   Rest
#> 6   10.705273177 14.30892204   Rest
#> 7    4.720016004  2.68496624   Rest
#> 8    8.929548596  1.66461561   Rest
#> 9    6.096904057  1.65776169   Rest
#> 10   2.143243185  7.62365293   Rest
#> 11   3.154054135  6.62750446   Rest
#> 12  13.473989788 15.72568724   Rest
#> 13   6.590345029 13.70618819   Rest
#> 14  11.618900619 10.06354433   Rest
#> 15   5.513714512 15.04750357   Rest
#> 16  12.662864451  8.19384257   Rest
#> 17  12.368291442  5.13720221   Rest
#> 18   6.732267824  2.96590513   Rest
#> 19  -0.535108596  9.24132298   Rest
#> 20  12.837642476  7.11836469   Rest
#> 21  12.896341437  1.06214089   Rest
#> 22   8.770703089 10.20560573   Rest
#> 23   8.781161999 13.91994084   Rest
#> 24   0.388312100  9.71030859   Rest
#> 25   9.417094655 13.10055148   Rest
#> 26  15.005504237  4.68043633   Rest
#> 27   9.204953168  1.99131514   Rest
#> 28   0.764945369  9.94400517   Rest
#> 29  14.709990898  6.14655564   Rest
#> 30   9.074901445 11.67425464   Rest
#> 31  10.939634714 16.38736726   Rest
#> 32  13.015175736 11.04615523   Rest
#> 33  10.455785138  7.13138449   Rest
#> 34  12.531230078  1.23888624   Rest
#> 35   4.487555255  5.10594864   Rest
#> 36   0.935850538 12.26302617   Rest
#> 37  10.471498900 14.91000443   Rest
#> 38   2.973073545 11.83071573   Rest
#> 39   3.031759137  3.05598274   Rest
#> 40   7.961298008  4.25324365   Rest
#> 41   9.545385731  5.46306993   Rest
#> 42  12.458770762  8.62796425   Rest
#> 43  12.834760085 10.21678304   Rest
#> 44  15.650747452 13.68697637   Rest
#> 45   3.121452520 14.99484588   Rest
#> 46   7.474943733 14.15147987   Rest
#> 47  12.968506672 11.70497950   Rest
#> 48   9.858170542 10.86939073   Rest
#> 49   2.126062722  7.31615197   Rest
#> 50  -0.232531525  2.88542025   Rest
#> 51   2.409530371  5.70107579   Rest
#> 52   5.765692394  7.96814354   Rest
#> 53  11.231563354  1.50108739   Rest
#> 54   6.066714981 13.56671237   Rest
#> 55   4.726906712  7.62289235   Rest
#> 56  12.148190548 13.61056017   Rest
#> 57  11.624046352  1.85443008   Rest
#> 58   7.127155150  0.31480777   Rest
#> 59  12.801444053 11.24232534   Rest
#> 60  10.338143415  6.44501468   Rest
#> 61   6.134957127 13.87288524   Rest
#> 62  12.243154779  2.84011152   Rest
#> 63   4.688064568  2.37503219   Rest
#> 64   1.992783764  1.21043732   Rest
#> 65   2.339778453  3.87867911   Rest
#> 66   8.058194894 12.79222462   Rest
#> 67  -0.863071904 -0.48283828   Rest
#> 68   5.391607613  5.14974087   Rest
#> 69   1.866440135 10.57383278   Rest
#> 70   2.844308093  4.85933302   Rest
#> 71   2.919948145  0.05929653   Rest
#> 72  15.248747055 -0.40312015   Rest
#> 73   9.141436877 12.48641857   Rest
#> 74  10.437182654  1.48011258   Rest
#> 75   5.207871289  8.78357858   Rest
#> 76   4.650009813  8.47054986   Rest
#> 77   5.989750276 10.05781217   Rest
#> 78  10.215177954 15.31684911   Rest
#> 79  11.508979330 14.35516925   Rest
#> 80   1.963450752  4.88109271   Rest
#> 81   0.401782523  8.95125913   Rest
#> 82   7.607630073  4.81705765   Rest
#> 83   2.206969999 11.82246390   Rest
#> 84   9.391755121  3.31080815   Rest
#> 85  13.738525080  8.84613713   Rest
#> 86  12.842434191  8.25779864   Rest
#> 87   2.519094554  3.82704751   Rest
#> 88  10.935813715  4.36717199   Rest
#> 89   0.947019833  4.17761919   Rest
#> 90   1.224106022  8.47212120   Rest
#> 91  15.148430027  9.52078525   Rest
#> 92   9.317679146  9.14079548   Rest
#> 93   6.908446223  3.52761214   Rest
#> 94   8.169764373  0.24490272   Rest
#> 95  15.694019699 15.56668223   Rest
#> 96   3.466108550  1.51255291   Rest
#> 97   5.855700113  1.46489084   Rest
#> 98   5.836257247 11.71093098   Rest
#> 99  -0.001318688 13.80103789   Rest
#> 100  8.980678125  8.66160988   Rest
#> 101  6.227192633  1.70348477   Rest
#> 102  7.262864011 16.70907839   Rest
#> 103 12.364488308  5.75021567   Rest
#> 104  5.038094383 13.73904860   Rest
#> 105  3.213745665  5.24343759   Rest
#> 106 15.037219236  4.13969715   Rest
#> 107  5.043446148  3.94683910   Rest
#> 108  5.860398757  0.54558040   Rest
#> 109  9.533611674 12.87548071   Rest
#> 110  5.953830070 12.26318951   Rest
#> 111  5.748938405 12.13739354   Rest
#> 112  0.208271595  6.54936956   Rest
#> 113 10.586671329  2.63003242   Rest
#> 114 11.193388440  1.75127162   Rest
#> 115  4.323659665 14.07019364   Rest
#> 116  0.257738356  4.27407260   Rest
#> 117  9.227410535 14.66360649   Rest
#> 118  1.786696879 14.04209149   Rest
#> 119  4.643335284 14.89696067   Rest
#> 120  2.810246731  6.00142299   Rest
#> 121 16.592175424  7.10903582   Rest
#> 122  3.199061522  7.10922155   Rest
#> 123  4.557124205  2.99411788   Rest
#> 124  5.658767066 12.42655369   Rest
#> 125  0.938635430 14.56678447   Rest
#> 126 11.552572524 14.55240348   Rest
#> 127 12.842386243  4.76221232   Rest
#> 128  5.012358089 10.81202846   Rest
#> 129  4.355359719  8.83003096   Rest
#> 130 14.327845781 13.84176669   Rest
#> 131  9.485833540 12.27757668   Rest
#> 132  1.117044320  1.81419513   Rest
#> 133  1.428837937 13.18151669   Rest
#> 134 10.719721738 11.91374245   Rest
#> 135 11.335077600  9.10954039   Rest
#> 136  9.520659716  6.59148771   Rest
#> 137 15.449621573  6.36317906   Rest
#> 138  6.891299754  1.67806434   Rest
#> 139 10.196897616  7.31574252   Rest
#> 140  6.345467097 15.32094688   Rest
#> 141  0.433934669  9.50423038   Rest
#> 142 15.231352042  2.30747386   Rest
#> 143  1.966110248  5.87641866   Rest
#> 144  7.762121155 13.47606693   Rest
#> 145  2.167449343 12.18433962   Rest
#> 146 11.828525057  9.96142002   Rest
#> 147  8.761663021 13.33235525   Rest
#> 148  9.913726921  3.68819643   Rest
#> 149  4.447247376 12.42458784   Rest
#> 150  2.521808117 10.75436451   Rest
#> 151  0.118429470  6.72811474   Rest
#> 152 15.143095224  2.89624015   Rest
#> 153 12.081957288  0.59098335   Rest
#> 154 11.899266769  3.98507792   Rest
#> 155 15.198597203 10.23108006   Rest
#> 156  4.262808965  2.11311271   Rest
#> 157  0.719115243 -1.95489225   Rest
#> 158 10.961457042 12.57208781   Rest
#> 159  2.119849530  3.43208062   Rest
#> 160 14.559010225  6.19726206   Rest
#> 161 12.642089606 13.09505112   Rest
#> 162  8.897735838 17.75772418   Rest
#> 163  0.835786758 12.78191862   Rest
#> 164  1.948085789  8.14481321   Rest
#> 165 12.190875445  0.45799199   Rest
#> 166  5.103737516  0.13647249   Rest
#> 167  4.924803757 15.57674470   Rest
#> 168  1.055490571  3.72643298   Rest
#> 169  8.574683418 14.08550248   Rest
#> 170 11.490488576  4.05664011   Rest
#> 171 10.161488910  6.12985069   Rest
#> 172  3.013172334 11.16596053   Rest
#> 173 13.363618538  5.20397782   Rest
#> 174  9.396752720 11.50348104   Rest
#> 175 10.449855794  5.16151528   Rest
#> 176 13.922222582  5.61746334   Rest
#> 177  9.283219884 15.84878810   Rest
#> 178  3.798495984 11.03185847   Rest
#> 179  2.295521014  4.94012869   Rest
#> 180 13.207816566 13.02476232   Rest
#> 181  8.579707422  5.83005970   Rest
#> 182  2.622624061  1.83431894   Rest
#> 183 -0.587543347  7.40669320   Rest
#> 184  2.273789542  3.12141126   Rest
#> 185 10.553088013 14.73978091   Rest
#> 186  8.426989759 17.08078656   Rest
#> 187 13.973648157  4.60958392   Rest
#> 188  3.430511297 12.42536361   Rest
#> 189 13.555879148  3.80520554   Rest
#> 190  0.562506279  0.24028485   Rest
#> 191 16.011364568 12.32334121   Rest
#> 192  1.899281751 17.16428640   Rest
#> 193  3.517627242 13.22174889   Rest
#> 194  6.752205457  4.52886681   Rest
#> 195 12.437829386  5.13605565   Rest
#> 196 15.270914802 10.20756056   Rest
#> 197 11.555247900  7.64518658   Rest
#> 198  9.619638707 10.23982675   Rest
#> 199 12.135542553  5.21019840   Rest
#> 200  2.945012249  5.00418282   Rest
#> 201 13.610596341 10.09446511   Rest
#> 202  2.441618553 12.07964208   Rest
#> 203 10.451512034  7.35299640   Rest
#> 204  3.454730679  9.25102094   Rest
#> 205 11.117156755 14.84921898   Rest
#> 206  1.190140648 13.68660145   Rest
#> 207  6.303797588  0.76403602   Rest
#> 208 15.117720250 10.50014377   Rest
#> 209 11.471754693  5.60036140   Rest
#> 210  6.174646196 12.15784870   Rest
#> 211  2.430038321  5.50193688   Rest
#> 212  9.265977213 15.23900949   Rest
#> 213  5.482291833 14.27993195   Rest
#> 214  2.482820598 15.93799298   Rest
#> 215  6.194703330 12.12634575   Rest
#> 216  9.350364425  6.17741232   Rest
#> 217  3.783299071 11.89913160   Rest
#> 218 11.779388268 12.29749224   Rest
#> 219  4.685361477  2.45536069   Rest
#> 220  3.165405963  2.06049666   Rest
#> 221  1.398203433 14.42721649   Rest
#> 222 12.748685090  6.38620343   Rest
#> 223  2.441607387  3.44944340   Rest
#> 224 11.711285505  4.04969809   Rest
#> 225 13.424592798  7.52132920   Rest
#> 226  4.674270419 14.11093884   Rest
#> 227 13.712929774 -0.16222456   Rest
#> 228  8.769416806  3.78970913   Rest
#> 229 14.131302175 15.33671315   Rest
#> 230  6.288188862 16.06820774   Rest
#> 231  8.894773319 -2.33366653   Rest
#> 232 12.334190314 13.27874010   Rest
#> 233  3.616296436 12.70353721   Rest
#> 234  7.016670625  4.06725645   Rest
#> 235  8.435361741 14.39431693   Rest
#> 236 11.413371183  2.35090326   Rest
#> 237 12.797003771  9.27045283   Rest
#> 238  6.918858592 11.55633929   Rest
#> 239  4.812953889  2.89960615   Rest
#> 240 11.194227979  2.15014517   Rest
#> 241 11.580122649  6.93277641   Rest
#> 242  3.281424598 13.30207569   Rest
#> 243  9.973841601 12.43898305   Rest
#> 244 15.765394282  9.58044518   Rest
#> 245 15.051864089  3.33936561   Rest
#> 246  1.799264604  8.18979093   Rest
#> 247 14.862239077  5.02256703   Rest
#> 248 15.547681652  3.91603230   Rest
#> 249  2.508861566  3.80951194   Rest
#> 250  5.031690873 10.25173727   Rest
#> 251 13.581980510  2.32663562   Rest
#> 252  6.780936171 -1.19772329   Rest
#> 253  5.040802914 12.71128509   Rest
#> 254 15.326964090  9.90428279   Rest
#> 255 14.990881033  4.13177595   Rest
#> 256  7.435592890 11.16596738   Rest
#> 257  2.447983601  4.14424797   Rest
#> 258  5.086287296  2.48302805   Rest
#> 259 10.849828177  8.34903031   Rest
#> 260 13.032241578  6.87705402   Rest
#> 261  9.550832478 13.39551828   Rest
#> 262  1.843978410 11.58863399   Rest
#> 263  5.670153383  0.72144416   Rest
#> 264 17.487022919  9.61204741   Rest
#> 265  3.912148658  3.83398378   Rest
#> 266  9.847955309  2.93315692   Rest
#> 267 13.470022145  5.27569786   Rest
#> 268  3.723358394 12.20649121   Rest
#> 269  8.939110112  9.54411811   Rest
#> 270  2.298105197  4.50066532   Rest
#> 271  6.107904060 12.89685938   Rest
#> 272 10.814673453  9.39230081   Rest
#> 273 11.361022613  6.96675677   Rest
#> 274  8.731036933  2.88282739   Rest
#> 275  4.702238711 -0.66150085   Rest
#> 276 15.306522090  5.15253872   Rest
#> 277  9.416546754 13.77372260   Rest
#> 278  6.846255087 11.61869999   Rest
#> 279 14.238090460  4.65976067   Rest
#> 280  4.181710594  8.10172349   Rest
#> 281 16.611786803  8.31783673   Rest
#> 282 12.109794005 -0.86514968   Rest
#> 283  4.407433678 12.33843205   Rest
#> 284 11.317311805  5.46369623   Rest
#> 285  3.428847257 12.31108432   Rest
#> 286 13.870513958 12.51406709   Rest
#> 287  4.643550544 12.47561039   Rest
#> 288  7.072900915 15.17740499   Rest
#> 289  6.525014740  0.34005951   Rest
#> 290  4.348803788  6.68178328   Rest
#> 291  0.839341641  4.97486278   Rest
#> 292 12.772251165  3.90967045   Rest
#> 293 10.980910490 11.35271849   Rest
#> 294  3.234483322  0.28717657   Rest
#> 295 14.160111710  2.96469220   Rest
#> 296  8.153270435  0.89982837   Rest
#> 297  6.106278186  4.25536098   Rest
#> 298  8.476987107 10.97658054 Circle
#> 299  5.195143452  6.60902950 Circle
#> 300  8.548662736  6.30974431 Circle
#> 301  8.068021325  8.90057844 Circle
#> 302  8.082808448  8.23578191 Circle
#> 303  6.784841801  6.95350261 Circle
#> 304  6.554746790  7.45312000 Circle
#> 305  6.960498000  9.75813162 Circle
#> 306  7.692549568  9.00629449 Circle
#> 307  4.700313181  7.24511879 Circle
#> 308  8.219781842  5.82679223 Circle
#> 309  6.129232366  7.30620119 Circle
#> 310  9.195779649 10.35641126 Circle
#> 311  5.795898433  6.83472064 Circle
#> 312  7.004122301  9.05162836 Circle
#> 313  8.720931601 10.43277860 Circle
#> 314 10.299483365 11.17467725 Circle
#> 315  5.487946936  7.07836357 Circle
#> 316  5.426531730  7.93978470 Circle
#> 317  9.363127809  9.27773719 Circle
#> 318  6.401744327  7.27482175 Circle
#> 319  8.155775186  7.89212390 Circle
#> 320  5.980173893  8.00050817 Circle
#> 321  7.718608843  8.11629929 Circle
#> 322  8.310343649  5.19884444 Circle
#> 323  5.572576860  6.34090343 Circle
#> 324  8.169161886 11.79998938 Circle
#> 325  7.356359098  9.40537250 Circle
#> 326  7.442611092  8.33945065 Circle
#> 327  7.930775793 10.52842704 Circle
#> 328  6.226501907  8.52908963 Circle
#> 329  5.359103723  9.53363937 Circle
#> 330  9.612154016  8.49869156 Circle
#> 331  5.418217154  8.50530875 Circle
#> 332  6.456603810  9.80037543 Circle
#> 333 10.982085585  5.85848134 Circle
#> 334  9.685238346 10.85375427 Circle
#> 335  5.756539272  8.98235828 Circle
#> 336  5.699025248  6.70363251 Circle
#> 337  9.285739922  7.82977233 Circle
#> 338  6.463166824  7.10910871 Circle
#> 339  9.159242175 11.46838830 Circle
#> 340  7.146116573 10.50495139 Circle
#> 341  9.692636772  8.18055487 Circle
#> 342  5.470140884  8.46942652 Circle
#> 343  6.683161398  9.27308355 Circle
#> 344  5.517135875  7.32079330 Circle
#> 345 10.541274367  6.86352991 Circle
#> 346  7.261199661  5.16261197 Circle
#> 347  9.682210551  7.41701888 Circle
#> 348  8.720008373  8.57499426 Circle
#> 349  6.712573116  8.37996602 Circle
#> 350  8.085300398  6.02208615 Circle
#> 351  8.800308593  4.16388487 Circle
#> 352  6.743638897  9.82773764 Circle
#> 353  9.619308112  8.15185689 Circle
#> 354 10.475158106  9.02325453 Circle
#> 355  7.843412532  7.04823105 Circle
#> 356  9.648812170 11.23150128 Circle
#> 357  7.563931125 10.67718435 Circle
#> 358  6.740909318 10.46816429 Circle
#> 359  5.972491882  8.09419138 Circle
#> 360  5.980397695  9.97148415 Circle
#> 361  7.858908929 10.78435895 Circle
#> 362  7.944483145  5.24194619 Circle
#> 363 10.975064589 10.35619626 Circle
#> 364  9.916203737  9.11007272 Circle
#> 365  5.018878604  8.79030098 Circle
#> 366  6.224940663  6.12086428 Circle
#> 367  6.338777483  7.25844372 Circle
#> 368  9.806384951  9.16668523 Circle
#> 369  8.560742822  5.50398908 Circle
#> 370 11.596717046  7.36141940 Circle
#> 371  6.312389579  9.79058610 Circle
#> 372  6.794598157  5.62051131 Circle
#> 373 10.047287515 10.65914789 Circle
#> 374  8.997098060  6.40228280 Circle
#> 375  9.442695063 10.68332596 Circle
#> 376  8.460995734  8.89110777 Circle
#> 377  5.389943757  9.26348509 Circle
#> 378 11.362233164 10.80698070 Circle
#> 379  5.563816177 10.10031478 Circle
#> 380  9.383504307  9.07915504 Circle
#> 381  5.253141773  8.45233011 Circle
#> 382  7.584605312  9.84340703 Circle
#> 383  8.493428133  7.48623907 Circle
#> 384  7.048719775  7.37638237 Circle
#> 385  6.778965877  7.65299781 Circle
#> 386  6.568868288 11.03256366 Circle
#> 387  4.026851163 10.73621307 Circle
#> 388 10.545976009  9.47659988 Circle
#> 389  6.931290179  8.04919171 Circle
#> 390  6.497237486  9.01552906 Circle
#> 391  6.311006764  9.51199693 Circle
#> 392  5.212041040  8.01652331 Circle
#> 393  7.335631933  8.58775883 Circle
#> 394  8.305981702  5.17073137 Circle
#> 395  5.264750205  8.40562477 Circle
#> 396  7.700374056  8.90913309 Circle
#> 397  7.756673372  3.79744543 Circle
#> 398 10.000628224  6.64767389 Circle
#> 399  5.975039534  9.41625262 Circle
#> 400  8.401196539  7.89564506 Circle
#> 401  9.205535579  6.28339887 Circle
#> 402 10.107633744 10.03156824 Circle
#> 403  7.181629940  8.21452337 Circle
#> 404  6.119036891  8.75821994 Circle
#> 405  8.274165627 10.99151243 Circle
#> 406  7.930600916  8.96546336 Circle
#> 407  7.880954009  6.36876940 Circle
#> 408  8.738339902  6.87382572 Circle
#> 409  6.913230111  6.80022724 Circle
#> 410  5.234670146  9.22738233 Circle
#> 411  8.740811794  7.62752287 Circle
#> 412  7.432389558  8.89933660 Circle
#> 413  5.984248530  6.00587049 Circle
#> 414  8.319965025  7.70100395 Circle
#> 415  5.319247102  9.02646367 Circle
#> 416  7.389728781  4.59779664 Circle
#> 417  8.931896281  9.21392855 Circle
#> 418 11.210128286  6.07622233 Circle
#> 419  6.174903565  9.59854699 Circle
#> 420  5.981151663  8.99570075 Circle
#> 421  5.722247492  7.97946385 Circle
#> 422  4.408298607  8.80815223 Circle
#> 423  4.538790596  6.21414614 Circle
#> 424  7.542032979 10.42848300 Circle
#> 425  7.229425311 10.26069639 Circle
#> 426  7.832154785  9.21663636 Circle
#> 427  8.372943993  9.37359515 Circle
#> 428  8.472838518  6.22508707 Circle
#> 429  6.543359939  7.58126915 Circle
#> 430  9.190810735  9.57696992 Circle
#> 431  8.260551447  7.30705427 Circle
#> 432  7.431681123  8.56244456 Circle
#> 433 10.816698120  8.46820196 Circle
#> 434  6.486902045  5.99952018 Circle
#> 435  8.620674950  9.75133730 Circle
#> 436  8.782533191  9.43578899 Circle
#> 437  7.656904908  6.03011503 Circle
#> 438  4.879804683  8.44041317 Circle
#> 439  8.681823019  9.72980293 Circle
#> 440  4.671532456  8.20735527 Circle
#> 441  7.365583557 10.41181275 Circle
#> 442  7.249956953  6.77592920 Circle
#> 443  9.330030559 10.52031526 Circle
#> 444  8.428759866  9.49956928 Circle
#> 445  7.103556796  6.23519894 Circle
#> 446  7.306030168  9.77805674 Circle
#> 447  8.194817235  9.11733830 Circle
#> 448  6.055913067 10.08865203 Circle
#> 449  8.335169594  6.82134824 Circle
#> 450  5.924777268  7.36382210 Circle
#> 451 11.702803375  7.71634985 Circle
#> 452 10.220536594  8.79294185 Circle
#> 453  7.179196079  5.50530251 Circle
#> 454  6.076334388  7.88083464 Circle
#> 455  8.619691574  8.82542165 Circle
#> 456  5.898627958  7.22845359 Circle
#> 457  8.406255941  7.65322100 Circle
#> 458  9.155371796 10.25034736 Circle
#> 459  9.785371857  8.26049049 Circle
#> 460  7.466939889  8.47550943 Circle
#> 461  8.781308738  9.81830622 Circle
#> 462  8.279501139  8.07926646 Circle
#> 463 10.937446838  9.92467005 Circle
#> 464  8.832361092  9.05053799 Circle
#> 465  4.262840762  7.69743777 Circle
#> 466  6.258536987  9.69397341 Circle
#> 467  8.170486925 10.17295031 Circle
#> 468  8.460844543  7.38206445 Circle
#> 469  8.045098029 10.37754289 Circle
#> 470  7.362462276  6.41727022 Circle
#> 471  5.568186258 10.01304420 Circle
#> 472  7.208285206  7.35359239 Circle
#> 473  7.945663842  7.88172946 Circle
#> 474  8.257821981 10.60021311 Circle
#> 475  4.422602346  9.30329396 Circle
#> 476  7.357093812  8.22147566 Circle
#> 477  9.907545861  7.57297594 Circle
#> 478  7.956143124  9.67196144 Circle
#> 479  7.170191613 10.18959224 Circle
#> 480 10.621867748  8.16817233 Circle
#> 481  9.620935343  8.56068480 Circle
#> 482 10.481420314  7.93330807 Circle
#> 483  8.950803877  9.66152365 Circle
#> 484 10.556466327  7.68023591 Circle
#> 485  7.992088502  5.87315681 Circle
#> 486  8.021771925  9.44928538 Circle
#> 487  9.460034201 10.63579152 Circle
#> 488 10.113758442  7.57553648 Circle
#> 489  6.383622648  9.16061485 Circle
#> 490  7.824870374 10.50127394 Circle
#> 491  5.060362156  9.09547659 Circle
#> 492  7.413091041 10.25991827 Circle
#> 493  6.241358480  8.20917454 Circle
#> 494  8.226692241  6.22737966 Circle
#> 495  8.472748478  6.07086390 Circle
#> 496  9.805447564  6.40722996 Circle
#> 497  7.303119109  8.99949079 Circle
#> 498  4.154480171  9.20708706 Circle
#> 499  8.432914966  9.38967670 Circle
#> 500  6.608114038 10.97691726 Circle
#> 501  9.587476419  6.07365183 Circle
#> 502  9.236997099  8.72268028 Circle
#> 503  7.883672669  8.04415848 Circle
#> 504  6.448251513  8.65346022 Circle
#> 505  9.356058782  6.35595245 Circle
#> 506  9.479529554  6.21173460 Circle
#> 507  8.891040595 10.03522611 Circle
#> 508  9.223628419  5.92179976 Circle
#> 509  7.437718599  9.13000991 Circle
#> 510  9.371503158  8.60570552 Circle
#> 511  9.110015608  8.72011170 Circle
#> 512  4.044552423  6.20658214 Circle
#> 513  8.053460277 10.36998019 Circle
#> 514  6.103259282  8.07896161 Circle
#> 515  8.725541679 10.24093329 Circle
#> 516  9.601204282 10.55647475 Circle
#> 517  5.962836198  8.64443612 Circle
#> 518  8.905927461  8.46503719 Circle
#> 519  7.028164249  8.83804394 Circle
#> 520 11.090148129  6.64113482 Circle
#> 521  7.027613615  5.49547917 Circle
#> 522  5.581668938 10.45576606 Circle
#> 523  9.385784866  9.68805914 Circle
#> 524  7.084099069  8.98072533 Circle
#> 525  8.436784519  4.49163509 Circle
#> 526  7.096547427  6.03129855 Circle
#> 527  7.541814089 10.44356966 Circle
#> 528  7.945379207  7.67992693 Circle
#> 529  9.752349669  5.50550103 Circle
#> 530  4.874609947  6.46760713 Circle
#> 531  8.929148607  8.96847765 Circle
#> 532 10.042741063 10.58889549 Circle
#> 533  7.647485518  9.05654044 Circle
#> 534  6.484901538  8.09940760 Circle
#> 535  7.282047136  8.45869403 Circle
#> 536  6.080886893  5.61017858 Circle
#> 537  6.343731548  8.39567023 Circle
#> 538 10.056986321  5.99524719 Circle
#> 539  6.832364409  8.66731784 Circle
#> 540  8.670830751  7.80269373 Circle
#> 541  9.803377648 10.03653576 Circle
#> 542  7.185229364  9.42659610 Circle
#> 543  7.013478837  8.50585980 Circle
#> 544  6.021613190 11.19158761 Circle
#> 545  8.951389110  9.36022510 Circle
#> 546  9.564645851  9.80568713 Circle
#> 547  8.933867212 10.40757060 Circle