ROSE Algorithm

A thin wrapper around ROSE::ROSE() that generates a synthetic balanced sample by enlarging the feature space of minority and majority class examples.

Usage

rose(
  df,
  var,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1
)

Arguments

df

A data.frame or tibble. Must have 1 factor variable with exactly 2 levels and remaining numeric variables.

var

Character, name of variable containing the 2-level factor variable.

over_ratio

A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows as the majority level. See vignette("ratio", package = "themis") for more details.

minority_prop

A numeric value between 0 and 1 for the proportion of synthetic observations from the minority class. Defaults to 0.5, which generates an equal split of minority and majority synthetic observations. This parameter controls the class balance within the synthetic data, while over_ratio controls the total size of the synthetic data.

minority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the minority class. Defaults to 1.

majority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the majority class. Defaults to 1.

Value

A data.frame or tibble, depending on type of df.

Details

The factor variable used to balance around must only have 2 levels.

The ROSE algorithm works by selecting an observation belonging to class k and generating new examples in its neighborhood, which is determined by a smoothing matrix H_k. Smaller values of minority_smoothness and majority_smoothness shrink the entries of H_k, producing tighter neighborhoods. This is a cautious choice when there is a concern that excessively large neighborhoods could blur the boundaries between classes.

This function is a thin wrapper around ROSE::ROSE(). For full details on the underlying implementation, see that function's documentation.

References

Lunardon, N., Menardi, G., and Torelli, N. (2014). ROSE: a Package for Binary Imbalanced Learning. R Journal, 6:82–92.

Menardi, G. and Torelli, N. (2014). Training and assessing classification rules with imbalanced data. Data Mining and Knowledge Discovery, 28:92–122.

See also

Examples

rose(circle_example[, c("x", "y", "class")], var = "class")
#>               x           y  class
#> 1    8.46815613  2.18937232   Rest
#> 2   11.66001604 11.74303416   Rest
#> 3    4.19304334 12.70785242   Rest
#> 4    7.27584415  1.19109554   Rest
#> 5   13.49940099  1.08127909   Rest
#> 6    7.74490640  3.37540370   Rest
#> 7    3.49449657 10.24334497   Rest
#> 8   14.80993440 14.26412968   Rest
#> 9    7.34442565  7.91953080   Rest
#> 10   8.90719366 12.25578047   Rest
#> 11   1.24068368  4.31212532   Rest
#> 12  11.17181481 15.50081716   Rest
#> 13  14.09686789 13.64274285   Rest
#> 14   2.37014983  0.46728572   Rest
#> 15  10.34536000  9.26807724   Rest
#> 16   7.31929335 17.27694452   Rest
#> 17   7.73416448 16.05222240   Rest
#> 18   8.13566432  5.95261920   Rest
#> 19  11.58170710  7.04742951   Rest
#> 20   5.75362968 13.14314803   Rest
#> 21   0.54077839  7.64180186   Rest
#> 22  13.24394946  4.52458410   Rest
#> 23  15.89231666  7.23604648   Rest
#> 24  11.63513180  8.69188244   Rest
#> 25  15.12142290  8.85029424   Rest
#> 26   6.19468589  5.81405540   Rest
#> 27   7.04356850  8.76482476   Rest
#> 28   7.07370538  1.15536906   Rest
#> 29   8.12583875 12.80272146   Rest
#> 30   2.12067916 15.68467688   Rest
#> 31   3.99190583  5.44270402   Rest
#> 32   1.66127758  3.00007894   Rest
#> 33  14.09669584  4.21760505   Rest
#> 34   2.67754953  0.75405845   Rest
#> 35  13.07260828 11.86007064   Rest
#> 36   2.01366049  0.18861942   Rest
#> 37   4.13604573 12.08975904   Rest
#> 38   3.48039958  7.92882698   Rest
#> 39  12.33380200 14.15749887   Rest
#> 40   3.21561851 14.54547454   Rest
#> 41  10.69414916  9.91810864   Rest
#> 42  11.23974901  8.25956835   Rest
#> 43   8.39291745 13.54608229   Rest
#> 44   7.09118715 13.02426742   Rest
#> 45  15.01378614  8.99140952   Rest
#> 46   2.76025342  3.94376014   Rest
#> 47   5.23287792 10.74850868   Rest
#> 48   6.65754406 13.08349052   Rest
#> 49  14.00551194 11.77866612   Rest
#> 50   2.99562633 14.85481895   Rest
#> 51   1.96829770  1.54205754   Rest
#> 52   3.39074839  1.30818417   Rest
#> 53  12.40776514  7.06445106   Rest
#> 54   0.19782620  4.58783074   Rest
#> 55   2.55662898  8.04011846   Rest
#> 56  -1.20085127  2.14202616   Rest
#> 57  15.92858197  4.21590218   Rest
#> 58   6.59946866  1.57452149   Rest
#> 59   7.04078468  3.43410318   Rest
#> 60  11.97707200 14.80023090   Rest
#> 61   7.16892860  4.86161479   Rest
#> 62   6.59398326  7.89081983   Rest
#> 63   4.59453452  8.86844370   Rest
#> 64  12.25580598  4.09353274   Rest
#> 65   8.31476979 14.90055780   Rest
#> 66   3.57171973  4.90081693   Rest
#> 67   7.94731637 10.00970636   Rest
#> 68   0.23206456  9.61400618   Rest
#> 69   4.72664227 11.53155124   Rest
#> 70  10.37230559  9.64299052   Rest
#> 71   0.79005688 10.09011347   Rest
#> 72  10.81683301  8.57891590   Rest
#> 73  10.92294486  2.39927091   Rest
#> 74   8.01398408 15.40068927   Rest
#> 75  16.90719012  9.25427438   Rest
#> 76  16.10959646 15.11261077   Rest
#> 77   9.04837469 14.51343018   Rest
#> 78  11.32893016 14.44023680   Rest
#> 79  18.07680151 10.88370632   Rest
#> 80   4.84766725 10.45940895   Rest
#> 81  11.51598515  9.08689104   Rest
#> 82   6.75235039  0.03107838   Rest
#> 83  11.49083788  3.24461126   Rest
#> 84   4.06711684  6.57544863   Rest
#> 85   2.45447069 13.18077240   Rest
#> 86  11.11146597  1.56607331   Rest
#> 87  10.91892578 10.42354311   Rest
#> 88  10.46235723  1.67788965   Rest
#> 89   6.26336510  4.90444679   Rest
#> 90   9.83965847 10.64923913   Rest
#> 91   4.91854549 13.26682672   Rest
#> 92   4.75240253 10.31883089   Rest
#> 93   2.42188346  4.91779894   Rest
#> 94   8.92189632  3.85043817   Rest
#> 95  -0.69850352  8.50075638   Rest
#> 96   2.62693928  9.91623051   Rest
#> 97  11.64755218 11.69864306   Rest
#> 98   6.30932995  7.92436075   Rest
#> 99   5.47829360 12.06354801   Rest
#> 100 13.88872397 10.51745127   Rest
#> 101  2.73825115 12.31802229   Rest
#> 102 14.22416614 -0.71636773   Rest
#> 103 13.81234295  7.18813838   Rest
#> 104 -0.02825943  7.23502752   Rest
#> 105  6.18103407 15.05681837   Rest
#> 106 11.62094031 12.34474756   Rest
#> 107  4.51721736  1.42188172   Rest
#> 108  6.76865280  1.95105918   Rest
#> 109 11.52350734  1.61987940   Rest
#> 110  2.51971599 11.20271357   Rest
#> 111 11.50252328  9.10430318   Rest
#> 112  8.19656123 13.64423185   Rest
#> 113  9.74888592 12.30583164   Rest
#> 114 14.87123379  0.92793039   Rest
#> 115  9.44177340 10.45871505   Rest
#> 116  2.40924200 13.85336034   Rest
#> 117 13.67070639 12.09790991   Rest
#> 118 12.12086610  2.84136647   Rest
#> 119  6.67658113 13.23574169   Rest
#> 120  6.32352384 11.84232229   Rest
#> 121 11.44436748 11.81584894   Rest
#> 122  5.56351556 12.82434953   Rest
#> 123  7.03620395 11.90906683   Rest
#> 124 10.41422352 11.64935497   Rest
#> 125  7.59011439 10.79360916   Rest
#> 126  0.25881007 14.71373601   Rest
#> 127  3.40437765  3.21021005   Rest
#> 128  2.22998129  8.24567908   Rest
#> 129  6.99976242  6.58262104   Rest
#> 130 11.34826790  2.87288648   Rest
#> 131 13.64860264 16.55603491   Rest
#> 132  0.55338907  1.70174346   Rest
#> 133 15.67391155 13.20249554   Rest
#> 134  4.95339383  0.21742435   Rest
#> 135  1.23338195  4.59690048   Rest
#> 136  2.94081281  1.56761501   Rest
#> 137  4.39821650  6.61806713   Rest
#> 138  6.05598633 14.78245079   Rest
#> 139 13.22056727  5.30720084   Rest
#> 140 12.95858534  1.17539809   Rest
#> 141  8.57710144  2.56076530   Rest
#> 142  6.33730406  4.21607808   Rest
#> 143  2.15658035  5.61609507   Rest
#> 144  7.77321461 16.97644116   Rest
#> 145 14.38696562 10.74010231   Rest
#> 146  1.12129863 16.31064672   Rest
#> 147  4.09753746 12.04257832   Rest
#> 148  7.93957017 11.72818799   Rest
#> 149 10.84298435  7.83046831   Rest
#> 150  7.43861858 12.22504998   Rest
#> 151 14.50604808 11.61166065   Rest
#> 152 10.16935793  2.94910533   Rest
#> 153 13.48992862  9.83642681   Rest
#> 154  4.58802722  8.13665434   Rest
#> 155  8.69464558  1.46892717   Rest
#> 156  7.12515576  3.12522094   Rest
#> 157  8.91097248 12.69568557   Rest
#> 158 11.13100528 10.68805106   Rest
#> 159 12.90577040  7.90989136   Rest
#> 160  7.31496557  5.78593990   Rest
#> 161  3.03971217 10.81873987   Rest
#> 162  9.50909800 14.90836980   Rest
#> 163  9.48305919 13.08794963   Rest
#> 164 10.29042776  7.65340275   Rest
#> 165 16.43901842 13.56175812   Rest
#> 166 11.30291653 10.47046809   Rest
#> 167 10.58907867 15.01592925   Rest
#> 168  7.76269715 11.55767523   Rest
#> 169 13.95804708  8.67428434   Rest
#> 170  6.90046083  9.48214368   Rest
#> 171  3.92622330 11.74381932   Rest
#> 172 14.71040723  0.92789338   Rest
#> 173 15.85611913  1.56586183   Rest
#> 174  4.33335143 13.06508800   Rest
#> 175  6.38566591 15.37024983   Rest
#> 176 15.79518707  9.59863552   Rest
#> 177 13.47088810  8.86798384   Rest
#> 178 11.71913667  7.32469874   Rest
#> 179 17.11414512 15.15037627   Rest
#> 180 11.98759252  1.93269975   Rest
#> 181  5.40828192  4.99466173   Rest
#> 182  5.70094538 11.58573392   Rest
#> 183 13.42532098  2.26982076   Rest
#> 184  2.46379806  2.02144190   Rest
#> 185 12.85435277  2.22500089   Rest
#> 186  4.93910161  9.75838923   Rest
#> 187  4.82404372  9.68192318   Rest
#> 188  9.81193157  9.62475047   Rest
#> 189  7.65394492  1.56680461   Rest
#> 190  5.20578124  3.03383866   Rest
#> 191 10.46807079  1.15712911   Rest
#> 192 14.23342703  7.23837628   Rest
#> 193  5.94491701  0.53050473   Rest
#> 194  5.88742903 10.67322364   Rest
#> 195 10.14572700 12.15240459   Rest
#> 196 12.23941413  6.74556966   Rest
#> 197 -0.75187581 12.64739511   Rest
#> 198  6.58622893 -0.49682113   Rest
#> 199 11.11506252  2.62881626   Rest
#> 200 15.44235434 12.26732166   Rest
#> 201  4.16180158  5.19839440   Rest
#> 202 14.36659727 12.52118695   Rest
#> 203 13.15170972  5.40232802   Rest
#> 204  1.12975394  8.39502646   Rest
#> 205  5.48974819 11.03817407   Rest
#> 206  3.98836467  2.05952687   Rest
#> 207  4.89922382 11.26200021   Rest
#> 208  8.66102330 12.87241823   Rest
#> 209 11.27489210 10.15704715   Rest
#> 210  3.86143128 16.11402390   Rest
#> 211  4.78885381  4.15061860   Rest
#> 212  4.78735432  2.67962180   Rest
#> 213 10.92568122 11.62610422   Rest
#> 214  1.12719353  4.33369228   Rest
#> 215  4.80809989 13.17343789   Rest
#> 216  5.67547733  2.85862604   Rest
#> 217  3.88233394  7.00381375   Rest
#> 218  5.43577550 13.69578321   Rest
#> 219 12.19316136 11.14175115   Rest
#> 220  6.18170270  7.59622360   Rest
#> 221  1.93695496  7.95800770   Rest
#> 222  8.00193853 10.71930897   Rest
#> 223  1.27883803  3.01831342   Rest
#> 224  3.13901072 12.94395201   Rest
#> 225  5.54886486 14.34287270   Rest
#> 226  6.04477883 13.79149723   Rest
#> 227 11.61465248 10.08799889   Rest
#> 228 12.53843368 12.90123155   Rest
#> 229  2.20133602  1.17406649   Rest
#> 230 10.92469919  9.07745640   Rest
#> 231  1.88836057  4.32741407   Rest
#> 232  3.44053475  5.64207244   Rest
#> 233  3.62153981  5.32304636   Rest
#> 234 13.34704084  4.55169037   Rest
#> 235  0.91409672  9.88807016   Rest
#> 236 10.51976489 11.32720252   Rest
#> 237  8.82799922 12.46614142   Rest
#> 238 11.12384529  2.49250058   Rest
#> 239  5.32388811 11.52898781   Rest
#> 240 13.74657950 10.24347311   Rest
#> 241  4.87910544 12.17208101   Rest
#> 242 10.07502133  7.29917888   Rest
#> 243 10.83131440  3.32819187   Rest
#> 244  3.52645897 13.39894544   Rest
#> 245  8.88353486 -0.75681529   Rest
#> 246 17.12756992  8.97262278   Rest
#> 247  1.56510624 10.23245706   Rest
#> 248 12.16372282  7.85837871   Rest
#> 249 11.22931967  2.65745736   Rest
#> 250  5.72245644 13.30906398   Rest
#> 251  7.31662414  0.59928815   Rest
#> 252  3.65718878  5.21739620   Rest
#> 253 13.19131880  3.84368405   Rest
#> 254 10.74286112  1.60570952   Rest
#> 255  4.18334776  3.43056530   Rest
#> 256  4.98861600  8.49457218   Rest
#> 257  6.79315580 12.36608697   Rest
#> 258 15.67773808  2.38621163   Rest
#> 259  3.72154656 12.08368984   Rest
#> 260  3.89654126 -0.51093587   Rest
#> 261  3.31320159  7.49341165   Rest
#> 262  8.52575962  4.32845860   Rest
#> 263 13.57940357  6.65319561   Rest
#> 264 11.23852278  6.84351938   Rest
#> 265  4.65840065  9.62907854   Rest
#> 266  5.54126752  2.76411256   Rest
#> 267 11.98699877 12.34495957   Rest
#> 268  3.79195726  2.80362449   Rest
#> 269  7.71994867 -0.11479501   Rest
#> 270  4.23328756  5.92691840   Rest
#> 271  2.64604080  8.58812393   Rest
#> 272  7.04080295 11.54743442   Rest
#> 273 11.70022648  9.84728516   Rest
#> 274  1.17226322 15.55990602   Rest
#> 275  5.07804242  5.31252623   Rest
#> 276  2.58932575 12.59377258   Rest
#> 277 12.89344083 11.97312354   Rest
#> 278 11.53334533 15.27717764   Rest
#> 279  8.45950996 13.92959549   Rest
#> 280  8.33473545 12.52406565   Rest
#> 281  1.59868567 -1.19985109   Rest
#> 282  4.63727115  5.88067056   Rest
#> 283  5.22851014 15.56023816   Rest
#> 284  8.54762780  9.76938839   Rest
#> 285  2.75064874  3.43078433   Rest
#> 286  1.41133812  8.75835699   Rest
#> 287 10.97517278 11.01706449   Rest
#> 288 13.70593025 11.57993797   Rest
#> 289  2.38737854  7.31326227   Rest
#> 290 13.76314015 12.83764682   Rest
#> 291  6.45842179 13.94788828   Rest
#> 292  4.81946152  3.89723465   Rest
#> 293  4.65289350  8.27150398   Rest
#> 294 15.75774520 12.61288139   Rest
#> 295  2.14268059  4.73765239   Rest
#> 296  3.50858406  9.42637489   Rest
#> 297 10.44999831  4.84178649   Rest
#> 298 15.03888818  9.43620590   Rest
#> 299 -0.83074055 13.54439875   Rest
#> 300  1.58358030 15.41475462   Rest
#> 301  2.11988185  3.47611806   Rest
#> 302  3.51592612 12.10977349   Rest
#> 303  1.87506732  5.87779893   Rest
#> 304  9.01203814  2.22327117   Rest
#> 305  1.40481035  5.92699969   Rest
#> 306  0.10563342  1.88037234   Rest
#> 307 10.81435208 13.88848421   Rest
#> 308 10.51443834 14.68884761   Rest
#> 309  6.67696778  1.23745084   Rest
#> 310  8.57766207 14.72945722   Rest
#> 311  3.58880543  5.08624603   Rest
#> 312  2.90353351  8.50188409   Rest
#> 313  4.39364235 10.98382375   Rest
#> 314  9.74638279 12.05464237   Rest
#> 315  1.13705874  8.91816654   Rest
#> 316 12.20103534 14.40617366   Rest
#> 317  0.68686642 12.36442598   Rest
#> 318  4.83084258 14.22961665   Rest
#> 319  3.78491491 11.75826392   Rest
#> 320 15.14594445 13.64838875   Rest
#> 321  2.23396583  7.80351573   Rest
#> 322  9.45405591  3.34760277   Rest
#> 323  7.72341587  3.60183268   Rest
#> 324 11.12608035 12.45774044   Rest
#> 325  5.21551545  8.91336876 Circle
#> 326  5.96304343  9.72630684 Circle
#> 327  7.91867343 10.21493363 Circle
#> 328  8.40447117 10.31796848 Circle
#> 329  5.01503606  7.76311448 Circle
#> 330  7.27534493  9.91283710 Circle
#> 331 10.10557454  6.31786068 Circle
#> 332  7.30618086  5.32639737 Circle
#> 333  7.89840743 11.03146922 Circle
#> 334  8.47346728  6.16576118 Circle
#> 335  8.58735730  9.98921140 Circle
#> 336  7.05915295  8.06694565 Circle
#> 337  8.00331491  8.26411933 Circle
#> 338  5.91772968  8.59228052 Circle
#> 339  7.22931167  8.80214444 Circle
#> 340  6.23397314  7.29473173 Circle
#> 341  8.67471539  9.22417538 Circle
#> 342  7.23710194 10.38177630 Circle
#> 343 10.26009877  4.76171349 Circle
#> 344  5.37059158  7.66912485 Circle
#> 345  9.64330429 10.49359325 Circle
#> 346  8.17394068  6.06637963 Circle
#> 347 10.57874139  7.21960538 Circle
#> 348  9.09422412  9.31578032 Circle
#> 349  6.31200379  8.14703635 Circle
#> 350  9.64646426  9.90586760 Circle
#> 351  5.69955597  6.29187693 Circle
#> 352  8.58969443  6.08256452 Circle
#> 353  4.63418605  6.62509340 Circle
#> 354  6.59316417  8.53376420 Circle
#> 355  4.87560349  7.89495343 Circle
#> 356  7.68712951  8.38967335 Circle
#> 357  7.39428586 10.55936780 Circle
#> 358  9.69398927  4.24577693 Circle
#> 359  8.05004942  9.35252386 Circle
#> 360  8.20088589 10.39571018 Circle
#> 361 10.10290393  9.54257171 Circle
#> 362  8.34620942 10.47850743 Circle
#> 363  8.35284789  8.70090182 Circle
#> 364  9.27830421 10.11185955 Circle
#> 365 10.22782572  7.95558465 Circle
#> 366  7.04904810  7.88987638 Circle
#> 367  5.19316310  7.90275424 Circle
#> 368  6.93286582 10.56209836 Circle
#> 369  6.85125657  8.14891515 Circle
#> 370  8.43060909  8.33825921 Circle
#> 371  5.32310356  8.59483464 Circle
#> 372  4.51731400  8.83535283 Circle
#> 373  8.54056793  8.54731296 Circle
#> 374  4.09958269  7.07753318 Circle
#> 375  9.52040454  6.15032700 Circle
#> 376  6.45608869  5.70832567 Circle
#> 377  5.83245575 10.50376651 Circle
#> 378  8.19192713  6.99764442 Circle
#> 379  3.37489204  9.60919871 Circle
#> 380  6.77867651  6.72667890 Circle
#> 381  9.21956001  8.79054289 Circle
#> 382  9.10306333 10.21162469 Circle
#> 383  8.94860987  8.40703495 Circle
#> 384  5.86602514  6.25957027 Circle
#> 385  9.08986371  6.81996991 Circle
#> 386  8.23270326  5.55620769 Circle
#> 387 10.25010731  6.31179109 Circle
#> 388  7.64299902  9.33511075 Circle
#> 389  9.54107354  6.41886844 Circle
#> 390 10.70186450  8.44351703 Circle
#> 391  6.57683568  8.59372804 Circle
#> 392  9.09347200 10.55830872 Circle
#> 393  9.07263617  6.72518220 Circle
#> 394  9.80523260  4.82531086 Circle
#> 395  8.78221722  8.99862553 Circle
#> 396  6.41263353  7.11458611 Circle
#> 397  8.16422705  6.15725321 Circle
#> 398  8.04058456  5.65500797 Circle
#> 399  5.16190661  6.18118687 Circle
#> 400  8.88225969  9.98779174 Circle
#> 401  9.18463993 10.32676391 Circle
#> 402  8.93326541  9.68000512 Circle
#> 403  6.26660932  8.02761584 Circle
#> 404  6.34760539  7.58302856 Circle
#> 405  7.31302114  9.85368955 Circle
#> 406  9.48627616  5.50546661 Circle
#> 407  5.48567989  8.84933480 Circle
#> 408  3.85297444  9.63217801 Circle
#> 409  5.07479114  6.73952444 Circle
#> 410  8.19749178  6.35447078 Circle
#> 411  5.38233911  9.19811906 Circle
#> 412  6.80369516  9.14904569 Circle
#> 413 10.17629427  8.45921803 Circle
#> 414  9.40085885  6.71566586 Circle
#> 415 11.31322173  8.84597391 Circle
#> 416  9.27673890  6.65406005 Circle
#> 417  6.18805641  4.22620022 Circle
#> 418  9.18933540  6.76239144 Circle
#> 419  9.02543821  7.50016609 Circle
#> 420 11.61576683  7.14725299 Circle
#> 421  7.10150337  9.20631071 Circle
#> 422  9.88207209  7.99191016 Circle
#> 423  6.70190831  6.06454556 Circle
#> 424  8.49829747  8.32456718 Circle
#> 425  8.60905584 10.41696155 Circle
#> 426  7.91436152  6.96595318 Circle
#> 427  7.50853357 11.58496456 Circle
#> 428  9.60992524  4.66753905 Circle
#> 429  5.22987754  8.36545701 Circle
#> 430  9.05619848  8.73223698 Circle
#> 431  6.00853373  8.90979845 Circle
#> 432  7.05717181  9.84640844 Circle
#> 433  6.48582533  5.67973646 Circle
#> 434  8.53687123  6.37146525 Circle
#> 435  7.42050804  7.29663491 Circle
#> 436  7.38566529  8.25137669 Circle
#> 437  8.60773349  8.69141114 Circle
#> 438  7.53847693 10.28701893 Circle
#> 439  9.11148197  9.80195278 Circle
#> 440  9.45413244  7.49358023 Circle
#> 441  5.03590720 10.83745044 Circle
#> 442  9.65708808  9.42791422 Circle
#> 443 11.80769924  5.17402160 Circle
#> 444  8.84733656  5.95056409 Circle
#> 445  8.83990894 10.42701947 Circle
#> 446  9.13983232  7.95886415 Circle
#> 447  9.95587626 10.41822866 Circle
#> 448  6.81243236  5.51243355 Circle
#> 449  6.51018101  7.40238490 Circle
#> 450  7.95402209  8.24355441 Circle
#> 451  9.38638778  9.11182974 Circle
#> 452  9.67964035 10.81789564 Circle
#> 453  8.16445946  9.88937808 Circle
#> 454  5.32826796  7.82011782 Circle
#> 455  6.64057946  6.65713112 Circle
#> 456  6.95259055  8.07722806 Circle
#> 457  7.79212195  6.41781574 Circle
#> 458  6.18736860  8.37310957 Circle
#> 459  9.19573127  6.60734428 Circle
#> 460  7.03140489 10.26184553 Circle
#> 461  8.95251060  6.83621353 Circle
#> 462 10.73362610  8.58178459 Circle
#> 463  9.19174185  9.82080686 Circle
#> 464  7.96888922  7.56618205 Circle
#> 465  8.89820026  8.93541741 Circle
#> 466  8.47661472  6.93965379 Circle
#> 467  6.44041944  9.40945512 Circle
#> 468 10.69912762  6.11003821 Circle
#> 469 10.93351734  8.79824263 Circle
#> 470  6.74863848  6.04768042 Circle
#> 471  8.41254120  9.17247969 Circle
#> 472 10.80612617  6.59685905 Circle
#> 473  8.36127099  9.11899403 Circle
#> 474  9.04058650  4.78876342 Circle
#> 475 12.65792057  9.54303481 Circle
#> 476  8.30686044  7.52703827 Circle
#> 477  5.22749145  9.83153834 Circle
#> 478  9.33502286  6.96903803 Circle
#> 479  8.87442968  6.20826254 Circle
#> 480  7.38714668  5.48208955 Circle
#> 481  8.17507536  6.09257599 Circle
#> 482  8.15989473  6.43226136 Circle
#> 483  6.67474058 10.73661953 Circle
#> 484  9.09925967  8.60433935 Circle
#> 485  6.59529042  7.64271625 Circle
#> 486  9.75750513  8.62514570 Circle
#> 487  8.66834847  6.35388728 Circle
#> 488  9.40576356  6.96280062 Circle
#> 489  6.28605572  7.16892029 Circle
#> 490  9.00028143  6.11956527 Circle
#> 491  8.03747403  4.93809283 Circle
#> 492 10.91461962 10.29940929 Circle
#> 493  8.76495342  6.05719354 Circle
#> 494  7.55758521 10.15472474 Circle
#> 495  7.41118013 11.40398923 Circle
#> 496  4.76099267  7.14787896 Circle
#> 497  8.41016665  9.82094769 Circle
#> 498  9.45831002  7.85207339 Circle
#> 499  5.52151035 10.11231358 Circle
#> 500  7.88141088  9.06524843 Circle
#> 501  9.90998363  7.17883610 Circle
#> 502 11.17567366  8.49944140 Circle
#> 503  9.18042636  9.26804288 Circle
#> 504  8.58450823 10.68921996 Circle
#> 505  9.41277853  8.59565904 Circle
#> 506  7.44741950  7.90126503 Circle
#> 507 10.71755050  7.53210539 Circle
#> 508  6.88312964  5.79473286 Circle
#> 509  8.50488128 10.50952213 Circle
#> 510 10.95576903  9.17808762 Circle
#> 511  7.22862142  7.66378616 Circle
#> 512  4.82162537  6.21598659 Circle
#> 513  8.51112860  8.05840277 Circle
#> 514  7.73667517  7.71654736 Circle
#> 515  7.82358514 10.25650578 Circle
#> 516  8.45518443  5.79835589 Circle
#> 517  9.38574766  7.16042690 Circle
#> 518  6.92321166  9.46072915 Circle
#> 519 10.02858341  6.59357283 Circle
#> 520  9.44495302  6.89639041 Circle
#> 521 10.40376626  7.63522206 Circle
#> 522  9.63283828 10.14461676 Circle
#> 523  4.88709399  6.66851316 Circle
#> 524  6.27726637  7.35102711 Circle
#> 525  8.77234225  7.46372667 Circle
#> 526 10.11250621 10.39972479 Circle
#> 527  7.91962580  7.47764254 Circle
#> 528  8.31837023  9.87264072 Circle
#> 529  8.62046765 10.90052532 Circle
#> 530  9.42173053  6.06053075 Circle
#> 531 10.76886912  9.62811945 Circle
#> 532  7.99002245  6.55760195 Circle
#> 533  8.42251665  7.34097797 Circle
#> 534  8.85187844 10.96806186 Circle
#> 535  7.79116529  9.88222057 Circle
#> 536  6.54599529  7.28106945 Circle
#> 537  7.59083400  9.37926724 Circle
#> 538  6.37568482 11.03533714 Circle
#> 539  6.12112998 10.10633816 Circle
#> 540  7.56050618 10.28553657 Circle
#> 541  6.57401842  8.74772222 Circle
#> 542  5.67660470  6.71992776 Circle
#> 543  9.95038658  8.79218626 Circle
#> 544 10.40796653  9.38439669 Circle
#> 545  8.47973551 11.51511529 Circle
#> 546  9.88302968  7.76308137 Circle
#> 547  9.74476107  9.41948100 Circle
#> 548  7.93677173  6.83080130 Circle
#> 549  6.84557045  7.39225292 Circle
#> 550  5.78181919  9.21354405 Circle
#> 551  8.46710446  9.16076954 Circle
#> 552  7.86373647  7.91248090 Circle
#> 553 10.26344637  8.79122603 Circle
#> 554  6.99841207  9.31132698 Circle
#> 555  7.71443811  9.16218342 Circle
#> 556  8.06847941  9.03435589 Circle
#> 557 10.53638422  7.89840569 Circle
#> 558  8.15555167  9.65845017 Circle
#> 559  7.40109839  6.96286215 Circle
#> 560  7.91752638  9.69036504 Circle
#> 561  5.09209013  9.79204209 Circle
#> 562  7.72436492  6.35133557 Circle
#> 563  7.91653379  9.40363890 Circle
#> 564  7.84161111  7.75347112 Circle
#> 565  6.31838617  9.65171855 Circle
#> 566 11.37556099 10.68161460 Circle
#> 567  7.21294291  6.48589784 Circle
#> 568  8.61462803  8.68164262 Circle
#> 569  8.27315347  6.70884350 Circle
#> 570  9.70361233  9.95454005 Circle
#> 571  9.31947136  6.03219965 Circle
#> 572 10.45002677  8.64827352 Circle
#> 573 10.59065391 11.03706023 Circle
#> 574  5.43108348  7.20489816 Circle
#> 575  9.16129315  9.04655671 Circle
#> 576 10.29504512  7.19527458 Circle
#> 577  7.92788223  8.54933216 Circle
#> 578  5.69142122  8.51059883 Circle
#> 579  6.16372083  7.27788018 Circle
#> 580 11.55463204  9.07830388 Circle
#> 581  6.80716767  7.36287034 Circle
#> 582  9.45258164  9.42070500 Circle
#> 583  7.68452249 11.85079643 Circle
#> 584  7.21333326  9.08729018 Circle
#> 585  8.78244803  9.33844229 Circle
#> 586  6.53585605  7.18160263 Circle
#> 587  5.73951805  7.81467402 Circle
#> 588  7.02549849  5.00556497 Circle
#> 589  6.05655847  7.94608312 Circle
#> 590  7.26269033  5.94963207 Circle
#> 591  9.74698330  7.10333229 Circle
#> 592  4.77611387  7.78317837 Circle
#> 593  8.22443232  5.28292196 Circle
#> 594  9.00154403  8.79917302 Circle
#> 595  7.92531947  9.13521208 Circle
#> 596  4.51114843  8.06738435 Circle
#> 597  9.40741585  7.06841309 Circle
#> 598  7.09660644  9.86407213 Circle
#> 599 10.42739535  6.12366041 Circle
#> 600  4.26790607  5.61564547 Circle
#> 601  7.72366383  9.61264245 Circle
#> 602  9.29018340  5.36667954 Circle
#> 603 10.48266326  5.89160485 Circle
#> 604  7.31014071  8.83032007 Circle
#> 605  6.37013509  5.93067851 Circle
#> 606  9.25423958  5.47422420 Circle
#> 607  9.78153072  5.14409708 Circle
#> 608  8.82681113  6.36998963 Circle
#> 609  6.19690454 10.09123340 Circle
#> 610  8.51369174  5.71876066 Circle
#> 611 10.94654079  9.20334499 Circle
#> 612  8.83501388 11.35915378 Circle
#> 613 10.91840370  7.32039435 Circle
#> 614  7.55660398  9.76209405 Circle
#> 615  8.98996144  9.77094854 Circle
#> 616  8.30813613 10.60398578 Circle
#> 617  7.50883739  7.36372969 Circle
#> 618  7.67301200  8.36468596 Circle
#> 619  7.79841859  9.95662016 Circle
#> 620 10.38772413  5.13258452 Circle
#> 621  8.44905884  9.19267203 Circle
#> 622  9.46782670  8.25203017 Circle
#> 623  6.62020539  7.40252358 Circle
#> 624  8.09428507 10.03557935 Circle
#> 625  8.57351070  6.24131892 Circle
#> 626  9.22666186  6.68915161 Circle
#> 627  6.69714465  8.92370205 Circle
#> 628  5.10016793  8.54537878 Circle
#> 629  6.93466200 10.13029259 Circle
#> 630  8.14658222  9.58791501 Circle
#> 631  8.64317288  8.09623978 Circle
#> 632  7.19151758  9.84700722 Circle
#> 633  8.69699594  8.88937294 Circle
#> 634  6.66719875  7.76646440 Circle
#> 635  8.96652291  5.93084602 Circle
#> 636  7.43976319  4.67910657 Circle
#> 637  9.71717704  9.60582169 Circle
#> 638  6.85134122  6.43048784 Circle
#> 639  4.54909515  8.88839336 Circle
#> 640  8.82170515 10.47480351 Circle
#> 641  9.11096810  6.42741748 Circle
#> 642  7.18426381  4.64958842 Circle
#> 643  9.07664313  8.85958303 Circle
#> 644  9.86945133 10.69272158 Circle
#> 645  4.86826215  9.69351833 Circle
#> 646  3.77052070  7.56333161 Circle
#> 647  6.85466796  6.50902765 Circle
#> 648  7.05720384  8.83256745 Circle
#> 649  7.67017225  9.97290037 Circle
#> 650  7.46126808  5.59883913 Circle
#> 651  6.08409011  5.42046279 Circle
#> 652  8.98859707  8.08638002 Circle
#> 653 10.08096062  7.03866071 Circle
#> 654  7.47332351 10.92214964 Circle
#> 655  9.82340962  9.86576702 Circle
#> 656  7.66786032  9.72429239 Circle
#> 657  6.76076191  7.83166432 Circle
#> 658  9.08452877  8.30725409 Circle
#> 659  9.80268400 11.16187190 Circle
#> 660  7.24318880  9.10420460 Circle
#> 661  6.88443798  6.13411892 Circle
#> 662  8.67480426  9.30435673 Circle
#> 663  7.72455039  6.55012898 Circle
#> 664  7.50796257  9.06326469 Circle
#> 665  4.24384773  7.15875221 Circle
#> 666  6.74300984  9.89873333 Circle
#> 667  8.22912329  9.07799344 Circle
#> 668  7.73082246  6.71154927 Circle
#> 669  8.91891703  4.88355247 Circle
#> 670  9.54169853  6.96897212 Circle
#> 671  6.86681206 10.84278376 Circle
#> 672  9.21179593 11.30658404 Circle
#> 673  9.48809985  6.30106524 Circle
#> 674  6.43436982  8.58126125 Circle
#> 675  9.09340988  8.87140870 Circle
#> 676  8.19452093 10.19104608 Circle
#> 677  5.37226873  8.50498478 Circle
#> 678  9.09638959  7.40300474 Circle
#> 679  7.16970163  7.41877000 Circle
#> 680 10.10566997  9.46987798 Circle
#> 681 10.08676699  8.53237265 Circle
#> 682 10.54024911  7.92036458 Circle
#> 683  8.29745745  5.86719643 Circle
#> 684  6.25124012  9.25188764 Circle

rose(circle_example[, c("x", "y", "class")], var = "class", over_ratio = 0.8)
#>               x          y  class
#> 1   13.11995692  6.7821610   Rest
#> 2    6.39717318 11.4583920   Rest
#> 3   14.44748430 10.8189463   Rest
#> 4    3.29134017  9.7461230   Rest
#> 5    7.43764676 15.3165631   Rest
#> 6    5.77596437 11.6193679   Rest
#> 7   11.76380840  9.3977287   Rest
#> 8    5.53632905  9.3570367   Rest
#> 9   11.05191034 12.3852308   Rest
#> 10   3.24913015  3.1006740   Rest
#> 11   0.33065981 10.4418057   Rest
#> 12   8.68851919  7.5729950   Rest
#> 13   4.63472308 12.8900633   Rest
#> 14   1.33923469 11.2043553   Rest
#> 15  11.07552798  5.5794371   Rest
#> 16  11.54204981 10.3164778   Rest
#> 17   9.89124202 14.0453389   Rest
#> 18   1.47099410  7.9912937   Rest
#> 19  13.84399568  3.3510006   Rest
#> 20  12.31206672 11.0537062   Rest
#> 21  15.11607246 10.5949232   Rest
#> 22  12.02429657  5.5230800   Rest
#> 23  13.47582548  5.4864654   Rest
#> 24   8.48043627 17.4433441   Rest
#> 25  12.93357401 14.8823846   Rest
#> 26  14.56002117 12.5467838   Rest
#> 27   3.89305263 11.7389816   Rest
#> 28  17.72622255  5.2654478   Rest
#> 29  16.44645144  9.8929540   Rest
#> 30   8.87156547 12.6967619   Rest
#> 31  10.58308918 14.0636812   Rest
#> 32   6.67856191 15.8356832   Rest
#> 33   4.55737062 14.2210923   Rest
#> 34   7.24411768  4.4197485   Rest
#> 35  14.26726335  7.6257260   Rest
#> 36   7.21051793  9.4317407   Rest
#> 37   8.14872500 15.6526327   Rest
#> 38  10.08472259 16.3704356   Rest
#> 39   2.49465154 11.6725608   Rest
#> 40   9.86081166 -1.1358638   Rest
#> 41  13.82405049  2.0389051   Rest
#> 42  12.50588129 13.6956915   Rest
#> 43   7.68169957  4.3021144   Rest
#> 44   2.61095006  9.8054442   Rest
#> 45  10.37647532 13.4813058   Rest
#> 46  16.24384939  1.4781965   Rest
#> 47  15.43771771  5.1366797   Rest
#> 48   3.38492787 10.6710589   Rest
#> 49   1.90220211  6.7747430   Rest
#> 50  14.51559922 10.7866650   Rest
#> 51  10.64572650 10.4881610   Rest
#> 52  11.38580907  3.6534007   Rest
#> 53   6.68366169 11.4692882   Rest
#> 54   7.53955633 14.3186994   Rest
#> 55   3.57252391  9.9449706   Rest
#> 56   3.91434830 10.6059666   Rest
#> 57   2.29201324  8.0962911   Rest
#> 58  13.97128889 11.1073538   Rest
#> 59  15.80626274 -1.3389862   Rest
#> 60  16.08751342 13.3423704   Rest
#> 61  11.37122617  9.4304505   Rest
#> 62   7.05286377 12.0927738   Rest
#> 63   9.44986471  4.2053943   Rest
#> 64  12.24734842  9.1324818   Rest
#> 65  10.76320906 12.1597446   Rest
#> 66   0.33223290  0.9393747   Rest
#> 67   3.88931277 -1.0246616   Rest
#> 68   3.52238409  4.4885870   Rest
#> 69   6.56659401 17.4847152   Rest
#> 70  10.21478572  1.4562349   Rest
#> 71   5.39207185  4.8173437   Rest
#> 72   7.07110084 14.8031069   Rest
#> 73  15.36255849  1.0815824   Rest
#> 74   4.63299850 12.7108919   Rest
#> 75   6.64949783 13.8418366   Rest
#> 76  17.44213309  1.2452662   Rest
#> 77   5.09597712  2.2214087   Rest
#> 78  10.58327825 13.4904879   Rest
#> 79   7.32522242  4.8019040   Rest
#> 80   4.76664354 12.7030696   Rest
#> 81  13.79786509 14.8348636   Rest
#> 82  16.70701770  8.3243251   Rest
#> 83   6.29736233  4.2183728   Rest
#> 84   8.22666715  1.8142003   Rest
#> 85   2.37078687  8.3363008   Rest
#> 86   2.93284711  7.6930838   Rest
#> 87   9.90735084 -0.1878908   Rest
#> 88   7.04459474  1.6982210   Rest
#> 89  12.48755053 13.2770639   Rest
#> 90   1.66199966 -0.7630248   Rest
#> 91  11.46103939 15.9836570   Rest
#> 92  10.73335077 15.1249951   Rest
#> 93   2.18813066  9.6971943   Rest
#> 94   1.68399164  6.3597096   Rest
#> 95   7.91601393 14.9546803   Rest
#> 96  16.38459206  6.7750126   Rest
#> 97  16.04018406 11.5852787   Rest
#> 98   3.70227922 -0.1468308   Rest
#> 99   3.97449633  5.4042194   Rest
#> 100  8.71244830  4.5533550   Rest
#> 101  4.66219319  9.9844102   Rest
#> 102  7.24034569  1.1985848   Rest
#> 103 12.64586268 11.8109948   Rest
#> 104 10.28693980 11.6919735   Rest
#> 105  1.88805086  4.2925717   Rest
#> 106  1.16299983  8.1012416   Rest
#> 107  2.01649609  0.4548762   Rest
#> 108  9.19693613 13.4629804   Rest
#> 109  7.24967325 13.3408544   Rest
#> 110  8.75393443  5.7155919   Rest
#> 111 14.56169077  6.4556850   Rest
#> 112  3.03965667  0.1681632   Rest
#> 113  8.87911466  6.7112335   Rest
#> 114  5.45479989  4.5614671   Rest
#> 115  6.85342589  3.4935197   Rest
#> 116 13.86042450  2.1927141   Rest
#> 117 14.96361787  2.9686379   Rest
#> 118  2.93817554 12.5766125   Rest
#> 119  4.02361405 13.6994866   Rest
#> 120 14.45427647 15.0239390   Rest
#> 121  5.85570050 14.5596250   Rest
#> 122 14.05323759 13.3114146   Rest
#> 123  9.09365584  3.4475558   Rest
#> 124  8.29175858  3.3990532   Rest
#> 125  0.08991853  6.1775235   Rest
#> 126  4.02794362  9.9236433   Rest
#> 127  6.54263889 14.0944458   Rest
#> 128  2.19349272 12.7925156   Rest
#> 129  2.14148787 10.0568545   Rest
#> 130 12.97188062  3.7270717   Rest
#> 131  9.10557919  2.5182757   Rest
#> 132  9.11659282  1.8671199   Rest
#> 133 13.35468555  8.2096344   Rest
#> 134 13.25468818  3.2251793   Rest
#> 135  2.70902985 13.7722844   Rest
#> 136  4.88160766  8.6534532   Rest
#> 137  2.79232504 17.1773880   Rest
#> 138  6.46427180 10.9299604   Rest
#> 139  3.24669505  4.6386952   Rest
#> 140  4.72339450 11.6850351   Rest
#> 141 13.87191894  5.7473703   Rest
#> 142 12.93524941 14.0966708   Rest
#> 143 11.27453966 15.7163470   Rest
#> 144 14.46690638 15.0295214   Rest
#> 145  9.49931268 12.3313045   Rest
#> 146  3.60522109  4.7707898   Rest
#> 147 12.93854492 13.9202875   Rest
#> 148  4.85541611  5.6170188   Rest
#> 149 14.37376869  7.8051621   Rest
#> 150  3.87009162  1.0362962   Rest
#> 151  9.21017584 12.8177071   Rest
#> 152  0.95482218  9.8446254   Rest
#> 153  2.01275391  7.4155494   Rest
#> 154 10.20811595 11.3094954   Rest
#> 155  4.01372667 13.4140149   Rest
#> 156  4.51774424  5.7127424   Rest
#> 157 16.50724542  0.6244713   Rest
#> 158  3.71497818  0.5649499   Rest
#> 159 10.84748625  2.9171343   Rest
#> 160  3.98567187  7.1771730   Rest
#> 161  4.82422077  4.6973842   Rest
#> 162 12.25573938  2.4772014   Rest
#> 163 14.23228901 10.5902919   Rest
#> 164 12.39118288  0.4351951   Rest
#> 165 12.24282685  5.7867678   Rest
#> 166  5.68965576  8.0267012   Rest
#> 167 14.72121729  7.7992546   Rest
#> 168 12.41947232 16.4510667   Rest
#> 169  6.58339034  7.9333847   Rest
#> 170  9.80444285 12.1468485   Rest
#> 171  3.23989185  6.6752560   Rest
#> 172  9.03835774 11.4505394   Rest
#> 173 10.81703561 13.3967961   Rest
#> 174  0.13696353  9.1317981   Rest
#> 175  6.57949684  2.9165543   Rest
#> 176  8.46176713 13.7648041   Rest
#> 177  8.76903066 14.0623916   Rest
#> 178 11.93350870 10.9522277   Rest
#> 179  9.39125922  1.3868509   Rest
#> 180  0.91235546  4.8879012   Rest
#> 181  9.52876273  2.3311945   Rest
#> 182 10.07207013 12.9842577   Rest
#> 183  1.93691429 13.2809910   Rest
#> 184 14.26759171  7.7281949   Rest
#> 185  3.19226720  6.6953858   Rest
#> 186 12.90512063 17.1409094   Rest
#> 187 11.97636632  5.6548233   Rest
#> 188  3.90524839  1.0709252   Rest
#> 189 12.82398497 11.4112566   Rest
#> 190  7.26636086  4.9444718   Rest
#> 191  1.27743517  9.4544805   Rest
#> 192  3.16328198  1.9727061   Rest
#> 193  3.69866163  5.7028342   Rest
#> 194 12.64543307  9.3043129   Rest
#> 195 14.01467888 10.3846829   Rest
#> 196  2.90827298 16.2487940   Rest
#> 197 14.39443230  8.9534395   Rest
#> 198  3.36013600  8.6102288   Rest
#> 199  3.11020370  3.4796399   Rest
#> 200  2.73288345 10.2960955   Rest
#> 201  6.06857399 16.6910952   Rest
#> 202  5.67842754  3.8965036   Rest
#> 203  4.33907607  2.2302803   Rest
#> 204  4.97716275 11.4474977   Rest
#> 205  4.66061629 13.8523417   Rest
#> 206 14.13294048  3.3978773   Rest
#> 207 12.41495657  2.2546980   Rest
#> 208  4.73621226 15.8840274   Rest
#> 209  2.87059055 10.5036866   Rest
#> 210 11.47648887 12.5633548   Rest
#> 211 11.68872989  8.6388502   Rest
#> 212  5.42630684  5.3000439   Rest
#> 213  6.03846651 14.9819967   Rest
#> 214  4.95920102  1.2443255   Rest
#> 215 14.80609710 11.8497596   Rest
#> 216 16.06906247  5.4519256   Rest
#> 217 13.91984514 12.2577138   Rest
#> 218  9.02311092 13.0792075   Rest
#> 219 11.90452559 14.1482799   Rest
#> 220  6.83621499  6.1477619   Rest
#> 221 14.60961637  8.0852712   Rest
#> 222  2.42515537 11.9031563   Rest
#> 223  5.58493651 12.2261580   Rest
#> 224  2.67802155  9.9951553   Rest
#> 225  8.45469928 13.6010505   Rest
#> 226  1.99827310  6.0649206   Rest
#> 227 13.08443518  5.6708729   Rest
#> 228  4.27095915 18.9729795   Rest
#> 229 14.38254677  1.6437530   Rest
#> 230 11.12475273  4.5151383   Rest
#> 231  2.40296865 10.1229857   Rest
#> 232  7.37691816 12.2101213   Rest
#> 233  1.69758214  7.8767351   Rest
#> 234  8.17611337  3.8947524   Rest
#> 235 11.44979988  5.5932231   Rest
#> 236  5.77378672  4.2086376   Rest
#> 237  3.33436126  3.2237579   Rest
#> 238 14.47577448  8.2730241   Rest
#> 239  6.03096859  7.6153636   Rest
#> 240  1.75333801  4.2462311   Rest
#> 241  4.95394057  2.8885559   Rest
#> 242 10.95804181  7.5982360   Rest
#> 243  8.47893338 12.5199111   Rest
#> 244  2.34033749  0.8771414   Rest
#> 245  7.20157914  5.5734572   Rest
#> 246  9.29363778  3.5431945   Rest
#> 247  3.15560687  3.2705294   Rest
#> 248  6.56551974 13.4709752   Rest
#> 249  3.84954337 -0.7459063   Rest
#> 250 11.91325088 -0.2557097   Rest
#> 251  4.23764348 13.9969827   Rest
#> 252  5.51331657 12.1994676   Rest
#> 253  2.31716671  1.9613336   Rest
#> 254 12.39587340 14.0917397   Rest
#> 255  6.49909361  0.2169299   Rest
#> 256  6.59608071 12.6192586   Rest
#> 257  2.67867435 10.4097194   Rest
#> 258  3.94999095 11.4367929   Rest
#> 259  1.70412018  1.1322151   Rest
#> 260 15.80802245  6.2239669   Rest
#> 261 13.68927892  7.2390177   Rest
#> 262 13.58688951 14.2397686   Rest
#> 263  4.14682561 10.4833420   Rest
#> 264 15.95846355 14.0670426   Rest
#> 265  5.26476761  8.7206063   Rest
#> 266 10.32409297 11.1564921   Rest
#> 267 10.41775994  8.3776239   Rest
#> 268  8.31518736 10.5280531   Rest
#> 269  6.11244021 13.6038487   Rest
#> 270  3.97046857  3.4240112   Rest
#> 271  0.28697570 16.0233293   Rest
#> 272  4.26914606  9.9932660   Rest
#> 273  1.35268849  4.7082491   Rest
#> 274 12.69408220  5.1334875   Rest
#> 275  3.54092169 10.7517285   Rest
#> 276 10.77690886  4.3382728   Rest
#> 277 11.84886532 14.3085525   Rest
#> 278  4.86144321 13.5051076   Rest
#> 279  2.76887134  2.3221158   Rest
#> 280 14.44936018  8.1521760   Rest
#> 281  6.76304579 12.1902462   Rest
#> 282 12.99209543  7.5077403   Rest
#> 283  9.03254298  7.2691990 Circle
#> 284 10.04221860 10.0221556 Circle
#> 285  8.96165856  5.1725515 Circle
#> 286  7.37840963  9.1141514 Circle
#> 287 10.01105554  8.8741183 Circle
#> 288  9.47626024  6.4775017 Circle
#> 289 10.00542032  8.6001500 Circle
#> 290  9.23339535 10.7377777 Circle
#> 291  7.70027758  5.4955787 Circle
#> 292  4.36913058  8.0762293 Circle
#> 293  9.15828649  8.0943156 Circle
#> 294  8.03571454  8.6775117 Circle
#> 295 11.05366556  7.3744645 Circle
#> 296  4.83820966  5.2797836 Circle
#> 297  5.70435814  6.1535142 Circle
#> 298  6.44513749  9.7521112 Circle
#> 299  5.38980876  8.3342668 Circle
#> 300 10.25207695  6.8346622 Circle
#> 301  8.48501103  8.9178691 Circle
#> 302  9.73484732  6.7830220 Circle
#> 303  9.62928938  6.7871788 Circle
#> 304  6.63039920  9.3141170 Circle
#> 305  7.22721678  6.0070282 Circle
#> 306  5.89769358  9.2446589 Circle
#> 307  8.94720434  9.0685501 Circle
#> 308  9.65216254  9.8574585 Circle
#> 309  5.00685928  6.1983819 Circle
#> 310  7.75960331  7.5960247 Circle
#> 311  5.90892460  5.8207585 Circle
#> 312  4.16028333  8.5373534 Circle
#> 313  8.08560408  8.4436326 Circle
#> 314  8.44446029  6.3683631 Circle
#> 315  6.26998158  8.7932471 Circle
#> 316  8.51627612  5.6220691 Circle
#> 317  8.44163990  7.3011953 Circle
#> 318 10.90624185  9.9449553 Circle
#> 319  7.60195931  8.9647537 Circle
#> 320  4.77204525  7.6342848 Circle
#> 321  7.47750693  9.9810109 Circle
#> 322  8.85526464  9.4657569 Circle
#> 323  9.90707951  7.3048346 Circle
#> 324  6.54718853 11.4907826 Circle
#> 325  8.64857321  5.9213095 Circle
#> 326  6.11040153  7.3027010 Circle
#> 327  4.74594757  7.7791536 Circle
#> 328  8.36654232  9.2335454 Circle
#> 329  6.38988821  8.0109305 Circle
#> 330 10.11944820  9.6460725 Circle
#> 331  4.90976699  7.7649259 Circle
#> 332  6.19149584  7.8222963 Circle
#> 333  7.21832912  9.0404844 Circle
#> 334  9.04859251  5.8557455 Circle
#> 335 10.90219476 10.2610649 Circle
#> 336  7.78336312  5.9328654 Circle
#> 337 10.40797310  8.7611620 Circle
#> 338  8.43017282  8.4061616 Circle
#> 339  7.45090590  9.5785912 Circle
#> 340  9.43277838  8.3365324 Circle
#> 341  3.83411372  7.7088746 Circle
#> 342  9.72999634 10.3542960 Circle
#> 343  9.53323396  9.5780579 Circle
#> 344  8.11901555  9.0733318 Circle
#> 345  7.01079073  7.2570864 Circle
#> 346 11.38794635  8.7860951 Circle
#> 347  6.02296509  7.2049251 Circle
#> 348  9.06957388  7.3033461 Circle
#> 349 10.04453747 10.4469445 Circle
#> 350  5.50399636  5.8557761 Circle
#> 351  6.67989478  9.8897062 Circle
#> 352  8.74007088  9.7351494 Circle
#> 353  8.72638172  9.8197589 Circle
#> 354  5.34289680  6.6723636 Circle
#> 355  8.67007219 10.4600820 Circle
#> 356  5.18330887  8.9825186 Circle
#> 357  8.13983493  8.1346419 Circle
#> 358  5.02895198  4.6640870 Circle
#> 359  7.87071184  7.6105229 Circle
#> 360  8.89053352  9.9387335 Circle
#> 361  6.48011921 10.3924502 Circle
#> 362  7.11404859  9.2965365 Circle
#> 363  6.38882508  9.8929648 Circle
#> 364  4.16286851  7.8310411 Circle
#> 365  7.99109208  8.2547415 Circle
#> 366  9.67790925  6.4267384 Circle
#> 367  7.57356600  4.4415164 Circle
#> 368  9.05960553  6.6842444 Circle
#> 369  7.00556033  7.2115152 Circle
#> 370  6.65932320  6.2871666 Circle
#> 371  8.27279444  9.4688553 Circle
#> 372  8.31803776  5.2856326 Circle
#> 373  9.88975150  9.4083357 Circle
#> 374  7.10822577  9.8146192 Circle
#> 375  7.59659461  9.7894926 Circle
#> 376  7.95529228 10.7551174 Circle
#> 377  9.94376416  8.4019352 Circle
#> 378  9.66504599  7.2238337 Circle
#> 379  6.38368379  8.6571920 Circle
#> 380  8.92670131  6.4656407 Circle
#> 381  4.17896185  7.8225649 Circle
#> 382  7.25437368  7.9116294 Circle
#> 383  8.34343612  5.9728065 Circle
#> 384  8.17145600 10.0571570 Circle
#> 385  5.68817934  8.1759225 Circle
#> 386 10.47646547 10.3710354 Circle
#> 387  8.19128577  7.4153255 Circle
#> 388  9.03908214  8.9109797 Circle
#> 389  7.34576184 10.9675992 Circle
#> 390  8.90630362  8.3906147 Circle
#> 391 10.18527834  8.2045702 Circle
#> 392  8.73754546  7.7101878 Circle
#> 393  6.33683471  8.6090896 Circle
#> 394  7.62927208  4.0934333 Circle
#> 395  8.02677885 10.5425401 Circle
#> 396  9.59946981  5.2702400 Circle
#> 397  8.60856907  9.5649016 Circle
#> 398  6.64708586  5.9905309 Circle
#> 399  5.57514137  9.5753845 Circle
#> 400  9.43026998 10.0256804 Circle
#> 401  8.06650451  7.4170062 Circle
#> 402  5.95917831  6.6371050 Circle
#> 403  7.73576507  8.1794781 Circle
#> 404  9.11946768  9.5192581 Circle
#> 405  7.88000185  8.3826334 Circle
#> 406 10.29917539  7.8311287 Circle
#> 407  7.54881403  9.8768173 Circle
#> 408  8.99002459  9.4164304 Circle
#> 409  7.87261343 10.4293812 Circle
#> 410  6.78476088  6.6793235 Circle
#> 411  9.60597029  8.9808033 Circle
#> 412  8.31801930  4.9649654 Circle
#> 413  8.37778614  6.0929318 Circle
#> 414  5.03236973  6.1572355 Circle
#> 415 10.93515369  8.3654722 Circle
#> 416  8.22274180  9.1842501 Circle
#> 417  9.11019269  4.8735013 Circle
#> 418  6.06137571  9.0455990 Circle
#> 419  7.41296163  6.0583924 Circle
#> 420  9.40538699  9.5464067 Circle
#> 421  6.36058140  8.5132289 Circle
#> 422 11.01825470  7.8600097 Circle
#> 423  5.48194792  8.5699191 Circle
#> 424  7.76085105  5.0385676 Circle
#> 425 10.51489063  7.4481066 Circle
#> 426  6.47054393  8.1875060 Circle
#> 427  7.40482653 10.5268401 Circle
#> 428  4.62496739  9.6924316 Circle
#> 429  6.99675671  9.7740966 Circle
#> 430  8.41984236  8.7271208 Circle
#> 431  9.79086910  9.4540512 Circle
#> 432  8.03998479  6.6093827 Circle
#> 433 10.64057967  8.4326557 Circle
#> 434  4.79781170 10.5583431 Circle
#> 435  9.17458044  7.5942594 Circle
#> 436  8.48535644  6.4832704 Circle
#> 437  7.87195273  9.9367387 Circle
#> 438  6.95116495  8.7987135 Circle
#> 439  7.44322758  8.2317537 Circle
#> 440  8.44862909  5.9704518 Circle
#> 441  7.91417505  6.4904292 Circle
#> 442  7.49033445  9.8883399 Circle
#> 443  7.62027996  5.2954315 Circle
#> 444  7.74166670 11.0365268 Circle
#> 445  7.74355214  9.6699746 Circle
#> 446  8.57498745  7.8403221 Circle
#> 447  6.72067693  7.5493280 Circle
#> 448  5.78335454  7.4356239 Circle
#> 449  8.01674576 10.6340247 Circle
#> 450  9.01266203  9.2090748 Circle
#> 451  6.09583315  8.3507124 Circle
#> 452  7.81996863 10.5603420 Circle
#> 453  7.33554980  9.8376732 Circle
#> 454  8.55966644  7.6978475 Circle
#> 455 10.06754809  6.1656306 Circle
#> 456  7.34700168  6.9388465 Circle
#> 457  9.71809378  7.7128242 Circle
#> 458 11.33371918  9.9464236 Circle
#> 459  7.48436412  9.2876598 Circle
#> 460  4.41669372 10.4448363 Circle
#> 461  4.87501132  8.3193481 Circle
#> 462  5.99194267  7.0114209 Circle
#> 463 10.35299534  9.4999849 Circle
#> 464  6.58990235  9.8475541 Circle
#> 465  8.54393848 10.5360281 Circle
#> 466  9.03939073  6.3137634 Circle
#> 467  9.44900704  9.3532885 Circle
#> 468  8.19395003  5.6094612 Circle
#> 469  9.31368994  8.1083904 Circle
#> 470  4.58235824  8.8972429 Circle
#> 471  9.12699246  6.2206885 Circle
#> 472  6.86421619 10.0923707 Circle
#> 473  6.88450467  6.9772768 Circle
#> 474  7.33457623  9.5306718 Circle
#> 475  7.44452284  7.0345542 Circle
#> 476  9.51107553 10.6777239 Circle
#> 477  7.87036870  6.7640131 Circle
#> 478  6.26865183  6.4039597 Circle
#> 479  9.74572909 10.0992715 Circle
#> 480  8.22031076  6.6049889 Circle
#> 481  8.26215983  6.6624456 Circle
#> 482  5.94870561  7.1688868 Circle
#> 483  6.99510758 10.7613430 Circle
#> 484  6.20557376  7.9632999 Circle
#> 485 10.71004678  6.6122624 Circle
#> 486  4.76243109  6.9590829 Circle
#> 487  8.88591078  6.2693344 Circle
#> 488  8.58823291  6.6864185 Circle
#> 489  6.23103201  7.3876014 Circle
#> 490  5.24228550  7.8237922 Circle
#> 491 10.24446639 10.6972541 Circle
#> 492  5.10045404 10.2924823 Circle
#> 493  6.53502238  8.3771624 Circle
#> 494  7.18865017 10.9377012 Circle
#> 495  6.69969839  8.6828621 Circle
#> 496 11.64598401  6.8993315 Circle
#> 497  8.46166936  5.7387811 Circle
#> 498 11.01911679  7.9131095 Circle
#> 499  6.99174228  9.4838778 Circle
#> 500  8.34761447  6.4660293 Circle
#> 501  8.31449453  6.4594876 Circle
#> 502  6.29334534  7.2069715 Circle
#> 503  8.47709798  7.7609799 Circle
#> 504  9.61083165  8.9690803 Circle
#> 505  5.73437339 10.2384214 Circle
#> 506  7.32492660  8.5851005 Circle
#> 507  7.69534930  8.2178577 Circle
#> 508  9.73761080 11.6777207 Circle
#> 509  8.60637733  9.4493841 Circle
#> 510  7.22921319  6.9204620 Circle
#> 511  7.45731494  6.4124960 Circle
#> 512  9.06019128  5.4954520 Circle
#> 513  8.47768114  7.8339833 Circle
#> 514  5.08913741  8.7482644 Circle
#> 515  6.84194243  7.9405463 Circle
#> 516 10.39828651  5.9915052 Circle
#> 517  9.76337376  7.4081206 Circle
#> 518  6.76898532  6.3388287 Circle
#> 519  9.31864268  5.7419678 Circle
#> 520  5.26545311  8.2379011 Circle
#> 521  7.79385530  5.9411344 Circle
#> 522  7.59389943  9.5835486 Circle
#> 523  9.06186650  9.0664359 Circle
#> 524  7.25115335 11.1078616 Circle
#> 525 10.10381754  5.2682458 Circle
#> 526  6.19127140  7.5495023 Circle
#> 527  6.04461445  6.6228548 Circle
#> 528  8.07281625  7.6815472 Circle
#> 529  9.81912049  6.0655492 Circle
#> 530  8.48556741  8.6578919 Circle
#> 531  9.14714412  5.9516067 Circle
#> 532  8.89879127 10.2249247 Circle
#> 533  7.43341137  7.1059730 Circle
#> 534  7.26596045  8.8950737 Circle
#> 535  9.62117306  7.9682791 Circle
#> 536  6.86208839  5.7300095 Circle
#> 537  8.59974173  4.8365722 Circle
#> 538  9.43820612  5.7403184 Circle
#> 539  4.71840279  4.9826835 Circle
#> 540  6.40870564  9.5507926 Circle
#> 541  4.26401468  7.2625882 Circle
#> 542  6.47679897 10.0738046 Circle
#> 543  6.68485931  6.5691085 Circle
#> 544  8.97877643  6.7754841 Circle
#> 545  9.43866784  8.6303613 Circle
#> 546  8.70514249 11.2135781 Circle
#> 547  7.69378538  8.4535453 Circle