ROSE Algorithm

A thin wrapper around ROSE::ROSE() that generates a synthetic balanced sample by enlarging the feature space of minority and majority class examples.

Usage

rose(
  df,
  var,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1
)

Arguments

df

A data.frame or tibble. Must have 1 factor variable with exactly 2 levels and remaining numeric variables.

var

Character, name of variable containing the 2-level factor variable.

over_ratio

A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows as the majority level. See vignette("ratio", package = "themis") for more details.

minority_prop

A numeric value between 0 and 1 for the proportion of synthetic observations from the minority class. Defaults to 0.5, which generates an equal split of minority and majority synthetic observations. This parameter controls the class balance within the synthetic data, while over_ratio controls the total size of the synthetic data.

minority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the minority class. Defaults to 1.

majority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the majority class. Defaults to 1.

Value

A data.frame or tibble, depending on type of df.

Details

The factor variable used to balance around must only have 2 levels.

The ROSE algorithm works by selecting an observation belonging to class k and generating new examples in its neighborhood, which is determined by a smoothing matrix H_k. Smaller values of minority_smoothness and majority_smoothness shrink the entries of H_k, producing tighter neighborhoods. This is a cautious choice when there is a concern that excessively large neighborhoods could blur the boundaries between classes.

This function is a thin wrapper around ROSE::ROSE(). For full details on the underlying implementation, see that function's documentation.

References

Lunardon, N., Menardi, G., and Torelli, N. (2014). ROSE: a Package for Binary Imbalanced Learning. R Journal, 6:82–92.

Menardi, G. and Torelli, N. (2014). Training and assessing classification rules with imbalanced data. Data Mining and Knowledge Discovery, 28:92–122.

See also

Examples

rose(circle_example[, c("x", "y", "class")], var = "class")
#>               x           y  class
#> 1   12.95910362  6.70142128   Rest
#> 2    8.39161860  1.88574553   Rest
#> 3    3.75114882 10.27031779   Rest
#> 4   11.86363706 13.08582837   Rest
#> 5   10.58746591 11.04046510   Rest
#> 6   10.26828744 12.83908252   Rest
#> 7    1.93867851  4.00296793   Rest
#> 8   10.55304879  4.89552970   Rest
#> 9    7.14298896 10.52157845   Rest
#> 10   5.83116908 12.52132615   Rest
#> 11   5.21995878  9.51362857   Rest
#> 12   0.20952651 -0.13567746   Rest
#> 13  12.62821217 13.10533850   Rest
#> 14  12.74104135  1.69725908   Rest
#> 15   9.50957911 10.92379237   Rest
#> 16   0.35952418  3.36190279   Rest
#> 17   9.40363425 13.40195847   Rest
#> 18  12.35948696  9.20698654   Rest
#> 19   0.79822347 14.08798093   Rest
#> 20   4.09478686  1.64102233   Rest
#> 21  13.49849774  3.12429117   Rest
#> 22  15.34249129  4.51418376   Rest
#> 23   1.83056664  1.20281606   Rest
#> 24  12.38529233  1.62238465   Rest
#> 25   0.92482696 11.88692813   Rest
#> 26  14.92322411  4.04495750   Rest
#> 27   6.09678654 11.28607379   Rest
#> 28  12.16668273 11.63420751   Rest
#> 29   3.92504750 10.65238515   Rest
#> 30   1.39401239  3.16099895   Rest
#> 31  13.40694658 -2.60592173   Rest
#> 32  10.33153197 15.47134360   Rest
#> 33  14.35451644  8.65676716   Rest
#> 34  12.84752428 13.39853679   Rest
#> 35  14.14535112  4.87511337   Rest
#> 36   5.41608562  6.83775480   Rest
#> 37   2.05161914  5.59375792   Rest
#> 38   8.15822641 15.74844547   Rest
#> 39   7.07527177  7.22939850   Rest
#> 40   4.47929270  3.84975824   Rest
#> 41  16.61883090 13.62247490   Rest
#> 42   4.99815805  9.72870233   Rest
#> 43   0.54838156  2.01892470   Rest
#> 44   2.75705753  9.65375824   Rest
#> 45   6.43734949 13.68655758   Rest
#> 46  10.49507610  3.28363116   Rest
#> 47  11.32980033  1.87698265   Rest
#> 48   0.08529301  4.48866861   Rest
#> 49   3.61490858 12.42016555   Rest
#> 50  14.82282735 12.95989009   Rest
#> 51   1.97923065  9.16731644   Rest
#> 52   0.26185203  1.45683781   Rest
#> 53   7.01395003  2.29337811   Rest
#> 54   5.65899638 12.46470398   Rest
#> 55  10.14199748 10.96357569   Rest
#> 56   4.35454306  8.56348091   Rest
#> 57  12.94240973  1.47918323   Rest
#> 58   3.13942529  4.57930415   Rest
#> 59  15.07955114  6.92959542   Rest
#> 60   3.80311270  9.01490818   Rest
#> 61  10.96637192  5.93019921   Rest
#> 62  14.36551172 13.12951011   Rest
#> 63   5.10167125 13.50632199   Rest
#> 64   5.48603656  0.42358673   Rest
#> 65  14.33174333  9.42790169   Rest
#> 66   4.72405077 10.11841002   Rest
#> 67   0.21192444 13.08872832   Rest
#> 68   0.98698042  9.81218498   Rest
#> 69   1.58443030  7.80720396   Rest
#> 70   3.47701911  8.91926438   Rest
#> 71   5.26054043 12.82562542   Rest
#> 72   5.99505502  1.48950466   Rest
#> 73  16.48206594  5.36620420   Rest
#> 74   7.57439695  8.63601687   Rest
#> 75  12.69301565 12.62432896   Rest
#> 76   3.48973722 10.00166502   Rest
#> 77   3.72835781  3.36938834   Rest
#> 78   8.84677036  7.27969037   Rest
#> 79  13.34284951  8.02776592   Rest
#> 80   8.11081618 10.41581008   Rest
#> 81   0.90588654  7.06901832   Rest
#> 82  -0.71969521 10.25840028   Rest
#> 83  -0.05643767 -0.14859229   Rest
#> 84  11.24180916  8.18076108   Rest
#> 85   3.69341628 15.20238544   Rest
#> 86   8.72967356 10.30852483   Rest
#> 87  12.11185538  8.99119696   Rest
#> 88   5.98310050 14.29697325   Rest
#> 89  13.96637886  9.18503621   Rest
#> 90   9.40992782  1.26509537   Rest
#> 91  14.22359974  9.66862115   Rest
#> 92   6.81524640 11.79699520   Rest
#> 93  10.00196696 10.08642566   Rest
#> 94  15.98180104 13.75379381   Rest
#> 95  -1.53830730  7.90864995   Rest
#> 96  14.58802381  3.69590707   Rest
#> 97   9.79953938  4.84692275   Rest
#> 98   4.21268881 12.80674130   Rest
#> 99   2.38478979 12.68587555   Rest
#> 100  0.14640532  9.98650369   Rest
#> 101  8.69059199 14.59097264   Rest
#> 102  9.58262162 12.49195785   Rest
#> 103  5.32857851  9.10788702   Rest
#> 104  8.46542835  6.85800787   Rest
#> 105  5.80144381  3.86211971   Rest
#> 106  2.12503459 10.64328456   Rest
#> 107 -0.52853960  7.51462617   Rest
#> 108 11.29975971 13.42343395   Rest
#> 109  6.00182354 11.63974315   Rest
#> 110  7.04653141 11.30849758   Rest
#> 111  3.44414978  4.57423426   Rest
#> 112  9.16644374 10.11174159   Rest
#> 113  8.70629762 12.67482996   Rest
#> 114 11.45597196 10.56626993   Rest
#> 115 -0.19701643  5.99463271   Rest
#> 116 -0.84866450  4.51888097   Rest
#> 117  5.19800225  5.42762239   Rest
#> 118 11.18627541 13.77978988   Rest
#> 119  1.65870215  8.06439816   Rest
#> 120 12.24956774  7.84068423   Rest
#> 121  7.11718636 17.48040593   Rest
#> 122 11.27773823 11.54673538   Rest
#> 123 14.02397192  3.81666677   Rest
#> 124  7.05622675  0.38323665   Rest
#> 125  1.77516652  9.23999370   Rest
#> 126  3.57452906 10.94713577   Rest
#> 127  7.22555990  6.29762905   Rest
#> 128  4.46890562 11.10750550   Rest
#> 129  6.71307768 12.38527589   Rest
#> 130 14.88982811 16.26395315   Rest
#> 131  5.79398006 14.56815538   Rest
#> 132  9.13742316  7.29507985   Rest
#> 133  3.13893627 14.52731276   Rest
#> 134  6.65359415  9.65056563   Rest
#> 135 13.21326492  4.31528703   Rest
#> 136 -0.48192623 12.14728930   Rest
#> 137 15.35823335 10.21164934   Rest
#> 138  8.67777694 10.48208080   Rest
#> 139 10.55737096  2.46887489   Rest
#> 140 -0.31788626 13.57560219   Rest
#> 141 15.32480325 13.05833110   Rest
#> 142 14.68501499  8.36909436   Rest
#> 143 11.17944715  4.80334311   Rest
#> 144  2.34180204  4.52102887   Rest
#> 145 16.45994657  8.90760004   Rest
#> 146 10.03315205 13.96283980   Rest
#> 147 10.94702235 14.91977608   Rest
#> 148  2.81402904  7.54997928   Rest
#> 149  7.83354713  2.75712620   Rest
#> 150 15.97087512  0.46741345   Rest
#> 151  4.33971814 13.81960436   Rest
#> 152  4.23727060 12.30621546   Rest
#> 153  0.57223240  3.27611094   Rest
#> 154 13.28828693  9.92338025   Rest
#> 155  9.44511951 17.26497984   Rest
#> 156  3.83708357  4.24348985   Rest
#> 157  3.42995140 13.50716017   Rest
#> 158  3.11445803 12.05179397   Rest
#> 159  0.73520883 13.04048168   Rest
#> 160  5.48195975 14.52344418   Rest
#> 161  1.85058900  4.08572207   Rest
#> 162 12.88976504  0.53783580   Rest
#> 163 13.27282342  2.76391340   Rest
#> 164  2.41309501 10.74098742   Rest
#> 165  3.54112981  2.39347221   Rest
#> 166 10.98409888  3.70276223   Rest
#> 167 10.74405520 15.70191188   Rest
#> 168 10.45977919 14.56896900   Rest
#> 169 13.53948418  5.00390448   Rest
#> 170 10.69834492 17.62636884   Rest
#> 171  4.72502348 15.67574176   Rest
#> 172  4.83659429 15.96027186   Rest
#> 173 11.94610094 11.58698168   Rest
#> 174  7.08252709 11.93055224   Rest
#> 175 11.32683158  5.93050131   Rest
#> 176  5.89343329 12.95941097   Rest
#> 177 11.20940756 13.18808609   Rest
#> 178 14.03237589 12.97490035   Rest
#> 179  8.62236639  3.78954739   Rest
#> 180  3.54398585  3.65640092   Rest
#> 181  3.18483707  7.87408928   Rest
#> 182  0.13278731  9.49488637   Rest
#> 183 11.92065742 10.82991794   Rest
#> 184  7.35677701  0.61558071   Rest
#> 185 15.00884419 11.38581610   Rest
#> 186 13.78265784 14.45501471   Rest
#> 187  2.75754508 14.81235169   Rest
#> 188  7.59080655 13.62862201   Rest
#> 189  6.00315735  4.63239767   Rest
#> 190 14.28367954  1.07877777   Rest
#> 191  9.36288379 11.65547921   Rest
#> 192 -1.02367229 14.97586397   Rest
#> 193  9.99882380 12.18684165   Rest
#> 194 14.85040513  6.68364934   Rest
#> 195  1.63697112 14.59752425   Rest
#> 196 15.37616508 14.03236443   Rest
#> 197  2.83029016 14.85416818   Rest
#> 198 10.82561455  9.45244592   Rest
#> 199  6.97798927  1.76225597   Rest
#> 200  4.36949912 15.02711079   Rest
#> 201  6.68867092 17.10372902   Rest
#> 202  2.22449626  5.72851200   Rest
#> 203  5.56339193 13.21587090   Rest
#> 204  2.13452154 11.56034019   Rest
#> 205  3.66502996 10.58794668   Rest
#> 206  4.35219548  3.36298929   Rest
#> 207  9.62626025 12.72561644   Rest
#> 208  8.26092356  9.86032082   Rest
#> 209 10.61153293  7.33470416   Rest
#> 210  9.67705889 10.93415152   Rest
#> 211  4.82611948  5.39364474   Rest
#> 212 10.67830450 16.06211141   Rest
#> 213  3.53335036  4.26330214   Rest
#> 214  2.28405741  5.54707708   Rest
#> 215  2.28926527  3.67979518   Rest
#> 216  5.32477453  6.43386901   Rest
#> 217  8.23493802  4.24609754   Rest
#> 218  5.19937483  0.29149352   Rest
#> 219 10.74575572  0.40309145   Rest
#> 220  9.51019114  4.48240068   Rest
#> 221 14.11725152  5.10042256   Rest
#> 222  8.52709361  0.96921371   Rest
#> 223  9.86855877 12.63247470   Rest
#> 224  3.46892230 10.50022073   Rest
#> 225  8.61698315  4.62999846   Rest
#> 226 13.09637847  3.84957327   Rest
#> 227  0.09290335 12.54005406   Rest
#> 228  8.54736620 16.67527581   Rest
#> 229 10.81851695 12.52940991   Rest
#> 230  9.54576124  7.30954019   Rest
#> 231 17.22801658  8.70122065   Rest
#> 232  5.92674386  1.50689988   Rest
#> 233  4.37101683  5.23374269   Rest
#> 234  3.54219730  9.26886548   Rest
#> 235  4.74302532  9.61201777   Rest
#> 236  5.07897317 12.41117491   Rest
#> 237  7.19661952 11.77117482   Rest
#> 238 15.02396598  8.26587905   Rest
#> 239  0.31678759 13.91822623   Rest
#> 240  6.68318982 12.47623160   Rest
#> 241  4.32879302  0.88060464   Rest
#> 242  2.60084102 12.67712327   Rest
#> 243  4.65866604  7.85456586   Rest
#> 244 14.59616682  3.83889758   Rest
#> 245 10.99968064  3.01314650   Rest
#> 246  2.73283720  4.24478219   Rest
#> 247  1.15831876 12.35923386   Rest
#> 248 12.13206052  4.51141148   Rest
#> 249 13.02338985  4.42763855   Rest
#> 250  3.92427534 15.43775820   Rest
#> 251  6.54422630 12.69353978   Rest
#> 252 14.34796052  3.77173058   Rest
#> 253  7.77447712  3.98162290   Rest
#> 254  6.89650051 18.18645502   Rest
#> 255 11.47862844 10.87267658   Rest
#> 256  9.94312391  4.04479136   Rest
#> 257 13.76546176  9.52506145   Rest
#> 258 10.43793688  4.96713445   Rest
#> 259 10.24910990  7.35836209   Rest
#> 260  7.75158836 11.75045917   Rest
#> 261  9.95465194  9.56158837   Rest
#> 262  0.28509932  0.01002563   Rest
#> 263 12.74108204  4.75887721   Rest
#> 264 10.42210896  8.31770014   Rest
#> 265  1.28066549 10.78867229   Rest
#> 266  0.46639180  1.37512288   Rest
#> 267 13.81602489  4.88458169   Rest
#> 268  3.72468375 -1.34895429   Rest
#> 269  7.34524714  1.04670410   Rest
#> 270 14.90940219  7.33835451   Rest
#> 271 11.65061928 11.67555044   Rest
#> 272 10.94714044  9.51653832   Rest
#> 273 12.74536486  5.93892268   Rest
#> 274  3.03158709 12.55576874   Rest
#> 275  6.06018715  0.07240212   Rest
#> 276 14.41760531  9.72682932   Rest
#> 277  6.68855687 12.81381350   Rest
#> 278  5.23224383  2.37882086   Rest
#> 279  3.87429421  4.18220307   Rest
#> 280 10.11523236  2.61969883   Rest
#> 281  6.52846760  8.08363457   Rest
#> 282  8.16155262 11.57745111   Rest
#> 283 13.75546335  2.75373329   Rest
#> 284  8.74740281 11.01534317   Rest
#> 285  6.76452259 13.01845769   Rest
#> 286  5.46804383  5.69728352   Rest
#> 287  1.16478986 15.80750759   Rest
#> 288  3.07394019 11.55323946   Rest
#> 289 13.03544594  3.94572015   Rest
#> 290  0.98861307 12.47941467   Rest
#> 291  7.49500669 14.24363989   Rest
#> 292 12.98581460  4.63233765   Rest
#> 293  7.42510737  1.21916557   Rest
#> 294  9.85480117  5.13937434   Rest
#> 295  5.31101626  4.42691349   Rest
#> 296 14.69038791  9.36460081   Rest
#> 297 13.44319043 11.68308602   Rest
#> 298 16.32219841 11.59002234   Rest
#> 299  8.17661013  4.88171846   Rest
#> 300  9.25760242 11.40127102   Rest
#> 301 11.50912014 10.74237956   Rest
#> 302  7.45299529 15.14029056   Rest
#> 303  4.73249663  9.09890958   Rest
#> 304  0.89016960  2.13690344   Rest
#> 305  7.67134587 13.08484188   Rest
#> 306 15.50264247  6.46849408   Rest
#> 307  0.62361634  0.21190801   Rest
#> 308 13.53002911 10.52401010   Rest
#> 309  8.21049929 12.30607014   Rest
#> 310 11.80719512  3.38742737   Rest
#> 311  3.30766954  4.17380294   Rest
#> 312  8.05371730  2.77536230   Rest
#> 313  6.92569908 12.72249813   Rest
#> 314 10.56702167  8.82260152   Rest
#> 315 14.26612175  3.50890702   Rest
#> 316  1.78197329  4.25518048   Rest
#> 317 14.91262605  6.10134095   Rest
#> 318 11.41606050 13.39767912   Rest
#> 319  3.82475564 -0.52156740   Rest
#> 320  4.81935026  7.53789347   Rest
#> 321 12.19073083 11.74177867   Rest
#> 322  8.81777663  1.21293756   Rest
#> 323  4.33738781 15.74345081   Rest
#> 324  2.49516186  7.75755314   Rest
#> 325 13.02983941 14.58495560   Rest
#> 326  6.68710374 10.87699511   Rest
#> 327 14.11387146  4.36013770   Rest
#> 328  1.31033091 14.44323251   Rest
#> 329  3.08588802  9.44828343   Rest
#> 330  4.47470544  6.58224983   Rest
#> 331  4.32102473  3.25901037   Rest
#> 332  2.19392717  2.67150027   Rest
#> 333 13.87134626  3.33662971   Rest
#> 334  1.96689788  8.11938260   Rest
#> 335  2.29535211  4.34023561   Rest
#> 336  6.28728154  7.43301086   Rest
#> 337  2.61730380  7.33878180   Rest
#> 338 -0.01202511  6.09986743   Rest
#> 339  1.35493999 17.42631464   Rest
#> 340  4.57944040 14.94539270   Rest
#> 341  6.15025240 10.07956511   Rest
#> 342 15.66725543 13.95035044   Rest
#> 343 11.00030672  9.79258376 Circle
#> 344  9.56513988  7.76860910 Circle
#> 345 10.05051095  8.19772122 Circle
#> 346  5.75420626  8.52226664 Circle
#> 347  4.92997413  6.28103918 Circle
#> 348  6.76027730  6.30431551 Circle
#> 349  9.94045309  9.63665304 Circle
#> 350  9.73054397  5.23234321 Circle
#> 351  8.04943831  8.77188720 Circle
#> 352  8.43741409 10.45579618 Circle
#> 353 11.33007662  8.57103895 Circle
#> 354  7.82469632  9.90177695 Circle
#> 355 10.14732579  9.00954555 Circle
#> 356  5.95252321  5.63563892 Circle
#> 357  9.62804441  5.68592840 Circle
#> 358  9.10916971 11.17304417 Circle
#> 359  8.53029509  6.52051538 Circle
#> 360  6.59554577  9.39745680 Circle
#> 361  6.14295452 10.08568149 Circle
#> 362  5.29991101  6.40269861 Circle
#> 363 10.02125260 10.05713880 Circle
#> 364  7.98475916 10.07850162 Circle
#> 365  6.77339317  7.38718162 Circle
#> 366  7.96064015  5.38131587 Circle
#> 367  6.35894969  8.13769559 Circle
#> 368  8.59455188  8.25390649 Circle
#> 369  9.74607473  9.25059082 Circle
#> 370  6.74644758  8.27551398 Circle
#> 371  8.89064924  8.06127327 Circle
#> 372  6.74128860  9.63904665 Circle
#> 373  5.59157960  9.54440211 Circle
#> 374 10.00108311 10.05343551 Circle
#> 375 10.20021450 10.31297487 Circle
#> 376  5.75011051  8.51260275 Circle
#> 377  6.36545042  7.87091505 Circle
#> 378  7.27452190  9.19726202 Circle
#> 379  8.99671074 11.01127727 Circle
#> 380 10.47434365  8.08521143 Circle
#> 381 10.47489975  8.78351085 Circle
#> 382  7.61478064  8.19408908 Circle
#> 383  5.63684842  8.16857741 Circle
#> 384 10.47666443  9.37804037 Circle
#> 385  8.52467390  6.66372890 Circle
#> 386  8.20858289  7.95204673 Circle
#> 387  9.25272211  8.34892449 Circle
#> 388  8.89943625  7.21603321 Circle
#> 389  7.61645511  6.39326571 Circle
#> 390  5.85218929  6.62750277 Circle
#> 391  9.90973710  7.66339663 Circle
#> 392  7.06431228 10.08881583 Circle
#> 393  5.67561428  9.11821717 Circle
#> 394  7.84520961  5.35377943 Circle
#> 395  5.47767229  7.93314041 Circle
#> 396  8.42503383  9.75521869 Circle
#> 397 10.49276746  5.43334358 Circle
#> 398  8.33070711  7.96103609 Circle
#> 399  7.17613554  8.81568378 Circle
#> 400  9.61796799  8.10649832 Circle
#> 401  7.04303615  5.50234139 Circle
#> 402  5.31128325  9.19450286 Circle
#> 403  9.31826904 10.56681517 Circle
#> 404  6.84823024  8.79439042 Circle
#> 405  9.98618185  9.26458704 Circle
#> 406  7.27452229 10.77454335 Circle
#> 407  5.93013935  6.46173527 Circle
#> 408  7.51787577 10.82904383 Circle
#> 409  7.30878448  8.58325671 Circle
#> 410  6.07652196  7.73074935 Circle
#> 411  9.88671399 10.94625689 Circle
#> 412  6.39420843  9.02224971 Circle
#> 413  8.75394240  5.15768846 Circle
#> 414 10.46294434  8.76344322 Circle
#> 415  8.69065396  8.96167553 Circle
#> 416  6.81724588 11.45184412 Circle
#> 417  6.85344149  9.71293818 Circle
#> 418  8.88629589  6.67246707 Circle
#> 419  7.09577340  9.89892639 Circle
#> 420  6.08173477  8.12824742 Circle
#> 421  7.02289340  7.32647486 Circle
#> 422  7.61914631  8.47751082 Circle
#> 423  8.50372851  8.02288084 Circle
#> 424  5.75877826  7.87685032 Circle
#> 425  6.95293389  7.82617501 Circle
#> 426  6.72941849  7.34856214 Circle
#> 427  7.71645556  6.45860172 Circle
#> 428  6.03346066  7.01768401 Circle
#> 429  6.83135687  9.49171528 Circle
#> 430  6.57256883  6.55267255 Circle
#> 431  7.75750255  6.87733087 Circle
#> 432  6.51529738 10.94137963 Circle
#> 433  8.06670937  9.73984326 Circle
#> 434 10.02951398  9.91869057 Circle
#> 435  8.40925522 10.01786381 Circle
#> 436  8.48247400  8.17361322 Circle
#> 437  7.87923328 10.34264224 Circle
#> 438  8.14289465  8.80550679 Circle
#> 439 10.16798212  7.09335209 Circle
#> 440  8.07560606  4.95160672 Circle
#> 441  9.15679067  8.01409856 Circle
#> 442  6.84460248  9.34387870 Circle
#> 443  5.75335297  6.94885579 Circle
#> 444  6.06867096  7.72580644 Circle
#> 445  9.07924114  6.19514824 Circle
#> 446  5.26045967  9.25922227 Circle
#> 447  6.43348340  9.07133634 Circle
#> 448  7.54245342  7.59161019 Circle
#> 449  6.65561699  9.01253259 Circle
#> 450  5.37637476  9.56373474 Circle
#> 451 10.95979530  6.57351816 Circle
#> 452  8.72835049 11.59796652 Circle
#> 453  7.67818764  5.27623466 Circle
#> 454  9.02664773  8.83686402 Circle
#> 455  8.61659832  9.01755686 Circle
#> 456  8.80625612  5.61003603 Circle
#> 457  8.46090212  6.93168530 Circle
#> 458  7.89434233  4.66303199 Circle
#> 459  7.14062607  9.54148271 Circle
#> 460 10.48921399 10.46692772 Circle
#> 461  6.22294958  7.54812533 Circle
#> 462  8.48065468  5.53220901 Circle
#> 463  6.49346574  6.13914493 Circle
#> 464  7.84358992 10.74564874 Circle
#> 465  9.38047454  4.95376071 Circle
#> 466  8.53932561  6.96259875 Circle
#> 467  8.29411422  5.56228015 Circle
#> 468  6.89133110 11.21702308 Circle
#> 469 11.40869502  6.05268991 Circle
#> 470  6.20712951  5.55388870 Circle
#> 471  5.62614922  7.39838885 Circle
#> 472  8.87971543  4.54202071 Circle
#> 473  9.54653552  7.46374444 Circle
#> 474  9.34487767  7.65506068 Circle
#> 475  8.74524868  9.67452556 Circle
#> 476  8.56981169  6.66495083 Circle
#> 477  9.72639899  7.69829168 Circle
#> 478  9.69555503  8.74184909 Circle
#> 479  7.10804113  8.88245528 Circle
#> 480  6.05671893  7.27067138 Circle
#> 481  4.55641098  7.67516609 Circle
#> 482  7.68504510  9.29508925 Circle
#> 483  5.77802685  5.29136419 Circle
#> 484  7.99903294  8.07808550 Circle
#> 485  9.13339544  7.79658438 Circle
#> 486  6.43820810  8.24706739 Circle
#> 487  8.57061619  7.80277984 Circle
#> 488 11.41862880 10.61298337 Circle
#> 489  6.66645976  6.33837404 Circle
#> 490  6.23407894  7.99062110 Circle
#> 491  7.33007609  8.30196737 Circle
#> 492  9.72163669  6.46340288 Circle
#> 493  7.85371313  6.66081100 Circle
#> 494  8.93350659  7.62035800 Circle
#> 495  9.89692243  8.05692640 Circle
#> 496 11.06336759  8.08884884 Circle
#> 497  9.28665072  6.82125870 Circle
#> 498  8.54228182  7.97495012 Circle
#> 499  8.78456433 11.04999861 Circle
#> 500  9.18442456  8.76466601 Circle
#> 501  6.85732075  9.76342133 Circle
#> 502  9.39346120  9.81446288 Circle
#> 503  7.55452609  6.03246753 Circle
#> 504  7.69200559 10.78100863 Circle
#> 505  5.53850646  8.72001053 Circle
#> 506  9.35400106 10.10302768 Circle
#> 507  9.29085155  8.41453074 Circle
#> 508  8.33764415  6.49334608 Circle
#> 509  6.62247955  7.55129310 Circle
#> 510  6.85253783  8.63539616 Circle
#> 511  8.98878787  6.84694142 Circle
#> 512  5.89971709  6.53638468 Circle
#> 513  8.18321237  9.11662529 Circle
#> 514  5.18182173  9.39225681 Circle
#> 515  8.29464437  8.06677978 Circle
#> 516  6.78765028  7.89824540 Circle
#> 517  8.75720323  9.12164495 Circle
#> 518  8.20257760  9.57020456 Circle
#> 519  6.98042002  8.25494230 Circle
#> 520  8.87034411  8.33722209 Circle
#> 521  8.52315742  8.86345478 Circle
#> 522 11.46177628  9.18722180 Circle
#> 523  6.45412703  8.93778250 Circle
#> 524  8.96505816  5.88370611 Circle
#> 525  6.72055832 10.93467741 Circle
#> 526  8.79976996 10.76187798 Circle
#> 527  6.52359933  8.38886253 Circle
#> 528 11.41571246  9.84150566 Circle
#> 529  6.76719559  8.37569236 Circle
#> 530  8.52798246  5.91481173 Circle
#> 531  8.06409069 10.65894915 Circle
#> 532  6.50821166 11.37835524 Circle
#> 533  7.99370422  4.94225971 Circle
#> 534  7.51387837  9.79123277 Circle
#> 535  6.17790245  7.65171285 Circle
#> 536  7.85109135 10.40940117 Circle
#> 537  9.19341989 10.54760617 Circle
#> 538  9.10390907  4.85760978 Circle
#> 539  7.58490632  7.53858704 Circle
#> 540 11.59899669  4.96753904 Circle
#> 541  6.64272203  9.01149836 Circle
#> 542  5.49627717  6.48809368 Circle
#> 543  6.34536192  8.05958011 Circle
#> 544  9.75698955  5.98952791 Circle
#> 545  8.78809201  9.66566736 Circle
#> 546 10.03900448  4.83199537 Circle
#> 547  8.35061967  6.15276038 Circle
#> 548  8.48732557  4.61169659 Circle
#> 549  7.63375197  8.81486203 Circle
#> 550  5.85234006  8.74397179 Circle
#> 551  9.88008159  7.81641445 Circle
#> 552  7.91514286  7.84045720 Circle
#> 553  5.55528047  9.53684448 Circle
#> 554  7.94079469  7.33726342 Circle
#> 555  7.71866857  5.90218951 Circle
#> 556  4.24964503  7.75859921 Circle
#> 557  8.85163242  7.09892226 Circle
#> 558  8.22803928  6.86076890 Circle
#> 559  4.66774843  8.54356758 Circle
#> 560  8.97830403 11.73877238 Circle
#> 561  8.00590312  6.73594426 Circle
#> 562  6.35344350  8.86628813 Circle
#> 563  7.25214861  8.22401185 Circle
#> 564  9.34464401 10.46475883 Circle
#> 565  6.20561459  6.78159059 Circle
#> 566  5.96481339  7.91660802 Circle
#> 567  6.03350107  9.86603016 Circle
#> 568  8.01388639  6.91674072 Circle
#> 569  6.88005512  8.91618661 Circle
#> 570  7.76503798  9.28338774 Circle
#> 571  6.49822109 10.83364181 Circle
#> 572  9.88566543  3.66743374 Circle
#> 573  7.41148885  7.74831195 Circle
#> 574  7.97954615  6.46397471 Circle
#> 575  9.29914628  7.36517643 Circle
#> 576 10.57110712  8.39926577 Circle
#> 577  8.62690260 10.34991321 Circle
#> 578  9.05154724  7.26224441 Circle
#> 579  8.01431058  7.93321417 Circle
#> 580  5.71788530  7.05854979 Circle
#> 581  8.73420821  7.12995538 Circle
#> 582  8.54066713  6.50110062 Circle
#> 583  5.84211663  6.80313304 Circle
#> 584  7.34150099  8.33326185 Circle
#> 585  8.65121595  5.09060271 Circle
#> 586  9.49171726  6.37206914 Circle
#> 587  9.78556214 11.23344928 Circle
#> 588  6.45623827  5.58187013 Circle
#> 589  5.87743359 10.01774159 Circle
#> 590  8.46120878  9.03482990 Circle
#> 591  5.79355633  9.45977527 Circle
#> 592  6.22134664  7.88727721 Circle
#> 593  7.58032495  7.83726290 Circle
#> 594  7.04369642  8.29574910 Circle
#> 595 10.75707588 10.68667637 Circle
#> 596  8.45528030  6.61389875 Circle
#> 597  6.18211234  7.06168155 Circle
#> 598  8.17568740  7.07611904 Circle
#> 599  9.30295929 10.44319663 Circle
#> 600  8.08071056  6.25800532 Circle
#> 601  7.41900453  9.87601527 Circle
#> 602  8.51050588  6.60234707 Circle
#> 603  7.67037705 10.91773169 Circle
#> 604  6.31641169  7.23414649 Circle
#> 605  8.13692137  6.83776690 Circle
#> 606 10.05920500  6.46827305 Circle
#> 607  7.20309070  9.88798936 Circle
#> 608 10.12988443  9.69570569 Circle
#> 609  9.83276930  7.68766849 Circle
#> 610  6.79733035 11.35575844 Circle
#> 611  8.38519983  7.80075178 Circle
#> 612  8.22476094  8.17970090 Circle
#> 613  9.03800382  9.21886400 Circle
#> 614  8.56170279  8.74530810 Circle
#> 615  4.82905255  8.30967717 Circle
#> 616  9.08352803  8.11429379 Circle
#> 617  6.55773915  8.79192858 Circle
#> 618  7.90510140  6.78681899 Circle
#> 619  8.96336030 11.25266310 Circle
#> 620  4.95247842  5.83833334 Circle
#> 621  8.47958708  4.58565837 Circle
#> 622  9.11741994  6.80133780 Circle
#> 623  6.91474699  7.25514642 Circle
#> 624  7.15325249  8.48774543 Circle
#> 625  9.15741844  7.24676137 Circle
#> 626  6.62384686  6.12053082 Circle
#> 627  6.39962368  7.85230270 Circle
#> 628  8.18748879  7.00010145 Circle
#> 629  7.60899635  6.63615853 Circle
#> 630  8.09654837  8.11942771 Circle
#> 631 10.33772060  8.38184073 Circle
#> 632  8.26628817  9.27455499 Circle
#> 633 10.30448139  5.32027818 Circle
#> 634  8.76859642  8.61184212 Circle
#> 635  8.41486984  4.95808394 Circle
#> 636 10.45558513  9.26772245 Circle
#> 637  9.97535916  7.54606512 Circle
#> 638  7.38442936  9.17595589 Circle
#> 639  6.52706708  9.24619945 Circle
#> 640  9.81098243  8.59219002 Circle
#> 641  8.13220603  8.99660062 Circle
#> 642  8.79736091  8.97375780 Circle
#> 643  8.46217109  7.14051442 Circle
#> 644  4.59782986  6.70909797 Circle
#> 645  9.44303931  9.81470417 Circle
#> 646  6.75310661  4.76267007 Circle
#> 647 10.14595021  9.46043571 Circle
#> 648 10.99370226  9.04252535 Circle
#> 649  8.76264875  5.79897464 Circle
#> 650  7.99393598 11.15430785 Circle
#> 651  9.57978988  9.65366338 Circle
#> 652  6.97732506  7.85212874 Circle
#> 653  7.96415282 10.66803238 Circle
#> 654  9.88744116 10.95185060 Circle
#> 655  7.81243247  9.55092796 Circle
#> 656  7.75241795  6.09487353 Circle
#> 657  7.23670922  6.01744373 Circle
#> 658  5.95520469  6.45685014 Circle
#> 659  5.98930587  8.51402974 Circle
#> 660  9.08924592  7.52104723 Circle
#> 661  8.39207941  9.88698541 Circle
#> 662  8.22409967  9.33311044 Circle
#> 663  8.28154211  6.58119512 Circle
#> 664  9.39184596  9.41728753 Circle
#> 665  6.03656726  8.07310618 Circle
#> 666  7.65053301  8.14629487 Circle
#> 667  9.55273416  7.17168854 Circle
#> 668 11.65600874  9.46399198 Circle
#> 669  5.94958135  8.52201590 Circle
#> 670  8.69552304  7.96286991 Circle
#> 671  8.97862328  6.39907975 Circle
#> 672  6.49227144 10.20662485 Circle
#> 673  6.04166863  8.33400844 Circle
#> 674  8.58732093  9.38003377 Circle
#> 675  8.37143965  8.08136743 Circle
#> 676 11.06046712  9.71797369 Circle
#> 677 11.19620992  6.47298704 Circle
#> 678  6.83833386  5.76362617 Circle
#> 679 10.13076624  8.41577789 Circle
#> 680  6.10154444  7.62224622 Circle
#> 681  6.12063905  7.91187824 Circle
#> 682  8.91215395 10.29610527 Circle
#> 683  9.16841458  8.45175127 Circle
#> 684 10.49176550  4.85025068 Circle

rose(circle_example[, c("x", "y", "class")], var = "class", over_ratio = 0.8)
#>               x           y  class
#> 1    4.91656861  7.78832180   Rest
#> 2    5.47945992 12.69638853   Rest
#> 3   11.52588485 12.72436633   Rest
#> 4   10.85931153 13.80601017   Rest
#> 5    9.28746653 12.62622602   Rest
#> 6    9.81068206 12.06830277   Rest
#> 7    5.17180841  1.70469011   Rest
#> 8    0.77227262  1.93840694   Rest
#> 9    8.11077018 16.59964819   Rest
#> 10   9.32778386 12.79671109   Rest
#> 11   1.88258493  1.12104970   Rest
#> 12   0.75295569  6.64719079   Rest
#> 13   9.57239268 10.70400922   Rest
#> 14  14.42501066 13.79550722   Rest
#> 15   3.28208251  6.41048946   Rest
#> 16  12.19883341 10.52812715   Rest
#> 17   1.84556442 12.70571656   Rest
#> 18   5.04587947  6.76654903   Rest
#> 19   5.82219672  7.32442308   Rest
#> 20  15.97035429 12.58925810   Rest
#> 21   2.71759770  6.76480325   Rest
#> 22   4.02203601  7.98305971   Rest
#> 23  11.33346660  5.07536329   Rest
#> 24  12.10131128 12.34648831   Rest
#> 25   0.36710444 16.53287180   Rest
#> 26   1.84762744 11.76223195   Rest
#> 27   1.07114050  2.08061225   Rest
#> 28   1.41073510 12.99843107   Rest
#> 29   2.30368657  9.35028303   Rest
#> 30  10.47030762  3.69368816   Rest
#> 31   2.35150118  6.79589260   Rest
#> 32   2.22945378 -2.23517858   Rest
#> 33  10.55972244 16.53392275   Rest
#> 34   6.89650068 14.46829693   Rest
#> 35   4.02325518  1.05061268   Rest
#> 36   6.08811655 10.92603950   Rest
#> 37   6.72510741  6.42542941   Rest
#> 38   3.01469812  8.13198898   Rest
#> 39   2.93877847 12.31678556   Rest
#> 40   9.80138562  8.67205348   Rest
#> 41   0.42260372  5.97340989   Rest
#> 42  14.40759271 11.36885015   Rest
#> 43   2.71098263 13.91338618   Rest
#> 44   4.53795883 11.88446872   Rest
#> 45   5.96288228 10.65272328   Rest
#> 46  12.82046421 15.05150102   Rest
#> 47   0.75300191  8.58963557   Rest
#> 48  12.11952580  3.82333719   Rest
#> 49   6.46968474  3.60474932   Rest
#> 50  10.57801201 14.08918366   Rest
#> 51   2.94054627  1.26322214   Rest
#> 52  10.04531164  7.16327567   Rest
#> 53   1.77631410 15.18361207   Rest
#> 54   4.54161058 12.88308439   Rest
#> 55  13.98045414  6.08320739   Rest
#> 56   6.92631560  0.99029704   Rest
#> 57   4.72425577 14.38409947   Rest
#> 58  12.59193282 -0.03087085   Rest
#> 59   9.62542481  1.86966027   Rest
#> 60  11.96601656 11.82749403   Rest
#> 61  14.25376420  6.01427997   Rest
#> 62   3.46788123 -1.25973881   Rest
#> 63   9.32338336 10.74466378   Rest
#> 64  12.48278813 13.88392999   Rest
#> 65  10.54304642 11.11712510   Rest
#> 66   7.35484780 14.25488233   Rest
#> 67   4.04060873  7.39649271   Rest
#> 68  12.12500614 10.24121188   Rest
#> 69   2.93221055 11.79360944   Rest
#> 70   6.87483556 15.41725562   Rest
#> 71   9.54650619 11.99126116   Rest
#> 72   4.71963956 12.23385068   Rest
#> 73   8.01768891  1.30407542   Rest
#> 74   7.82006557 11.49779410   Rest
#> 75   9.22452497 14.52386942   Rest
#> 76   2.77070681 10.20427811   Rest
#> 77   5.32731756  4.52187801   Rest
#> 78   1.72838781  8.58055793   Rest
#> 79   7.18122387 11.09284318   Rest
#> 80  11.04961070 13.79727209   Rest
#> 81   1.46241238  7.84561291   Rest
#> 82  14.29306383  5.25175365   Rest
#> 83   2.10182388 10.22909861   Rest
#> 84  13.20973135 14.17772088   Rest
#> 85   8.14867814  2.89332318   Rest
#> 86   3.92132351  4.48883582   Rest
#> 87   7.69328498 18.61342042   Rest
#> 88  10.03211999 11.12758141   Rest
#> 89   4.51303502  5.85863587   Rest
#> 90   5.89277747 14.44419320   Rest
#> 91   4.46055102 13.18319720   Rest
#> 92   5.95291928 12.80914986   Rest
#> 93   1.53907023 13.10066852   Rest
#> 94   4.64463884 11.36495958   Rest
#> 95   0.48717483 14.34023926   Rest
#> 96   8.94343531 15.76575148   Rest
#> 97  14.09818970 14.19762546   Rest
#> 98  13.48958147  2.38273873   Rest
#> 99   5.63042149  5.66038706   Rest
#> 100  7.82158811 11.90074780   Rest
#> 101  1.24200672  1.72831094   Rest
#> 102  9.63778105  6.12778684   Rest
#> 103 12.60468064 12.87450420   Rest
#> 104  2.23572794  0.84211846   Rest
#> 105 12.88708164 11.07870521   Rest
#> 106 12.32634688  6.81734762   Rest
#> 107  7.15733274  1.90967596   Rest
#> 108 11.63636247 14.41534222   Rest
#> 109  0.91983840  9.78046837   Rest
#> 110  5.03375968 13.63197240   Rest
#> 111 12.73057744  6.23476101   Rest
#> 112 13.00842269  3.33974082   Rest
#> 113 15.07935468 14.79810330   Rest
#> 114  6.57975298  5.68800942   Rest
#> 115  5.22221875  2.95848311   Rest
#> 116  6.45764673  4.39193813   Rest
#> 117  0.27299387 12.51192957   Rest
#> 118  4.61663668  2.32689139   Rest
#> 119  3.60685463  8.84209161   Rest
#> 120 11.32722366 12.91058986   Rest
#> 121  2.32952750 13.20657556   Rest
#> 122  1.17549286  5.64829831   Rest
#> 123  8.90933507 12.34798750   Rest
#> 124  6.12163611 10.81379579   Rest
#> 125 -0.81238669  1.64808406   Rest
#> 126  4.99289275  5.38969579   Rest
#> 127 13.94836132  9.04484837   Rest
#> 128  6.10415048 11.48492365   Rest
#> 129  8.71281167 -0.79454413   Rest
#> 130 14.62365821  6.30175660   Rest
#> 131  9.51591908  9.44905317   Rest
#> 132  8.16444215 12.38906578   Rest
#> 133 11.66104733  8.57199312   Rest
#> 134  3.48111028  1.29884470   Rest
#> 135 14.48824700  9.62581552   Rest
#> 136  6.78676503 14.09071311   Rest
#> 137  9.94614102  1.93329225   Rest
#> 138  1.98195623  9.62417632   Rest
#> 139  3.21195062  6.24627449   Rest
#> 140 14.21833125  7.84415644   Rest
#> 141 12.37610127  8.27313792   Rest
#> 142  7.33864017 13.96980840   Rest
#> 143  5.38809826  0.84981651   Rest
#> 144 11.04099846 12.65290782   Rest
#> 145 12.68531154 11.98410922   Rest
#> 146 16.44457438  6.86892502   Rest
#> 147  9.23935094 11.14940261   Rest
#> 148 13.21601192  7.03495306   Rest
#> 149  1.68586130 14.49876909   Rest
#> 150  7.31313976 13.88861189   Rest
#> 151  6.40805991  9.39685966   Rest
#> 152 11.42426077 11.46907283   Rest
#> 153  8.94881459 15.43973526   Rest
#> 154 10.74574454 12.17680496   Rest
#> 155  1.07597056  5.74912964   Rest
#> 156 13.19317670 14.11790522   Rest
#> 157 14.34608465  4.11669868   Rest
#> 158  1.78908422  3.55138643   Rest
#> 159  2.34829099 11.39189382   Rest
#> 160  9.58005757 13.61622757   Rest
#> 161  8.52010563 15.01385298   Rest
#> 162  2.73233485 11.53782490   Rest
#> 163  5.00528612  2.46902295   Rest
#> 164  4.38485077  8.35255069   Rest
#> 165  0.09777378  4.30093203   Rest
#> 166  0.58843925 10.14781915   Rest
#> 167 13.38972625 13.54298411   Rest
#> 168  6.48043488 10.26570494   Rest
#> 169  1.57661826  4.29078089   Rest
#> 170  7.23335282 12.62857518   Rest
#> 171 11.06782996  4.26975047   Rest
#> 172 13.68097398 11.18294801   Rest
#> 173  9.31799610  4.25584222   Rest
#> 174  5.03511725 11.82608203   Rest
#> 175  6.06945214 15.25833871   Rest
#> 176  7.59704899 12.72783978   Rest
#> 177  1.42040505  4.84974945   Rest
#> 178  4.47245245  1.57759313   Rest
#> 179 14.32799765  7.08885179   Rest
#> 180 13.10930989 14.96423455   Rest
#> 181  6.28419753  1.13002038   Rest
#> 182  9.49770928 16.29625634   Rest
#> 183  4.57948757 11.06584262   Rest
#> 184  7.35287137 11.56197964   Rest
#> 185  5.75336888  4.94551330   Rest
#> 186 14.67563233  4.79211140   Rest
#> 187 14.19628888  2.26816519   Rest
#> 188  5.97854471 14.89989973   Rest
#> 189  0.10495474  9.35000009   Rest
#> 190 15.71077499  0.66424466   Rest
#> 191  9.93924544 14.57583563   Rest
#> 192 10.93532027 11.94715337   Rest
#> 193  1.31038019 14.69160750   Rest
#> 194 11.81049859  3.66430773   Rest
#> 195  3.95781550 11.31310999   Rest
#> 196  5.59379100 12.75390876   Rest
#> 197 14.30649850  7.78012055   Rest
#> 198  4.64292803 15.39671728   Rest
#> 199 11.09816147  3.57283734   Rest
#> 200  5.37351192 12.90692028   Rest
#> 201  2.81803722  8.81949471   Rest
#> 202  6.05929908 10.72581565   Rest
#> 203  9.81470006 11.67216058   Rest
#> 204 14.87624396  6.65794698   Rest
#> 205  0.82729524  7.60591652   Rest
#> 206  8.81887939  5.85070006   Rest
#> 207  4.70735879 11.50041712   Rest
#> 208  9.79504943  5.90307886   Rest
#> 209 13.92408440  2.57762046   Rest
#> 210  8.75851798  3.80170852   Rest
#> 211  5.80515639 13.77354813   Rest
#> 212  3.69245755 11.00998101   Rest
#> 213  4.82842908  8.08431379   Rest
#> 214 16.69413751  6.19904009   Rest
#> 215  9.66223136 15.74367255   Rest
#> 216  2.67888361 12.16792784   Rest
#> 217  1.75281699  7.74543579   Rest
#> 218 13.59610302  6.07789859   Rest
#> 219 14.58434697  1.22725731   Rest
#> 220 -1.28954768  9.35717590   Rest
#> 221  8.06287812 13.95853139   Rest
#> 222  7.70779618 -0.20519973   Rest
#> 223  3.98036087  2.80228829   Rest
#> 224  8.45437187  4.73702623   Rest
#> 225  1.94391334  2.40814800   Rest
#> 226  9.81562281  4.44001662   Rest
#> 227 10.14386653  2.92936759   Rest
#> 228  6.43842928  4.66988974   Rest
#> 229 15.35824220  3.51702489   Rest
#> 230  4.26578734  1.81813458   Rest
#> 231  4.79387067 18.37574166   Rest
#> 232 12.93017196  9.44484440   Rest
#> 233  3.51812723 12.18763125   Rest
#> 234  8.55859497  3.86623252   Rest
#> 235 14.22310675  9.71481851   Rest
#> 236  4.28245393 10.67159150   Rest
#> 237  8.14336598 12.63083777   Rest
#> 238 13.98062536 13.61488902   Rest
#> 239  3.18261128 -1.94738402   Rest
#> 240  7.84705309 -1.96112862   Rest
#> 241 12.91305705 10.15607386   Rest
#> 242  6.46917667  2.50845826   Rest
#> 243 14.09005780  5.69467257   Rest
#> 244  2.49952528  3.63524234   Rest
#> 245  8.52561704  2.11514322   Rest
#> 246 12.80343682  3.59403743   Rest
#> 247 13.01526824  9.74727848   Rest
#> 248  5.65331607  4.71879498   Rest
#> 249  5.21834823  7.61732516   Rest
#> 250  4.32117474  8.04585362   Rest
#> 251 14.11278378 11.53277580   Rest
#> 252  2.08624714 11.43158308   Rest
#> 253  4.36932504  8.26329990   Rest
#> 254  7.10149206  1.70169235   Rest
#> 255 -0.07383601  4.51873748   Rest
#> 256 15.62664382  2.63073350   Rest
#> 257  5.20170818  9.84857489   Rest
#> 258  7.52134794  0.85680813   Rest
#> 259 12.21150911 12.67507731   Rest
#> 260  3.29066986 14.92959169   Rest
#> 261 15.56845681 13.84098249   Rest
#> 262  7.06497590 15.78606733   Rest
#> 263  6.94095840  8.77719549   Rest
#> 264  6.96782314 11.56567188   Rest
#> 265  3.71961858  1.73339830   Rest
#> 266  8.75119936  6.79704844   Rest
#> 267  3.99448956 11.84877505   Rest
#> 268 15.44851072  2.79914807   Rest
#> 269 11.87057342  1.38052962   Rest
#> 270  4.34313952 13.01669077   Rest
#> 271  8.52570719  8.42575643 Circle
#> 272  9.64979548  8.65010953 Circle
#> 273  8.08518237  6.65801567 Circle
#> 274  6.37292756  7.23919643 Circle
#> 275  7.43501189  9.25522864 Circle
#> 276 10.61953762  6.74905005 Circle
#> 277  6.83223809  9.82272246 Circle
#> 278 10.36605746  8.05102369 Circle
#> 279  5.17728833  8.02072349 Circle
#> 280  8.94768396  9.90016126 Circle
#> 281  8.56994731  5.25738649 Circle
#> 282  8.83354325  6.36969144 Circle
#> 283  5.58541957  6.97266582 Circle
#> 284  6.55677767  7.14347896 Circle
#> 285  7.41319363 10.90362970 Circle
#> 286  6.65708954  8.85302219 Circle
#> 287 10.74921193  7.98306276 Circle
#> 288  6.91673402  9.80589899 Circle
#> 289  9.30879891  8.24052099 Circle
#> 290  6.92945059  9.88037925 Circle
#> 291  8.36620900  5.93993510 Circle
#> 292  7.53523213  5.38001335 Circle
#> 293  9.47123098  8.93725358 Circle
#> 294  7.40808174  5.78320733 Circle
#> 295  5.19746671  9.10239753 Circle
#> 296  9.21909295 10.01335354 Circle
#> 297 11.09372096  6.61416645 Circle
#> 298  8.48897988  6.12268796 Circle
#> 299  8.49429174  9.07123013 Circle
#> 300 10.69976744 11.15551681 Circle
#> 301  4.66750617  9.04574240 Circle
#> 302  5.30395492  7.39124720 Circle
#> 303  7.32681853  6.66612872 Circle
#> 304  7.40876873 10.10143537 Circle
#> 305  5.92362562  8.45057263 Circle
#> 306  7.21728243  5.39866984 Circle
#> 307  5.67739007  6.49482385 Circle
#> 308  9.71140306  7.09605977 Circle
#> 309  8.64829710  8.59060853 Circle
#> 310  7.13527583 10.22254505 Circle
#> 311  9.15705678  9.59703740 Circle
#> 312  5.97302480  9.49128815 Circle
#> 313  6.48970925  8.84010247 Circle
#> 314  9.37393410  8.72045440 Circle
#> 315  9.54226525 10.11177707 Circle
#> 316  7.81417225  8.42024875 Circle
#> 317  7.17063363  5.46277801 Circle
#> 318  9.13023131  9.94381818 Circle
#> 319  8.12474897  6.84971716 Circle
#> 320  6.29458075  9.82788295 Circle
#> 321  4.24727073  7.26549987 Circle
#> 322  6.14122784 10.54856931 Circle
#> 323  7.16578931  9.87062099 Circle
#> 324  8.59698948  5.28825709 Circle
#> 325  7.95955437  5.10951701 Circle
#> 326  7.97064078  8.04653837 Circle
#> 327  7.19660919 10.37398266 Circle
#> 328  9.76384332 10.78125438 Circle
#> 329  9.51669898  6.92981646 Circle
#> 330  7.53991451  7.41775620 Circle
#> 331  9.31962599  8.09506771 Circle
#> 332  8.37142363 10.70459763 Circle
#> 333  4.67254069  7.50286471 Circle
#> 334  7.73137658  7.79845884 Circle
#> 335  6.72481623  7.93995262 Circle
#> 336  8.91635658 10.28689307 Circle
#> 337  8.56695490  8.91484874 Circle
#> 338  9.57639818  7.77999088 Circle
#> 339  8.39002744  6.66543395 Circle
#> 340  4.83714376  8.12186493 Circle
#> 341  9.48166096  9.96452414 Circle
#> 342  6.32414914  8.38739794 Circle
#> 343  5.78655228  7.49382329 Circle
#> 344  7.88952569  8.35425909 Circle
#> 345  6.17760916  7.96968503 Circle
#> 346  9.39016331 10.06211279 Circle
#> 347 10.23416862  8.63224489 Circle
#> 348  5.83243923  7.90192803 Circle
#> 349 10.05774438  6.84235578 Circle
#> 350  9.05489916  9.04944010 Circle
#> 351  6.08432945  8.60312721 Circle
#> 352  9.86331875  7.91038841 Circle
#> 353  6.83128640  8.31919782 Circle
#> 354  5.79784331  9.27567903 Circle
#> 355  7.73749065  8.66261574 Circle
#> 356  7.90816332 10.87731289 Circle
#> 357  8.60127180  6.86716140 Circle
#> 358  8.03408892  4.25834594 Circle
#> 359 10.38985010  6.62404264 Circle
#> 360  9.80333687  7.00573838 Circle
#> 361  8.28351257  6.78310202 Circle
#> 362  7.93194753 10.47675272 Circle
#> 363  5.61670987  7.59230548 Circle
#> 364  6.84039468  6.72789527 Circle
#> 365  7.29329797  9.76419097 Circle
#> 366  6.24043365  8.54025588 Circle
#> 367  7.42284315  7.83416117 Circle
#> 368  9.90939707 11.13789696 Circle
#> 369  6.32257801  9.69524097 Circle
#> 370  8.09933332  6.90312090 Circle
#> 371  7.76762847  9.08588737 Circle
#> 372  7.13794165  9.78392201 Circle
#> 373  5.99136587  6.67691202 Circle
#> 374 10.06006608  7.04567875 Circle
#> 375  8.65121341  6.43061613 Circle
#> 376 11.35253743  7.33446916 Circle
#> 377  8.47624842  6.48437304 Circle
#> 378  8.81910489  6.19910463 Circle
#> 379  7.52105007  5.22996566 Circle
#> 380  7.11297560  8.66077550 Circle
#> 381 10.14639377  9.02635188 Circle
#> 382  5.69742484  8.34139750 Circle
#> 383  9.65121222  9.91351895 Circle
#> 384  5.74227347  9.61673112 Circle
#> 385  7.50726450  9.61039688 Circle
#> 386  5.74572918  7.80697198 Circle
#> 387  6.63043924  5.90946481 Circle
#> 388  9.69454480  8.27321239 Circle
#> 389  8.06696749  8.93067045 Circle
#> 390  8.83196688 10.49067189 Circle
#> 391 10.85563400  8.18164743 Circle
#> 392  9.16275504 10.41625254 Circle
#> 393  7.04526831  6.90087819 Circle
#> 394  8.14057650  8.03414694 Circle
#> 395  4.66507000  8.22550906 Circle
#> 396  7.37291078  7.42159766 Circle
#> 397  6.22130231  8.49840792 Circle
#> 398  5.73143572  8.53637304 Circle
#> 399 10.54821983  8.65017362 Circle
#> 400  4.64927486  6.48722810 Circle
#> 401  7.85243074  4.31188355 Circle
#> 402  6.67732928  7.91754902 Circle
#> 403  6.71727627  7.09500914 Circle
#> 404 10.63876828  8.13256867 Circle
#> 405  8.10133976  8.73055134 Circle
#> 406  6.05009752  9.63581820 Circle
#> 407  8.08890794  8.71095795 Circle
#> 408 10.19363113  9.17572974 Circle
#> 409  7.83293384  9.61022541 Circle
#> 410  8.18926793  6.96368477 Circle
#> 411  6.30517266  7.38113572 Circle
#> 412  6.45485020  7.77629753 Circle
#> 413  7.77613610  8.51661546 Circle
#> 414  8.19939893  6.94511116 Circle
#> 415  8.98820006 10.25883595 Circle
#> 416  5.69314825  7.62563287 Circle
#> 417  5.99463227  8.59468265 Circle
#> 418  6.85958890  6.89771424 Circle
#> 419  8.89108661 10.05388376 Circle
#> 420  8.42971686  9.60122948 Circle
#> 421  8.72058306  7.05555781 Circle
#> 422 10.76444634  7.63216156 Circle
#> 423  6.59203769  9.38965243 Circle
#> 424  4.99507708  5.63542557 Circle
#> 425  6.77678490  8.43963132 Circle
#> 426  5.15323094  5.09675846 Circle
#> 427 10.11088964  6.13460345 Circle
#> 428  8.96862135 10.02013455 Circle
#> 429  8.24582819  7.93031541 Circle
#> 430  5.23500544  7.36647403 Circle
#> 431 10.31066036  7.88051694 Circle
#> 432  5.77040039 10.00360440 Circle
#> 433  7.27273986  6.21620341 Circle
#> 434 10.04859257  9.15576169 Circle
#> 435  8.44044693  8.93497852 Circle
#> 436  7.61409498  8.73726898 Circle
#> 437  9.25812805  7.37270959 Circle
#> 438  8.74302373  6.96493574 Circle
#> 439  7.23469541 10.53632786 Circle
#> 440  7.04991901  6.69101305 Circle
#> 441 10.44920041  9.77140434 Circle
#> 442  9.49632586  7.62603734 Circle
#> 443  6.90154886  6.62845914 Circle
#> 444  9.33498557  8.55301776 Circle
#> 445  9.44306235 10.28277345 Circle
#> 446  7.08445108  6.15935080 Circle
#> 447  7.30906992  9.27070724 Circle
#> 448 10.78920044  9.92291602 Circle
#> 449  8.09540595  8.47032344 Circle
#> 450  8.31407620 10.32552176 Circle
#> 451  9.87596802  7.31270508 Circle
#> 452  9.15135829  5.93339988 Circle
#> 453  8.57947170  9.53874886 Circle
#> 454  9.78452339  8.26399695 Circle
#> 455  9.83618515  7.98638556 Circle
#> 456  5.73580443  8.35917177 Circle
#> 457  6.62069749  6.56866458 Circle
#> 458  9.18238842 10.07319986 Circle
#> 459  6.54527960  8.55225829 Circle
#> 460  5.29751346  7.97986096 Circle
#> 461  7.05331653  5.90583610 Circle
#> 462  6.56778642  6.33246148 Circle
#> 463  8.74502929  8.99620602 Circle
#> 464  5.36293959  6.86287162 Circle
#> 465  5.68109798  8.20903627 Circle
#> 466  6.50137693 10.62400942 Circle
#> 467  7.02801847  6.64678928 Circle
#> 468  8.18431605  7.96204479 Circle
#> 469  8.36412196  6.83276822 Circle
#> 470  5.45580848  7.67098922 Circle
#> 471  8.95885331  7.82980935 Circle
#> 472 10.25800066  9.20334308 Circle
#> 473  9.77744643 10.22908481 Circle
#> 474  9.42733472  8.72314093 Circle
#> 475 10.19951453  6.03093100 Circle
#> 476  7.03973573  8.41649363 Circle
#> 477  6.40601783  6.26578418 Circle
#> 478 10.33961464 10.12259480 Circle
#> 479  5.35837335  7.42203559 Circle
#> 480  9.69073499 11.31747335 Circle
#> 481  7.60329618  5.32239677 Circle
#> 482  7.15111263 10.60251320 Circle
#> 483  6.52552649  8.46919514 Circle
#> 484 10.33857919  7.39379158 Circle
#> 485  8.97041496  7.48861048 Circle
#> 486 10.12748856 10.89194714 Circle
#> 487  6.36277177  8.11848063 Circle
#> 488  7.04584819  8.41578340 Circle
#> 489  7.54098222 10.09705947 Circle
#> 490  8.00202849  9.04953140 Circle
#> 491  8.72533958  6.08678194 Circle
#> 492  5.53152978  7.09668488 Circle
#> 493  6.94863293  8.29921756 Circle
#> 494 10.57935823  8.23801934 Circle
#> 495  7.73990233 10.30580478 Circle
#> 496  5.99875968  8.04496949 Circle
#> 497  7.66083355  5.57489598 Circle
#> 498  5.85823257  9.59353120 Circle
#> 499 11.32743978  8.37385518 Circle
#> 500  5.64639109  5.85165279 Circle
#> 501  8.59579975  7.61101565 Circle
#> 502  7.72457344  7.15179679 Circle
#> 503  8.62631194  9.45667610 Circle
#> 504  4.86467524  9.20151389 Circle
#> 505 10.41044619  6.07553860 Circle
#> 506  6.17993412  8.36590302 Circle
#> 507  9.96811053  6.63646493 Circle
#> 508  5.18954106  7.27702242 Circle
#> 509  7.74172532  9.11996924 Circle
#> 510 10.78336048  5.19922339 Circle
#> 511  8.59009627  9.71354222 Circle
#> 512  7.06765112  6.65634714 Circle
#> 513  8.01296188  9.76370490 Circle
#> 514  8.89531378  8.70292378 Circle
#> 515  7.04780018  9.24089049 Circle
#> 516  9.56118414 10.13785158 Circle
#> 517  8.84859756 10.02060867 Circle
#> 518  7.26397175  9.96055032 Circle
#> 519  6.38282609  7.53280748 Circle
#> 520  9.33010821  7.11789151 Circle
#> 521  6.45774866  7.68877294 Circle
#> 522 10.61479701 11.05387737 Circle
#> 523  5.38068234  7.56924252 Circle
#> 524  6.46588157  8.50068597 Circle
#> 525  9.82762140  6.34305166 Circle
#> 526  9.28362844 10.71051589 Circle
#> 527  6.46259901  7.83147214 Circle
#> 528  9.68969623  9.71082974 Circle
#> 529  5.69786533  7.94292629 Circle
#> 530  8.20396078 10.52841542 Circle
#> 531  8.70579474  7.36866438 Circle
#> 532  5.21990994  7.84451615 Circle
#> 533  8.90367306  6.37933221 Circle
#> 534  6.44836042  8.96912883 Circle
#> 535  7.57919034  6.14099512 Circle
#> 536  7.95446221  7.31792617 Circle
#> 537 11.49050796  8.25107267 Circle
#> 538  5.48272226  8.34977362 Circle
#> 539  5.73455448  6.61061419 Circle
#> 540  8.26793367  9.34732085 Circle
#> 541  5.99386198  8.06004074 Circle
#> 542  8.35170205  8.04879115 Circle
#> 543  5.18106630  7.55088381 Circle
#> 544  5.63871565  6.61506729 Circle
#> 545  7.20713774  9.43052018 Circle
#> 546  7.67649950  6.04778358 Circle
#> 547  6.71896604  5.88426404 Circle