ROSE Algorithm

A thin wrapper around ROSE::ROSE() that generates a synthetic balanced sample by enlarging the feature space of minority and majority class examples.

Usage

rose(
  df,
  var,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1
)

Arguments

df

A data.frame or tibble. Must have 1 factor variable with exactly 2 levels and remaining numeric variables.

var

Character, name of variable containing the 2-level factor variable.

over_ratio

A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows as the majority level.

minority_prop

A numeric. Determines the of over-sampling of the minority class. Defaults to 0.5.

minority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the minority class. Defaults to 1.

majority_smoothness

A numeric. Shrink factor to be multiplied by the smoothing parameters to estimate the conditional kernel density of the majority class. Defaults to 1.

Value

A data.frame or tibble, depending on type of df.

Details

This function is a thin wrapper around ROSE::ROSE(). For full details on the underlying implementation, see that function's documentation.

The factor variable used to balance around must only have 2 levels.

The ROSE algorithm works by selecting an observation belonging to class k and generates new examples in its neighborhood as determined by some matrix H_k. Smaller values of minority_smoothness and majority_smoothness have the effect of shrinking the entries of the corresponding smoothing matrix H_k. Shrinking would be a cautious choice if there is a concern that excessively large neighborhoods could lead to blurring the boundaries between the regions of the feature space associated with each class.

References

Lunardon, N., Menardi, G., and Torelli, N. (2014). ROSE: a Package for Binary Imbalanced Learning. R Jorunal, 6:82–92.

Menardi, G. and Torelli, N. (2014). Training and assessing classification rules with imbalanced data. Data Mining and Knowledge Discovery, 28:92–122.

See also

Examples

rose(circle_example[, c("x", "y", "class")], var = "class")
#>               x          y  class
#> 1   13.08172463  9.0034688   Rest
#> 2    2.47460001  2.3346730   Rest
#> 3    5.01289558 13.4216909   Rest
#> 4   13.85042660  7.5337545   Rest
#> 5   13.89639349  9.6648430   Rest
#> 6    6.93230224 14.5884874   Rest
#> 7    6.70689041 14.4942610   Rest
#> 8    4.34875512  2.4400285   Rest
#> 9    5.36598962  4.7790644   Rest
#> 10  14.95244541  9.7633019   Rest
#> 11   5.34993370 12.4242527   Rest
#> 12   6.48414387  2.9525913   Rest
#> 13  14.68002855  2.7732020   Rest
#> 14   8.09863272 13.6220936   Rest
#> 15   2.99493260  7.9742031   Rest
#> 16  14.63226882  1.8173572   Rest
#> 17  11.11195153 12.1477861   Rest
#> 18   3.35635880  4.4541817   Rest
#> 19   6.55543627  0.8852976   Rest
#> 20  12.23681685 13.5818523   Rest
#> 21   5.15036190 13.1447995   Rest
#> 22   5.11212285 12.2458546   Rest
#> 23  12.03186295  9.2738622   Rest
#> 24   3.39702117 10.4601064   Rest
#> 25   8.26089480  2.5440225   Rest
#> 26   2.19285939 11.4768084   Rest
#> 27   5.65679992  6.7180883   Rest
#> 28  -0.17901451  2.0818453   Rest
#> 29   0.96408917 12.2364937   Rest
#> 30   0.59301058  4.2378522   Rest
#> 31   0.97162998  1.9700397   Rest
#> 32   2.66108076 15.5378456   Rest
#> 33   3.63834260  3.7023108   Rest
#> 34   5.66739836  8.4784556   Rest
#> 35  14.84541641 10.6213331   Rest
#> 36   6.66878929 11.9723719   Rest
#> 37  10.14318394  7.6582415   Rest
#> 38  10.95184502 13.2398300   Rest
#> 39   7.47916433  2.9071641   Rest
#> 40   6.58926979  2.2518449   Rest
#> 41  12.36081665  2.0473490   Rest
#> 42   1.93242905  2.5913521   Rest
#> 43   5.01777136  6.4271916   Rest
#> 44  11.79241814  5.6344555   Rest
#> 45   3.99871826  9.6266593   Rest
#> 46   3.70988092  9.1368105   Rest
#> 47  10.41645232 13.8322918   Rest
#> 48  13.18351765 11.7070244   Rest
#> 49   0.23244124  4.0667210   Rest
#> 50  10.35233806  8.9579520   Rest
#> 51  13.86082366  4.6200438   Rest
#> 52   5.14447383 15.3116222   Rest
#> 53  13.51583017 12.6632251   Rest
#> 54  11.15488071 13.0284765   Rest
#> 55   5.40119007  1.2956755   Rest
#> 56   1.31615017  2.5797530   Rest
#> 57  12.17670693  3.2133675   Rest
#> 58   7.37906625 16.9454943   Rest
#> 59   4.09668546  6.1436798   Rest
#> 60  10.30601601 17.9446996   Rest
#> 61  11.38743746 11.5840027   Rest
#> 62   4.40715467  3.1912019   Rest
#> 63   6.15774830  1.4446085   Rest
#> 64  10.69311038  3.3101885   Rest
#> 65   7.37965849 13.1006180   Rest
#> 66   6.28582980 13.6810305   Rest
#> 67  11.56290363  2.9547931   Rest
#> 68  14.88365057 11.0877280   Rest
#> 69  13.70027536  4.1561313   Rest
#> 70   9.46380439 10.1873227   Rest
#> 71  15.35725730 15.3472018   Rest
#> 72  11.78458857 14.4242270   Rest
#> 73   5.37960630  3.3965164   Rest
#> 74   0.04980465 13.5778353   Rest
#> 75   4.77942748  6.9205513   Rest
#> 76   6.84847451 17.0601851   Rest
#> 77   1.00339023  1.4871527   Rest
#> 78  10.00493089  2.1924143   Rest
#> 79   4.98815116 10.1983950   Rest
#> 80   4.54899679 10.4367794   Rest
#> 81   5.59673571  5.3085170   Rest
#> 82   1.01954053  8.6892133   Rest
#> 83   5.44507938  2.1734735   Rest
#> 84  10.79453263  1.2610927   Rest
#> 85  -0.16523586 14.2046420   Rest
#> 86   4.72303077  9.7612167   Rest
#> 87   5.67150451 12.5446606   Rest
#> 88  16.47278365  4.5384900   Rest
#> 89   9.50078315 12.9480714   Rest
#> 90  13.94350120  9.8937277   Rest
#> 91  15.23292084  9.5714188   Rest
#> 92   6.17912907  8.5757333   Rest
#> 93  11.33764781 12.7031713   Rest
#> 94  10.46472677  4.7634942   Rest
#> 95   4.20110001 10.3928788   Rest
#> 96   2.81270667 15.0147068   Rest
#> 97   0.49751248 12.5918184   Rest
#> 98  12.40940131  4.3233542   Rest
#> 99  12.73811830  7.0629058   Rest
#> 100  6.93946273  9.6671086   Rest
#> 101  4.23094711 13.7389327   Rest
#> 102  6.33722323  2.2593598   Rest
#> 103  4.60069933 -0.2605124   Rest
#> 104 15.33398352  1.1155111   Rest
#> 105 11.38469916 13.3559068   Rest
#> 106  9.73273880  7.8568950   Rest
#> 107 11.87826866 12.9628610   Rest
#> 108 11.39344873  8.0790024   Rest
#> 109  0.26148198  4.4653254   Rest
#> 110  1.89424551  2.1340391   Rest
#> 111  9.67748448 14.5747410   Rest
#> 112 13.72284486  1.0364767   Rest
#> 113  7.58558792 14.8119705   Rest
#> 114 10.36914306  9.7586688   Rest
#> 115  6.42984747 14.1439435   Rest
#> 116  7.87306927  6.2483394   Rest
#> 117  4.35839522 10.3292732   Rest
#> 118  7.92991179 15.3295171   Rest
#> 119  2.45835560  6.9383301   Rest
#> 120  9.57204193 15.8468536   Rest
#> 121  7.82170606 13.0467139   Rest
#> 122  3.82276458  0.8159718   Rest
#> 123  9.61910426 -0.5612103   Rest
#> 124 11.55563880  1.6450444   Rest
#> 125  5.60494609  1.7362104   Rest
#> 126  4.06537175  3.3707115   Rest
#> 127  2.12259562  4.1288296   Rest
#> 128 11.55922241 20.3532716   Rest
#> 129  5.07114328 12.6712775   Rest
#> 130  5.41703754 13.8028346   Rest
#> 131 18.57731377  4.2500274   Rest
#> 132  6.94441719  5.0009413   Rest
#> 133 14.83032311  1.5471473   Rest
#> 134 14.90143671  9.2960075   Rest
#> 135  1.20523669  0.1387226   Rest
#> 136 12.79695588 13.5389430   Rest
#> 137  4.56543967 14.6323366   Rest
#> 138  1.61688871  4.6399198   Rest
#> 139  4.00808454  4.6930005   Rest
#> 140  3.52302715  5.0054702   Rest
#> 141 12.33080257  8.1857984   Rest
#> 142  2.64004120  6.6648118   Rest
#> 143 13.50648770  3.6104671   Rest
#> 144  6.90239425  5.6984721   Rest
#> 145 14.49503268  8.8967220   Rest
#> 146  2.02837044 13.3479460   Rest
#> 147  2.80653414  2.2263625   Rest
#> 148  6.07591127 11.7815752   Rest
#> 149 -0.07639280  2.9935512   Rest
#> 150 13.91850592  8.2815711   Rest
#> 151 13.88150179 11.5592062   Rest
#> 152  6.69578355  3.0943044   Rest
#> 153  6.33640846  1.2062775   Rest
#> 154  7.76035789 10.6375013   Rest
#> 155  0.75350112  7.7310681   Rest
#> 156  0.71494650  0.7790829   Rest
#> 157 11.04374003 15.4387641   Rest
#> 158  3.17611215 10.4456853   Rest
#> 159  1.71398346  1.7898877   Rest
#> 160  3.71126033  3.7921257   Rest
#> 161  8.17731370  2.2771224   Rest
#> 162  2.13875558 13.4383373   Rest
#> 163  9.14840662  4.6838562   Rest
#> 164  1.40520008  2.6444408   Rest
#> 165 15.91361225 11.1031131   Rest
#> 166  7.37514700 11.6927712   Rest
#> 167 17.23057051  2.8292179   Rest
#> 168 12.63511020  8.4649658   Rest
#> 169  6.37247873  3.8918162   Rest
#> 170  5.60283511  5.6458595   Rest
#> 171 10.13614041 10.4926652   Rest
#> 172 10.54405310 11.7632662   Rest
#> 173 12.86923175 11.1646785   Rest
#> 174  2.14388363 14.1747394   Rest
#> 175  2.22039787  9.3629974   Rest
#> 176 11.04785067  7.8749336   Rest
#> 177 11.71447884 17.0930661   Rest
#> 178 12.96433710  5.4989294   Rest
#> 179  7.91517025 10.1793066   Rest
#> 180  6.68189682 12.3412243   Rest
#> 181  7.32838012  1.0103109   Rest
#> 182  5.51945591  6.1578399   Rest
#> 183  8.57622396 14.5877268   Rest
#> 184  9.19876977  3.8483216   Rest
#> 185  5.25070838 14.8061776   Rest
#> 186  8.73497875 12.4232683   Rest
#> 187  2.95557613 15.0785041   Rest
#> 188  0.46415036  4.0619399   Rest
#> 189 10.46898620 10.5179323   Rest
#> 190 10.39406722 12.1678221   Rest
#> 191 12.02882363  8.7506667   Rest
#> 192 10.17116038  2.0855498   Rest
#> 193  5.20772894 13.9488110   Rest
#> 194  6.84761578  0.4621599   Rest
#> 195  5.15452466 13.3684291   Rest
#> 196  4.69501784  2.5961687   Rest
#> 197  2.12421634  8.2963276   Rest
#> 198  3.09488549 13.4748130   Rest
#> 199  4.38988587  0.7081768   Rest
#> 200 10.11752450  9.8759200   Rest
#> 201  7.88285127 13.8798335   Rest
#> 202  4.83929460  3.1190833   Rest
#> 203 11.05734454 15.9301368   Rest
#> 204 13.82348315  3.7315712   Rest
#> 205 10.77393717 11.6005174   Rest
#> 206  7.47047751 11.6683702   Rest
#> 207  7.52258314 13.7794884   Rest
#> 208  3.81226715  4.2664352   Rest
#> 209 13.33990665  5.0649035   Rest
#> 210 11.28822828 11.9934125   Rest
#> 211 13.61277418 13.2478742   Rest
#> 212 10.03598975 -1.0204460   Rest
#> 213 10.77062848 11.6715691   Rest
#> 214 10.53142175  2.1550236   Rest
#> 215 14.88052114  4.8436979   Rest
#> 216  5.59931445 10.4439770   Rest
#> 217  3.52150571  5.2359385   Rest
#> 218 11.33156262  9.1485034   Rest
#> 219  9.05747146 11.4925564   Rest
#> 220 13.45217955  7.0237251   Rest
#> 221  0.58031201  9.8493233   Rest
#> 222  9.38395842  1.6660489   Rest
#> 223 13.32344325  6.5022496   Rest
#> 224  7.87011973 10.5855191   Rest
#> 225  6.66593426 12.4238082   Rest
#> 226  4.29074969 14.8666098   Rest
#> 227  3.62712329 15.7753727   Rest
#> 228 12.87366157 12.0298581   Rest
#> 229  6.71933909 12.0399127   Rest
#> 230  1.66925890  7.9971426   Rest
#> 231 11.10339961 13.7564817   Rest
#> 232 11.93067266  7.7017232   Rest
#> 233  4.85364289 11.3763972   Rest
#> 234  6.08205139 16.0230065   Rest
#> 235 10.40358654 12.4823478   Rest
#> 236  5.99751978  4.7052056   Rest
#> 237 15.86846681 14.0524192   Rest
#> 238  1.78411987  7.0055451   Rest
#> 239  8.94649561  7.2847636   Rest
#> 240  6.66598862  6.2916671   Rest
#> 241 12.38515301  5.9856303   Rest
#> 242  7.75840043 15.5845589   Rest
#> 243  5.32749595 16.1175841   Rest
#> 244 14.42475314 12.5309754   Rest
#> 245 13.46413186  9.6063541   Rest
#> 246  4.74319703  3.7313916   Rest
#> 247  6.48164737  3.5089772   Rest
#> 248  2.31381030  7.9197762   Rest
#> 249  1.32031561 10.9140296   Rest
#> 250 15.86729033 12.2034726   Rest
#> 251  7.93261139 10.7068705   Rest
#> 252  8.04517088 14.4679686   Rest
#> 253  8.59558947  3.7605440   Rest
#> 254  7.98338303  5.0183293   Rest
#> 255 11.00817444 11.1594239   Rest
#> 256 12.25981483 12.0274299   Rest
#> 257 11.80958629 15.1087172   Rest
#> 258 11.89804170  4.5289603   Rest
#> 259  8.58916494 13.7117356   Rest
#> 260  5.32816153  0.9795712   Rest
#> 261  6.58843871 12.1135746   Rest
#> 262 12.66402907  5.4845014   Rest
#> 263  8.50231438 11.5514246   Rest
#> 264  1.85321792  4.6002817   Rest
#> 265  3.85594297  4.4411601   Rest
#> 266 14.51242449  2.5175129   Rest
#> 267 10.42614724 11.0903977   Rest
#> 268 14.19037753  6.3181580   Rest
#> 269  6.10825642  9.7256481   Rest
#> 270 13.46738024  3.6009824   Rest
#> 271  4.56049587 13.3329904   Rest
#> 272  2.17679406  1.2764453   Rest
#> 273  1.58182277  9.2581304   Rest
#> 274  1.71356803 13.4065310   Rest
#> 275 13.43691735  5.7028470   Rest
#> 276  6.11279375 10.9063490   Rest
#> 277  5.73818502 13.3709181   Rest
#> 278  2.85004165  9.2203347   Rest
#> 279  5.35956761 13.2498391   Rest
#> 280  5.30638582 11.8209022   Rest
#> 281 11.05910822  9.1458969   Rest
#> 282  7.67359531 12.6136616   Rest
#> 283 13.09865781 13.0448608   Rest
#> 284  0.50453836  2.4741875   Rest
#> 285  3.83999751  9.3931554   Rest
#> 286 10.11572261  1.1109550   Rest
#> 287  8.76386028 11.6267227   Rest
#> 288  2.50331474  9.2669270   Rest
#> 289  6.39989368 11.4137581   Rest
#> 290 11.35799515 12.0031408   Rest
#> 291  1.85859069  6.2430272   Rest
#> 292  2.59280840 12.2048474   Rest
#> 293 14.68980172  1.6222458   Rest
#> 294 14.99671925  7.1391295   Rest
#> 295  2.66348589  4.3941443   Rest
#> 296  5.96298189  2.2477893   Rest
#> 297  7.63441035  2.2806093   Rest
#> 298 12.81969414  3.0722462   Rest
#> 299  3.81667669 12.6844273   Rest
#> 300  2.93601811 13.9599228   Rest
#> 301 10.16281888  9.8612624   Rest
#> 302  3.92916434  8.3257404   Rest
#> 303 11.91825233  4.4708364   Rest
#> 304  7.46654671 13.1469317   Rest
#> 305  2.23356934  7.7241697   Rest
#> 306  9.25915294 11.6563553   Rest
#> 307  2.26355446 14.6856481   Rest
#> 308  4.60797794  9.3691897   Rest
#> 309 11.58195099 -0.4734698   Rest
#> 310  6.02389714 10.6959786   Rest
#> 311  2.78856420  6.1045425   Rest
#> 312  4.79155447  1.1208983   Rest
#> 313  7.07323246 12.9926393   Rest
#> 314  9.54484100 13.4193916   Rest
#> 315 10.44811019  9.3937234   Rest
#> 316 11.91375598  1.4778046   Rest
#> 317  3.00431861  1.8158376   Rest
#> 318  3.17842333  9.5919466   Rest
#> 319 11.04593934 12.9541323   Rest
#> 320  2.83305113  2.2336819   Rest
#> 321  4.12176769 10.0369642   Rest
#> 322  2.43613885  9.3601749   Rest
#> 323  2.64031076  7.8468566   Rest
#> 324  4.98302824  8.1370564   Rest
#> 325  1.66848089 15.4587365   Rest
#> 326 12.56696984  4.7666453   Rest
#> 327 13.20719401 -0.3890971   Rest
#> 328  3.21032518  3.5113408   Rest
#> 329  9.82554240 15.0662612   Rest
#> 330 10.77813156 13.9494023   Rest
#> 331  4.20570347 10.7695515   Rest
#> 332 10.64594602  8.0991815   Rest
#> 333  5.15159536  5.4036766   Rest
#> 334 13.46950936  9.8455501   Rest
#> 335  5.44242745 11.2050452   Rest
#> 336 11.35181392  3.3396972   Rest
#> 337 14.66580563 10.1509424   Rest
#> 338  1.09186648  4.0211416   Rest
#> 339  9.96544170  0.6412760   Rest
#> 340  7.60765015  0.6034569   Rest
#> 341 12.11507142  4.9056380   Rest
#> 342 13.93823755  2.3087477   Rest
#> 343 13.48125897 11.3162849   Rest
#> 344  8.28528483  8.4088226 Circle
#> 345 10.35010466  7.0416249 Circle
#> 346 10.38434404 10.3608676 Circle
#> 347  6.61868556 10.8717038 Circle
#> 348  6.28292731  6.4887182 Circle
#> 349  6.05803756  6.8718502 Circle
#> 350  7.66329492  6.0617184 Circle
#> 351  9.29394103  5.3222079 Circle
#> 352  8.91542960  9.4006377 Circle
#> 353  5.78586768  8.1673195 Circle
#> 354 10.87968700 10.7639663 Circle
#> 355  6.63377268  6.0741024 Circle
#> 356  7.80021680 10.1226182 Circle
#> 357 10.19056570 10.2224545 Circle
#> 358  6.05055257 10.0485329 Circle
#> 359  9.92252528 10.6989686 Circle
#> 360  8.64737574  9.5463822 Circle
#> 361  4.98078426  9.1064122 Circle
#> 362 10.79267425  6.3832403 Circle
#> 363  8.19046889  8.8812974 Circle
#> 364  6.97030071  6.9916532 Circle
#> 365  6.82715413  5.7027256 Circle
#> 366  6.49959024  9.6059334 Circle
#> 367  6.74238418  8.8858705 Circle
#> 368  5.34117609  7.3882493 Circle
#> 369  8.65048969  8.8450935 Circle
#> 370  6.41060767 10.7500039 Circle
#> 371  6.20250501  9.8710481 Circle
#> 372  9.50382260  8.9430858 Circle
#> 373  7.66268099  7.7869912 Circle
#> 374  5.73234058  5.8167261 Circle
#> 375  8.64988963  7.6034055 Circle
#> 376  7.46702069 10.0873507 Circle
#> 377  5.99391188  6.6321327 Circle
#> 378 10.07107894  8.7117076 Circle
#> 379 11.33933518  8.6284422 Circle
#> 380  6.32782574  6.6223697 Circle
#> 381  9.38063536  8.7157038 Circle
#> 382  5.56670243  9.8520940 Circle
#> 383  9.98890744  7.1660730 Circle
#> 384  8.14028870  6.1872181 Circle
#> 385  7.71317265  5.6599851 Circle
#> 386  6.99073532  9.8057084 Circle
#> 387  6.09847801  6.2493162 Circle
#> 388  6.75373823  8.4767832 Circle
#> 389  9.37423601 10.0829331 Circle
#> 390  7.24259289  9.7600733 Circle
#> 391  9.14699969  8.2117394 Circle
#> 392  8.27815792  6.0336658 Circle
#> 393  9.05424170 11.8476948 Circle
#> 394  8.53652485  8.0586296 Circle
#> 395  9.22269817  7.9779933 Circle
#> 396  7.70047350  4.5675913 Circle
#> 397  7.85868687  8.4516219 Circle
#> 398  8.86631014  8.3706228 Circle
#> 399  5.89333909  7.6792670 Circle
#> 400  7.21486808  7.5063662 Circle
#> 401  8.07218858  7.7690111 Circle
#> 402  7.80355080 10.3151010 Circle
#> 403  5.44130389  8.9808379 Circle
#> 404 10.01346443  6.4793940 Circle
#> 405  6.94561845 10.9998003 Circle
#> 406  7.49293653  8.9540705 Circle
#> 407  8.32857419  5.4429013 Circle
#> 408  5.58757528  8.7351482 Circle
#> 409  9.84369242  9.8483796 Circle
#> 410  9.13400026  6.8403155 Circle
#> 411  7.23668996  8.1726897 Circle
#> 412  6.67386749  8.5993461 Circle
#> 413  6.47994153  9.2222997 Circle
#> 414  8.64393668 10.1949374 Circle
#> 415  8.56095727  7.8634662 Circle
#> 416 10.74557618  8.5513962 Circle
#> 417  7.30744836  7.9504407 Circle
#> 418  8.75867489  9.8960397 Circle
#> 419  7.14371573  8.9119439 Circle
#> 420  7.41571950  6.9996496 Circle
#> 421  7.85491063  5.8283717 Circle
#> 422  6.73252235 10.9621948 Circle
#> 423  4.55512026  6.9710347 Circle
#> 424  8.96298693  9.5864947 Circle
#> 425  7.04865420  8.6723861 Circle
#> 426  6.98565096  8.0419950 Circle
#> 427  8.90982238  9.1056227 Circle
#> 428  8.24248475  6.6071843 Circle
#> 429 10.15934951  7.8863595 Circle
#> 430  9.89670069 10.0169343 Circle
#> 431  7.61433374 11.2242284 Circle
#> 432  7.44415054 10.0117151 Circle
#> 433  7.01470875 11.1996243 Circle
#> 434  6.86739238 10.5936957 Circle
#> 435  8.63043584 11.0295140 Circle
#> 436  4.76963045  8.2035887 Circle
#> 437  5.22212955  9.5387629 Circle
#> 438 10.67911240 11.1133427 Circle
#> 439  8.18344989  5.2301553 Circle
#> 440  9.38727716  9.1929808 Circle
#> 441  9.82346332  9.3316196 Circle
#> 442  6.38934237  8.6955589 Circle
#> 443  8.37137013  8.5469989 Circle
#> 444  5.56532151  8.4726220 Circle
#> 445  7.71924806  9.9725433 Circle
#> 446  9.68659603  7.6114409 Circle
#> 447 10.35875809 11.1243144 Circle
#> 448  4.70048498  5.6474418 Circle
#> 449  5.28159768  8.3585996 Circle
#> 450  9.37291549  8.2910520 Circle
#> 451  9.85479063  6.8739244 Circle
#> 452  7.30118746 10.0689612 Circle
#> 453 10.83626622  9.6261000 Circle
#> 454  4.85720857  6.1110049 Circle
#> 455  6.86649700  8.6740232 Circle
#> 456  9.86722241  8.9958422 Circle
#> 457  9.26776555  8.2557920 Circle
#> 458  5.36181966  7.3141833 Circle
#> 459  8.65098630  9.9377320 Circle
#> 460 10.24367737  5.5417697 Circle
#> 461  8.46733210  6.1166784 Circle
#> 462  6.93150077  9.1627730 Circle
#> 463  9.07228786  7.9857613 Circle
#> 464  8.45847977  6.4463083 Circle
#> 465  8.55121305  8.7640590 Circle
#> 466  6.89492139  9.7807550 Circle
#> 467  8.76485687  7.7151223 Circle
#> 468  9.47331868  5.9354136 Circle
#> 469  7.65271342  7.3611297 Circle
#> 470  7.34364057  9.6636974 Circle
#> 471  8.17639027  8.8594926 Circle
#> 472  8.08243359  7.6688579 Circle
#> 473  7.20527746  8.2577979 Circle
#> 474  8.63113164  5.8628075 Circle
#> 475  7.92523423  9.8723674 Circle
#> 476 10.21868075 10.3164997 Circle
#> 477  8.90239417  9.1391036 Circle
#> 478  9.90087472  6.4443917 Circle
#> 479  7.41691085  8.4551444 Circle
#> 480  8.00207435  6.5842405 Circle
#> 481  4.32097863  8.0970716 Circle
#> 482  4.91730987  9.6368484 Circle
#> 483  9.12924227  9.1259427 Circle
#> 484  8.75935007  5.2742440 Circle
#> 485  5.82182590 10.4271186 Circle
#> 486  8.19125542  7.2865793 Circle
#> 487  7.87464915  7.3734266 Circle
#> 488  8.01912232  6.0395963 Circle
#> 489  8.44973636  5.9164113 Circle
#> 490  6.07214162  9.6148472 Circle
#> 491  9.69979738  8.9444290 Circle
#> 492  7.80473187  7.2849123 Circle
#> 493  5.50979440  7.2736450 Circle
#> 494  9.26168647 11.1615821 Circle
#> 495  5.25371084  9.5396585 Circle
#> 496  8.52796289  6.7701999 Circle
#> 497  7.43153555  4.8844835 Circle
#> 498  6.71415237 10.1371930 Circle
#> 499  8.54548941  6.1458025 Circle
#> 500 10.39689894  7.6919184 Circle
#> 501  5.13411523  8.2223904 Circle
#> 502  7.75884243  6.3161875 Circle
#> 503  6.97772826 10.9622233 Circle
#> 504  9.17073213  9.1481621 Circle
#> 505  8.95030857 10.1589010 Circle
#> 506  7.43117814  9.7087594 Circle
#> 507  8.05464505  7.2039773 Circle
#> 508 10.71184705  7.5862885 Circle
#> 509  7.11146750  6.0920008 Circle
#> 510  5.48800731  7.1240721 Circle
#> 511  5.95154596 10.4728500 Circle
#> 512  9.10871087 11.1032418 Circle
#> 513  8.17288658  4.7372313 Circle
#> 514  6.08898669  8.4603616 Circle
#> 515  6.01208649  8.4762803 Circle
#> 516 11.79893448  7.1342334 Circle
#> 517  6.65360950 10.6327412 Circle
#> 518  6.10473730  8.4096173 Circle
#> 519  9.43426978  7.1352472 Circle
#> 520  9.49324279  8.3208194 Circle
#> 521 10.33695898  6.5078673 Circle
#> 522  6.81874317 10.2392997 Circle
#> 523  8.35171047  6.0107897 Circle
#> 524  5.16681962  6.6138886 Circle
#> 525  9.55681418  6.5612762 Circle
#> 526  7.10644723  8.2652712 Circle
#> 527  9.49641460 10.0586998 Circle
#> 528  8.32390080  5.8898412 Circle
#> 529  6.43840725  9.2341077 Circle
#> 530  7.32250273  9.8126086 Circle
#> 531  8.77480334  6.7318128 Circle
#> 532  6.27346746  7.6623323 Circle
#> 533  7.83026752 10.0424084 Circle
#> 534  3.06110256  7.2824882 Circle
#> 535  8.29682757  9.0130898 Circle
#> 536  6.35478795  7.0719131 Circle
#> 537  9.35928010  9.3771851 Circle
#> 538 10.76531980  5.9556729 Circle
#> 539  8.02853484 10.6999882 Circle
#> 540  8.00179588  7.5792472 Circle
#> 541  8.57457703  9.5955050 Circle
#> 542  9.00892656 11.0993675 Circle
#> 543  7.07841346  8.6429634 Circle
#> 544  5.38894899  8.3509315 Circle
#> 545  8.65025331 10.7560865 Circle
#> 546  8.23218693 10.9494508 Circle
#> 547  6.86908843  6.4252543 Circle
#> 548  6.36734061  7.7887234 Circle
#> 549  7.47239623  9.1735127 Circle
#> 550  6.63506466  7.6108625 Circle
#> 551 11.53492584  9.2238666 Circle
#> 552  7.56654884  9.8646959 Circle
#> 553  6.70288787  7.7224385 Circle
#> 554  9.49195489 10.3770500 Circle
#> 555  6.61120398 10.6220232 Circle
#> 556  6.12344419  7.9569457 Circle
#> 557  6.43128345 10.4340914 Circle
#> 558 11.07413991  9.6955750 Circle
#> 559  9.09446444 10.7408734 Circle
#> 560  7.58648523  7.8634835 Circle
#> 561  8.32094762  9.5891962 Circle
#> 562  8.01493145  9.8735849 Circle
#> 563 10.89414017  7.4862730 Circle
#> 564  5.33415051  8.9722952 Circle
#> 565  4.34579913  6.8752741 Circle
#> 566  8.90072738 10.6187337 Circle
#> 567  9.52197465 11.2126224 Circle
#> 568  9.93712979  9.7604037 Circle
#> 569  8.96882773  7.4693312 Circle
#> 570  5.83978924  8.7998514 Circle
#> 571  9.03142778  5.6779836 Circle
#> 572  7.62969334 10.7776781 Circle
#> 573  9.26655917  9.2296103 Circle
#> 574  9.47856955  7.9081559 Circle
#> 575  8.40534671  6.1797034 Circle
#> 576  7.96984845  5.8039954 Circle
#> 577  5.47406321  9.0885064 Circle
#> 578  9.35334033  9.3778258 Circle
#> 579  8.74594725  8.5350015 Circle
#> 580  6.29738775  9.9702995 Circle
#> 581  6.79626627  9.1835944 Circle
#> 582  8.20104194  3.3927398 Circle
#> 583  6.02285033  7.3052954 Circle
#> 584  8.11980154 12.4458590 Circle
#> 585  9.48481565 11.4832279 Circle
#> 586  9.70947432  9.4849534 Circle
#> 587  6.19179488  6.6809498 Circle
#> 588  8.92976096  9.2817114 Circle
#> 589 10.49756705 10.2702616 Circle
#> 590  9.53287982  6.9895041 Circle
#> 591  8.95909262  6.8252969 Circle
#> 592  8.33257744 10.7246434 Circle
#> 593  8.67460318  9.3193154 Circle
#> 594  4.83490596  7.4016944 Circle
#> 595  8.80048474  7.8379143 Circle
#> 596  5.12510346  6.3998865 Circle
#> 597  8.01474030  7.9614526 Circle
#> 598  6.21420857  7.0054749 Circle
#> 599  6.93551761  8.7538657 Circle
#> 600 10.81898753  8.2409878 Circle
#> 601  8.88244538  7.9127278 Circle
#> 602  4.83073110  7.8798344 Circle
#> 603  9.47103444  7.0222721 Circle
#> 604  6.63389565 10.3172392 Circle
#> 605  7.22994255  8.3078972 Circle
#> 606  6.56930648  5.3881722 Circle
#> 607  8.37943786  6.7109460 Circle
#> 608  6.75501315  8.8005501 Circle
#> 609 10.12483530  9.9641030 Circle
#> 610  5.53612464  7.3644431 Circle
#> 611  8.11215449  7.0788635 Circle
#> 612  9.27798370  8.0487793 Circle
#> 613  6.54927371 10.4399951 Circle
#> 614  9.27845226  6.9450775 Circle
#> 615 10.01613681  7.6756091 Circle
#> 616  5.52326055  9.2598074 Circle
#> 617  4.61671376  7.7063257 Circle
#> 618  8.60762022 10.4954313 Circle
#> 619  5.66099660  8.9216730 Circle
#> 620  8.43542248  5.7826760 Circle
#> 621  9.50729677 10.7425136 Circle
#> 622  9.91103164  9.2777376 Circle
#> 623  7.44238270  6.1775594 Circle
#> 624  7.36713718  6.4759110 Circle
#> 625  8.36560465  9.5710497 Circle
#> 626  5.18693787  5.6025304 Circle
#> 627  7.21629830  7.0116719 Circle
#> 628  9.74775977  6.2399088 Circle
#> 629  8.20716119  4.3659212 Circle
#> 630 10.27472209 10.8281584 Circle
#> 631  5.44541507  7.2667021 Circle
#> 632  7.51427497  6.2705555 Circle
#> 633  7.10027192  9.2221200 Circle
#> 634  9.14761070  8.9227845 Circle
#> 635  6.31970254  7.1793398 Circle
#> 636  9.00704830 10.8487094 Circle
#> 637  7.90296646  9.1502404 Circle
#> 638  7.58889761  9.4022842 Circle
#> 639  5.86614147  7.8697338 Circle
#> 640  7.12733872 10.1260660 Circle
#> 641  7.55485768  8.3357487 Circle
#> 642  4.84723494  6.6128268 Circle
#> 643  8.60227178  8.0264779 Circle
#> 644  8.62047098  9.0938437 Circle
#> 645  5.06628370  8.3199901 Circle
#> 646 10.89996706  9.4024624 Circle
#> 647  8.23501170  9.1786535 Circle
#> 648  8.29194515  8.5480971 Circle
#> 649  5.18870337  7.4006254 Circle
#> 650  6.63372210  8.9472033 Circle
#> 651  9.20862218  8.8519956 Circle
#> 652  7.99253716 10.7844282 Circle
#> 653  9.25196168  5.4218506 Circle
#> 654  7.98905669  9.2618626 Circle
#> 655  8.95518840  8.3544857 Circle
#> 656  9.64252870 11.8644449 Circle
#> 657  7.95948919 10.7395951 Circle
#> 658  9.48141076  9.8990790 Circle
#> 659  7.48280053  9.1224467 Circle
#> 660 11.84566994  8.1906357 Circle
#> 661 11.11978404  8.4789590 Circle
#> 662 10.76761411  6.4113528 Circle
#> 663  5.97085212  7.2702915 Circle
#> 664 11.04564218  8.3819170 Circle
#> 665  5.51364511  7.8135418 Circle
#> 666  9.75254247  9.0097339 Circle
#> 667  7.37401068  8.1301601 Circle
#> 668 10.17090598  9.1488259 Circle
#> 669  8.11331964  5.9324174 Circle
#> 670  9.75817589  6.0295078 Circle
#> 671  5.70194800  7.2963094 Circle
#> 672  6.68263723 10.3780465 Circle
#> 673  8.57149728 11.1717666 Circle
#> 674 11.99729246  8.7853780 Circle
#> 675  8.47561840  6.7152712 Circle
#> 676  6.40890228  7.4997710 Circle
#> 677  8.62619288 10.3493442 Circle
#> 678  8.54744941  8.9991090 Circle
#> 679 10.04775844 10.0982580 Circle
#> 680  7.05261620 10.1973212 Circle
#> 681  8.42902880 10.7090845 Circle
#> 682  9.91604696 11.2925778 Circle
#> 683  7.98441574 10.2741265 Circle
#> 684  9.20677265  8.2147606 Circle

rose(circle_example[, c("x", "y", "class")], var = "class", over_ratio = 0.8)
#>              x          y  class
#> 1   11.7667072 12.6577281   Rest
#> 2    7.0892870  2.1877948   Rest
#> 3    7.9595892  1.2473397   Rest
#> 4   13.6152031 10.8558946   Rest
#> 5   12.1396490  8.6775428   Rest
#> 6    8.1894482 12.1821896   Rest
#> 7   15.7066838  8.6019819   Rest
#> 8   11.8373725 10.4246157   Rest
#> 9    5.7131993  3.6819137   Rest
#> 10   6.7494605 14.8578778   Rest
#> 11   4.8084950  2.8567000   Rest
#> 12   6.4945614 13.3563514   Rest
#> 13   9.6690624 17.3582790   Rest
#> 14  14.7681717  6.0930347   Rest
#> 15   3.3468324  9.4708804   Rest
#> 16  13.6753726 13.1641773   Rest
#> 17  13.6587133 13.4906651   Rest
#> 18  14.2797553  6.6481380   Rest
#> 19  12.3247389 10.0113864   Rest
#> 20  12.9123216  8.7558641   Rest
#> 21   8.0596584 10.3333967   Rest
#> 22   8.6046983 16.7566604   Rest
#> 23   4.1502910 10.1856877   Rest
#> 24   1.5916118 12.9440717   Rest
#> 25   8.5440064  2.0095756   Rest
#> 26   8.8881870  2.3432720   Rest
#> 27   4.3730965 15.1643643   Rest
#> 28   9.8742703  4.5902000   Rest
#> 29  15.7247948  3.0810113   Rest
#> 30   4.1387425  2.8428110   Rest
#> 31  11.3002504 10.0791250   Rest
#> 32   7.5695590 13.4777877   Rest
#> 33  14.8570715  7.1682143   Rest
#> 34   9.6490424 14.0074403   Rest
#> 35   1.8005317  9.2770152   Rest
#> 36   8.9539374  4.0494635   Rest
#> 37   5.7085415 11.5133431   Rest
#> 38  12.9289171  2.3494396   Rest
#> 39   6.9042117  9.6925946   Rest
#> 40   1.2159436  5.8934131   Rest
#> 41   1.6543799  7.1532208   Rest
#> 42   8.2731125 13.8852787   Rest
#> 43  13.2787889  1.1568817   Rest
#> 44   0.7740273 12.9196883   Rest
#> 45  11.5581278  3.8349505   Rest
#> 46   3.5107405 15.3819025   Rest
#> 47   1.2482194 13.4395826   Rest
#> 48   8.0195646  0.1677711   Rest
#> 49   1.9532905 11.5253006   Rest
#> 50  12.7153348 10.4850806   Rest
#> 51   2.8973175 16.5833676   Rest
#> 52   4.4287676  4.1571797   Rest
#> 53   9.2622312  5.6498170   Rest
#> 54  14.5149772  4.5444689   Rest
#> 55   6.1592276  3.4040635   Rest
#> 56  -0.1146773  7.7811566   Rest
#> 57   6.6107264  0.6884451   Rest
#> 58   3.1312656  6.6536896   Rest
#> 59  13.9967218  2.8248096   Rest
#> 60   6.3283021  4.1209752   Rest
#> 61   7.3261357 13.1145861   Rest
#> 62   6.2703480  2.2712232   Rest
#> 63   9.9657902 12.9990516   Rest
#> 64  13.0071453  9.8901874   Rest
#> 65  11.2028319  1.9676322   Rest
#> 66   7.1092000 12.0771218   Rest
#> 67  11.8371173  2.2290220   Rest
#> 68  10.8679548 10.6183289   Rest
#> 69  10.5623820 14.2601186   Rest
#> 70   1.8466508  6.3605286   Rest
#> 71   7.5748350 15.4866340   Rest
#> 72  12.2873819  2.0865748   Rest
#> 73  14.4392378 10.6338660   Rest
#> 74   6.1438767  3.3705951   Rest
#> 75  12.8884168 16.5033345   Rest
#> 76   5.7535045 12.0476555   Rest
#> 77   4.6491719 -1.0706232   Rest
#> 78   4.3526162  2.9041730   Rest
#> 79  14.3477429 -0.7120180   Rest
#> 80  13.4833344  9.8507912   Rest
#> 81   6.1070076  1.8182291   Rest
#> 82  12.0608735  3.7277448   Rest
#> 83   5.0731847 14.9162565   Rest
#> 84   6.6708250  8.1799026   Rest
#> 85  13.9114063  2.9551432   Rest
#> 86   3.8388048  9.5591616   Rest
#> 87  12.3318439  4.9901051   Rest
#> 88  12.2255138  8.6149659   Rest
#> 89   9.2465982  9.7868247   Rest
#> 90  10.4866283  7.3300846   Rest
#> 91   9.9201476 10.9845092   Rest
#> 92   4.8155876  8.0981025   Rest
#> 93   2.6992207  3.3614508   Rest
#> 94   5.7808516  9.0454734   Rest
#> 95   5.1937049 12.7948081   Rest
#> 96  12.9287003  4.8129470   Rest
#> 97   8.3770220 13.7296734   Rest
#> 98  11.9383107  9.3298245   Rest
#> 99   8.5545826 16.1495525   Rest
#> 100 12.3871061 11.9913892   Rest
#> 101  9.7216000  5.2368493   Rest
#> 102  8.7957478 18.9419838   Rest
#> 103  7.8723480  3.8365956   Rest
#> 104 10.7324358 10.0104663   Rest
#> 105  7.4349765 15.5193040   Rest
#> 106 13.4194007 11.3462215   Rest
#> 107  4.4431319  3.9271765   Rest
#> 108  7.4641646 11.7122883   Rest
#> 109  9.0496221 13.7439620   Rest
#> 110 -1.0079466  7.7641814   Rest
#> 111 15.2078512 15.8435317   Rest
#> 112 14.2396071  8.3341325   Rest
#> 113  7.8222310 17.7979478   Rest
#> 114  9.8948615 15.9394061   Rest
#> 115  4.0733150  6.2368347   Rest
#> 116  5.5258649  8.3486634   Rest
#> 117  6.7702273 12.6119150   Rest
#> 118  2.1682508  7.6854645   Rest
#> 119 12.0545341 12.7668256   Rest
#> 120 14.6273352  4.7674681   Rest
#> 121  3.2381412  7.6419855   Rest
#> 122  9.6455404  4.5336377   Rest
#> 123  5.8718431 14.3170397   Rest
#> 124 15.9665757 16.1513999   Rest
#> 125 10.2152378  3.5191331   Rest
#> 126 15.7707528  7.6390621   Rest
#> 127  4.6220777 11.3556066   Rest
#> 128  9.5004720 14.2546342   Rest
#> 129 10.3754876  6.3060445   Rest
#> 130 10.2784003  7.8858625   Rest
#> 131  2.1050959  6.3307318   Rest
#> 132  5.5273223  1.4431217   Rest
#> 133  4.6927622  8.9756752   Rest
#> 134  3.4121952 14.4118020   Rest
#> 135 11.7548378  2.5448473   Rest
#> 136 12.7701518  0.9822972   Rest
#> 137 12.0057593  3.7789925   Rest
#> 138  2.5248886  9.8055396   Rest
#> 139  1.9655289 10.6020235   Rest
#> 140  0.5173008  8.7459238   Rest
#> 141  2.1217469  8.7295378   Rest
#> 142  5.3076029  3.7758702   Rest
#> 143  9.4763887 13.8974540   Rest
#> 144  0.1743561  3.5341828   Rest
#> 145 13.2788124  5.7396777   Rest
#> 146  1.4314676  7.9623356   Rest
#> 147  7.9378585 10.3968900   Rest
#> 148  8.8390260  6.2993131   Rest
#> 149  4.4110938 14.0580790   Rest
#> 150  1.7222256  2.3277985   Rest
#> 151 12.8081854  5.3279849   Rest
#> 152  6.2036973  8.0539710   Rest
#> 153  5.3617219 12.4933953   Rest
#> 154  2.8166205 17.4674647   Rest
#> 155  4.8886961  3.9351730   Rest
#> 156 12.6313519  2.7042774   Rest
#> 157  0.1808625 10.4837201   Rest
#> 158  5.3007462  0.5388264   Rest
#> 159  4.1622651  7.4419800   Rest
#> 160 12.3000940  9.6868085   Rest
#> 161  9.4918414 11.7644053   Rest
#> 162 11.8885325 16.3673857   Rest
#> 163  2.1329618 12.3686553   Rest
#> 164  6.0453535  4.2783540   Rest
#> 165  8.5008006 10.8180497   Rest
#> 166 10.1033580  3.5959782   Rest
#> 167  9.9989969 13.6225258   Rest
#> 168 13.6436004 10.1494006   Rest
#> 169 14.4880142  4.9329626   Rest
#> 170 16.0111253  1.1275624   Rest
#> 171 10.3820514  0.5040494   Rest
#> 172  7.9889007 11.2628233   Rest
#> 173  4.9702378 11.0365159   Rest
#> 174  6.7046528 10.9598849   Rest
#> 175  2.1988529  9.5703030   Rest
#> 176 12.9267353  8.1403488   Rest
#> 177  8.1195184  1.9702280   Rest
#> 178  4.3538798 10.5636414   Rest
#> 179 14.7958174  4.5748727   Rest
#> 180  4.2006596  2.6205227   Rest
#> 181  9.8524577 16.0708806   Rest
#> 182  5.7136394  3.1701405   Rest
#> 183  6.1136015  2.7421816   Rest
#> 184 12.4733940  0.2456047   Rest
#> 185  2.7307779  3.9407520   Rest
#> 186  2.6561518 15.0766951   Rest
#> 187 11.1003988 13.9577423   Rest
#> 188  9.9113402  9.9708068   Rest
#> 189  5.5600070 13.0210329   Rest
#> 190  5.5101710  4.0856379   Rest
#> 191 14.3825516  8.4382171   Rest
#> 192 10.6862855 11.6008951   Rest
#> 193  2.9418271 12.8244369   Rest
#> 194  5.5167349 10.8765447   Rest
#> 195 12.4511207  1.3001497   Rest
#> 196 15.2469511 -0.1689673   Rest
#> 197  5.0374018  9.0509271   Rest
#> 198  9.2623959  4.0980381   Rest
#> 199  8.4510416 13.2641798   Rest
#> 200 15.9445730  9.0630804   Rest
#> 201 12.7347373 12.6511265   Rest
#> 202 12.1641408 13.4593849   Rest
#> 203 17.0680222  5.7260933   Rest
#> 204  3.5091339  7.4637926   Rest
#> 205  4.9591922  2.7015997   Rest
#> 206  4.3426210  1.4998814   Rest
#> 207 10.9390944 11.4560758   Rest
#> 208  3.3564443  4.4161147   Rest
#> 209  7.9627703 12.5507503   Rest
#> 210  9.5990425 13.4292451   Rest
#> 211  6.9087223  6.6459347   Rest
#> 212  7.6079570  0.5045040   Rest
#> 213 15.5690045 14.4471339   Rest
#> 214 11.9292989  5.2701017   Rest
#> 215  1.1280562  8.0996926   Rest
#> 216  7.8373374 11.6333963   Rest
#> 217 11.8391513  3.2834383   Rest
#> 218  9.1573893  3.0764364   Rest
#> 219 11.2840221 13.9306292   Rest
#> 220  3.9429676 12.7713772   Rest
#> 221 15.6595222  9.9348854   Rest
#> 222  7.4543159  0.8946581   Rest
#> 223  4.3278839  5.2642683   Rest
#> 224 12.9642344  3.5506652   Rest
#> 225  6.5916700  4.7757094   Rest
#> 226 13.6157667 10.6397001   Rest
#> 227  3.7226461  9.0167651   Rest
#> 228 13.1692777 11.9433403   Rest
#> 229 12.8045923 10.0789210   Rest
#> 230 11.5408738  8.4055748   Rest
#> 231  4.3545980 10.8849843   Rest
#> 232  8.2807365  3.9142038   Rest
#> 233  8.4133111 12.8106128   Rest
#> 234  0.4405845  8.6145678   Rest
#> 235 14.5918537  5.5896071   Rest
#> 236 12.1697538 11.8873690   Rest
#> 237 10.7046602 16.9307452   Rest
#> 238  3.6305462 11.7900467   Rest
#> 239 12.0566200  5.6184507   Rest
#> 240 13.6668503  5.2805295   Rest
#> 241  3.1530218  7.6898813   Rest
#> 242 11.0141858  5.4515019   Rest
#> 243 14.1308190  4.4468585   Rest
#> 244 16.0389090  2.3833814   Rest
#> 245  4.2404404 11.4572076   Rest
#> 246  2.4377271 -0.7220831   Rest
#> 247  5.0753831 14.0461687   Rest
#> 248 14.4525251 -1.1890723   Rest
#> 249  8.3758833  3.2967222   Rest
#> 250 14.2930547 13.1084285   Rest
#> 251  3.2080137  4.3510517   Rest
#> 252 15.7502177  6.3190784   Rest
#> 253 10.2123674  2.2029156   Rest
#> 254  8.5124759 13.1291618   Rest
#> 255 -0.7773267 11.4995193   Rest
#> 256  6.8524769  5.7117575   Rest
#> 257 14.8512521  2.8846910   Rest
#> 258  4.2903508 13.4247816   Rest
#> 259 12.6036618 12.6361705   Rest
#> 260 -2.4672106  2.4829291   Rest
#> 261  3.3197241  4.0286345   Rest
#> 262  8.4890102  9.6041670 Circle
#> 263  4.9663937  7.9039727 Circle
#> 264  5.5130737  7.7278020 Circle
#> 265  6.4299863  7.8946676 Circle
#> 266  9.6834617  8.1709191 Circle
#> 267 10.6751716 10.1622261 Circle
#> 268  5.6034411  8.3777105 Circle
#> 269 11.4474238  9.9404901 Circle
#> 270  9.0682728  8.0754058 Circle
#> 271  6.2874891  7.9143901 Circle
#> 272  8.0183289  7.9407942 Circle
#> 273  9.6435609  9.5452527 Circle
#> 274  9.4883409  6.7590527 Circle
#> 275  5.1928849  9.2462501 Circle
#> 276  6.1598195  8.7230568 Circle
#> 277  9.1803906  6.6872751 Circle
#> 278  6.3967106  8.8824430 Circle
#> 279  9.2538457 11.9613788 Circle
#> 280  8.8818824  9.6429250 Circle
#> 281  9.5494120  7.9047119 Circle
#> 282  8.7656657  9.4947146 Circle
#> 283  9.0929017  7.9903075 Circle
#> 284  9.5953338  9.6696441 Circle
#> 285  8.5398188  6.8735083 Circle
#> 286  4.3413489  6.8907026 Circle
#> 287  8.5186094  8.4237276 Circle
#> 288  8.9722427  6.8555672 Circle
#> 289  9.8300173  5.6610186 Circle
#> 290  5.4917972  7.8939568 Circle
#> 291  8.2881977  8.8914677 Circle
#> 292  7.3316488  8.5207637 Circle
#> 293  6.6773087  6.7553424 Circle
#> 294  9.6755872  8.5542547 Circle
#> 295 10.5224246  6.5428906 Circle
#> 296  5.9099413  7.9955891 Circle
#> 297 10.2340327  7.9331467 Circle
#> 298  8.9991307  8.7447546 Circle
#> 299  8.4009386  6.2600826 Circle
#> 300 11.7423637 10.4058635 Circle
#> 301  8.1738877  5.1659534 Circle
#> 302  7.8865834  9.4026532 Circle
#> 303  5.6626134  6.9334342 Circle
#> 304  5.6932662  7.8863676 Circle
#> 305  6.8150769 10.2574836 Circle
#> 306  9.8042793 10.7438476 Circle
#> 307  8.3150235  8.7472434 Circle
#> 308  8.7908133  6.9518766 Circle
#> 309  5.6842163  6.7864785 Circle
#> 310  8.0486208  9.5714638 Circle
#> 311  9.4244170  9.0280263 Circle
#> 312  8.1926927  6.1428072 Circle
#> 313  5.9997258  6.4446994 Circle
#> 314  7.1770833 10.4561971 Circle
#> 315  7.7362395  9.6559767 Circle
#> 316 10.5429112  4.6896736 Circle
#> 317  6.7408679  8.2614128 Circle
#> 318  4.9180410  8.0673734 Circle
#> 319 10.2168357  6.5142245 Circle
#> 320  7.7790532  5.0996710 Circle
#> 321  5.3237247  7.9084004 Circle
#> 322  7.8591373  8.6251788 Circle
#> 323  9.5881403  6.0872756 Circle
#> 324  5.1589416  8.2271282 Circle
#> 325  6.4991577  9.9355778 Circle
#> 326  7.4635961  5.2697781 Circle
#> 327  9.9945251  7.9714738 Circle
#> 328  6.4829887  7.2954695 Circle
#> 329  7.2715093  9.8399688 Circle
#> 330  8.0547888  9.2094037 Circle
#> 331  7.4142709  7.6735002 Circle
#> 332  7.3745018 10.6862174 Circle
#> 333  5.2041098  8.9456925 Circle
#> 334  6.9660066  9.9294451 Circle
#> 335  7.1498046  9.8906068 Circle
#> 336  8.1978257  6.9029116 Circle
#> 337  8.5039054 10.0703896 Circle
#> 338  6.2219677  7.8981018 Circle
#> 339  9.5866875  9.7959185 Circle
#> 340  8.9836482  6.3979431 Circle
#> 341  8.8890100  8.0429423 Circle
#> 342 10.2517515 10.4010819 Circle
#> 343  5.8179176 10.3760383 Circle
#> 344  8.5643802  6.3083739 Circle
#> 345  6.4016754  8.3983251 Circle
#> 346  8.2202180  9.1883750 Circle
#> 347  8.7066943 10.1372602 Circle
#> 348  6.6699114  6.5102093 Circle
#> 349  8.6518088  7.5058150 Circle
#> 350  8.9728918  8.2186810 Circle
#> 351  4.6863502  8.1235691 Circle
#> 352  8.0724724  9.7549255 Circle
#> 353 10.8837754  9.6106785 Circle
#> 354  9.0748929  8.0595479 Circle
#> 355  4.8557338  7.7870977 Circle
#> 356  8.8474029  8.9505209 Circle
#> 357  6.5970279  7.7348016 Circle
#> 358  7.7548404  5.4543224 Circle
#> 359 10.4250370 10.4365689 Circle
#> 360  5.9352808 10.2291729 Circle
#> 361  9.0985875  7.9691076 Circle
#> 362  7.2496400  6.8967762 Circle
#> 363  7.3852907 10.9652046 Circle
#> 364  5.5372235  8.7187084 Circle
#> 365  8.5643515  6.5631168 Circle
#> 366  9.9589235  7.5959929 Circle
#> 367  9.0351056  8.9058319 Circle
#> 368  5.6096030  7.2517785 Circle
#> 369  9.8967966 11.1834676 Circle
#> 370  8.1947791  5.9699339 Circle
#> 371  9.6377120  6.3639876 Circle
#> 372  6.9469629  6.7107712 Circle
#> 373  7.1295395  6.7395972 Circle
#> 374  7.1712957  7.7472278 Circle
#> 375  5.6218075  8.9491003 Circle
#> 376  6.6132744  8.1566863 Circle
#> 377  7.3495469  7.9212502 Circle
#> 378  5.6094727  5.6757936 Circle
#> 379  8.4871807  6.5961777 Circle
#> 380 10.6646768  6.1013358 Circle
#> 381 10.4192434  9.2338347 Circle
#> 382  6.4252173  8.2921339 Circle
#> 383  6.4161033  6.4984240 Circle
#> 384  9.5301297  8.1966170 Circle
#> 385  8.8287033  8.2340869 Circle
#> 386  9.0512065  8.5497703 Circle
#> 387  7.4855346  7.8150849 Circle
#> 388  8.4529526  8.7545268 Circle
#> 389  8.7479220  8.4136370 Circle
#> 390  3.9792191  7.7808262 Circle
#> 391  6.0181208 11.1837033 Circle
#> 392  6.9863396  7.9568444 Circle
#> 393  8.4285018  6.0633307 Circle
#> 394  6.6320426  6.3923581 Circle
#> 395  7.6975276  8.0869180 Circle
#> 396  6.4502229  9.1708914 Circle
#> 397  8.1630805 10.6015240 Circle
#> 398  7.2789610  6.5170152 Circle
#> 399  8.1222420  8.4292177 Circle
#> 400  9.2922416  8.2045682 Circle
#> 401  6.1953491  7.4100362 Circle
#> 402  7.5590856  7.3077891 Circle
#> 403  8.1365066  8.6938880 Circle
#> 404  5.9423823  8.3330138 Circle
#> 405 10.3577453  9.7176557 Circle
#> 406  6.3063386  7.3698514 Circle
#> 407  7.4569706  7.9277424 Circle
#> 408  4.9945257  6.9805463 Circle
#> 409  7.7619119  7.1767983 Circle
#> 410  6.3734369  7.9661507 Circle
#> 411  6.1085111  7.5975585 Circle
#> 412  7.5908416  6.0775607 Circle
#> 413  5.6014375  7.7956433 Circle
#> 414  9.0011598  4.7332852 Circle
#> 415  5.1637376  9.1050392 Circle
#> 416  9.3393419  6.3415119 Circle
#> 417  8.4785873  9.4695721 Circle
#> 418  6.8431072  6.2083172 Circle
#> 419 10.7215943  7.9487038 Circle
#> 420  9.3595808  6.8142022 Circle
#> 421  6.5332558  7.8976615 Circle
#> 422 10.1732857  8.1306786 Circle
#> 423  6.2341036 11.0845997 Circle
#> 424  7.5597599  8.8005584 Circle
#> 425  5.4648252  6.6486678 Circle
#> 426  5.4462187  7.7226482 Circle
#> 427  7.8575067  9.5993313 Circle
#> 428  9.9997524  6.8046465 Circle
#> 429  5.2208339  8.5009654 Circle
#> 430  7.6710127  9.1046178 Circle
#> 431  5.5404514  9.6547020 Circle
#> 432  6.5596204  9.6817603 Circle
#> 433  7.5472852  8.0379023 Circle
#> 434 10.5859081  7.7772512 Circle
#> 435  7.7889023  8.0312628 Circle
#> 436  9.1879576  8.9364517 Circle
#> 437  8.3877202  5.2410108 Circle
#> 438  7.3762204 10.8724474 Circle
#> 439  7.4406273 10.2904288 Circle
#> 440  7.2891404  8.6847846 Circle
#> 441  7.7472437 10.2171290 Circle
#> 442  9.3332744  7.2386058 Circle
#> 443  7.1934704 10.2399411 Circle
#> 444  5.6204670  8.7299663 Circle
#> 445  5.7595651  7.6566548 Circle
#> 446  6.1721595  8.0096625 Circle
#> 447  5.8063228  7.5655179 Circle
#> 448  7.6915141  7.5193726 Circle
#> 449  8.6280489  6.4323794 Circle
#> 450 11.6197893  8.7052015 Circle
#> 451  5.9376153  9.6580717 Circle
#> 452  8.5013470  9.8685171 Circle
#> 453  7.0961193  8.7128519 Circle
#> 454  6.2622240  8.5840087 Circle
#> 455  9.0272192  7.7193356 Circle
#> 456  7.4473931  9.4948317 Circle
#> 457  8.8052597  9.3365891 Circle
#> 458  9.6653534  4.1818638 Circle
#> 459  8.8932366  9.5694696 Circle
#> 460  8.3445629  5.9925048 Circle
#> 461  7.7056781 10.4896104 Circle
#> 462  6.3039580  7.3210254 Circle
#> 463  5.2760553  5.7118907 Circle
#> 464  7.1510557  9.6565447 Circle
#> 465  5.2355955  7.1761969 Circle
#> 466  5.5147503  9.0331963 Circle
#> 467  6.7058453  8.9841166 Circle
#> 468  9.1672382  9.5685117 Circle
#> 469  5.5415424  9.4997430 Circle
#> 470  8.6904990  9.7082011 Circle
#> 471  8.0354180  7.6890266 Circle
#> 472  6.8427309  9.3292229 Circle
#> 473  6.9506260  9.0677750 Circle
#> 474  8.0350502 11.6196596 Circle
#> 475  7.8406828  5.3184981 Circle
#> 476 10.3380782  9.0164159 Circle
#> 477  5.4697108  8.2195853 Circle
#> 478  7.0685530 10.6648611 Circle
#> 479  9.2562471  8.0872175 Circle
#> 480  8.4327574  6.9226126 Circle
#> 481  6.4307055  9.3520432 Circle
#> 482  6.7540428  6.4225373 Circle
#> 483  7.3590417  6.0037445 Circle
#> 484  8.1766260  6.4735248 Circle
#> 485  9.2540723  8.1522088 Circle
#> 486  6.2225026  8.7309300 Circle
#> 487  6.6007209 10.6176641 Circle
#> 488  8.2247016  6.7318788 Circle
#> 489  7.0417617  7.2653646 Circle
#> 490  9.9450608  8.4169435 Circle
#> 491  7.2219814 10.2687844 Circle
#> 492  8.2872772  7.7513480 Circle
#> 493  8.0814381  5.5480537 Circle
#> 494  9.4684244  8.6654563 Circle
#> 495  7.4416244 10.3806489 Circle
#> 496 11.6668261  9.5239473 Circle
#> 497  8.8627410  8.8502163 Circle
#> 498 10.8040157  9.1688201 Circle
#> 499  8.6760443  6.3484269 Circle
#> 500  6.5291304  9.3310950 Circle
#> 501  7.1065216  6.7074003 Circle
#> 502  6.3492047  9.8267275 Circle
#> 503  7.0065779  8.3481654 Circle
#> 504  7.9881055  7.9193068 Circle
#> 505  8.3606913  6.8068585 Circle
#> 506  7.0449909 10.4274663 Circle
#> 507 10.5715734  8.3924223 Circle
#> 508  5.6243177  6.4088533 Circle
#> 509 10.2153429  4.9205925 Circle
#> 510  6.7617149  5.5093596 Circle
#> 511  7.0374856  8.5638387 Circle
#> 512 10.2654432  8.1724494 Circle
#> 513  9.8165494  9.5254120 Circle
#> 514  5.8409843  6.2303459 Circle
#> 515  8.1735467  7.6850550 Circle
#> 516 10.2584152  8.8970713 Circle
#> 517  7.3214467  8.9952490 Circle
#> 518  5.8199935  8.4817662 Circle
#> 519  6.7673977  7.1931804 Circle
#> 520  8.0015937  5.9734254 Circle
#> 521  6.2095248  8.0101570 Circle
#> 522  8.4197825  7.5206187 Circle
#> 523  8.9481727  9.1537316 Circle
#> 524  7.8344347 11.0533288 Circle
#> 525  8.2433768  9.1736816 Circle
#> 526  5.6550086  7.0384263 Circle
#> 527 10.6299149  8.0474821 Circle
#> 528  6.7879105  7.9712115 Circle
#> 529  8.6894524  9.7784896 Circle
#> 530  5.2686649 10.3652280 Circle
#> 531  5.3081791  8.2324173 Circle
#> 532  7.4429989 10.7958986 Circle
#> 533  9.1380925 10.4454639 Circle
#> 534  7.7015427 10.2129959 Circle
#> 535  8.5554406 10.4304620 Circle
#> 536  5.1059818  9.7201446 Circle
#> 537  7.9885555 10.2745851 Circle
#> 538  5.4297693  8.1197016 Circle
#> 539  7.6873777  8.4759870 Circle
#> 540  8.3663015 10.2905636 Circle
#> 541 10.2869023  5.7115120 Circle
#> 542  8.2346958  6.4573103 Circle
#> 543  4.6849603  7.5373141 Circle
#> 544  5.6789812  9.4814145 Circle
#> 545  7.8229919  7.7707613 Circle
#> 546  4.7903862  6.9344544 Circle
#> 547  6.2088557  7.4681980 Circle